refactor: modularize components and helpers
