+++ docs/assets/index-BQTPNGku.js
@@ -46,4 +46,4 @@
 `+f[e].replace(" at new "," at ");return l.displayName&&z.includes("<anonymous>")&&(z=z.replace("<anonymous>",l.displayName)),z}while(1<=e&&0<=u);break}}}finally{Fn=!1,Error.prepareStackTrace=a}return(a=l?l.displayName||l.name:"")?Za(a):""}function Br(l){switch(l.tag){case 26:case 27:case 5:return Za(l.type);case 16:return Za("Lazy");case 13:return Za("Suspense");case 19:return Za("SuspenseList");case 0:case 15:return Pn(l.type,!1);case 11:return Pn(l.type.render,!1);case 1:return Pn(l.type,!0);case 31:return Za("Activity");default:return""}}function Of(l){try{var t="";do t+=Br(l),l=l.return;while(l);return t}catch(a){return`
 Error generating stack: `+a.message+`
 `+a.stack}}function st(l){switch(typeof l){case"bigint":case"boolean":case"number":case"string":case"undefined":return l;case"object":return l;default:return""}}function pf(l){var t=l.type;return(l=l.nodeName)&&l.toLowerCase()==="input"&&(t==="checkbox"||t==="radio")}function Cr(l){var t=pf(l)?"checked":"value",a=Object.getOwnPropertyDescriptor(l.constructor.prototype,t),e=""+l[t];if(!l.hasOwnProperty(t)&&typeof a<"u"&&typeof a.get=="function"&&typeof a.set=="function"){var u=a.get,n=a.set;return Object.defineProperty(l,t,{configurable:!0,get:function(){return u.call(this)},set:function(c){e=""+c,n.call(this,c)}}),Object.defineProperty(l,t,{enumerable:a.enumerable}),{getValue:function(){return e},setValue:function(c){e=""+c},stopTracking:function(){l._valueTracker=null,delete l[t]}}}}function Uu(l){l._valueTracker||(l._valueTracker=Cr(l))}function xf(l){if(!l)return!1;var t=l._valueTracker;if(!t)return!0;var a=t.getValue(),e="";return l&&(e=pf(l)?l.checked?"true":"false":l.value),l=e,l!==a?(t.setValue(l),!0):!1}function Ru(l){if(l=l||(typeof document<"u"?document:void 0),typeof l>"u")return null;try{return l.activeElement||l.body}catch{return l.body}}var Gr=/[\n"\\]/g;function ot(l){return l.replace(Gr,function(t){return"\\"+t.charCodeAt(0).toString(16)+" "})}function In(l,t,a,e,u,n,c,i){l.name="",c!=null&&typeof c!="function"&&typeof c!="symbol"&&typeof c!="boolean"?l.type=c:l.removeAttribute("type"),t!=null?c==="number"?(t===0&&l.value===""||l.value!=t)&&(l.value=""+st(t)):l.value!==""+st(t)&&(l.value=""+st(t)):c!=="submit"&&c!=="reset"||l.removeAttribute("value"),t!=null?lc(l,c,st(t)):a!=null?lc(l,c,st(a)):e!=null&&l.removeAttribute("value"),u==null&&n!=null&&(l.defaultChecked=!!n),u!=null&&(l.checked=u&&typeof u!="function"&&typeof u!="symbol"),i!=null&&typeof i!="function"&&typeof i!="symbol"&&typeof i!="boolean"?l.name=""+st(i):l.removeAttribute("name")}function _f(l,t,a,e,u,n,c,i){if(n!=null&&typeof n!="function"&&typeof n!="symbol"&&typeof n!="boolean"&&(l.type=n),t!=null||a!=null){if(!(n!=="submit"&&n!=="reset"||t!=null))return;a=a!=null?""+st(a):"",t=t!=null?""+st(t):a,i||t===l.value||(l.value=t),l.defaultValue=t}e=e??u,e=typeof e!="function"&&typeof e!="symbol"&&!!e,l.checked=i?l.checked:!!e,l.defaultChecked=!!e,c!=null&&typeof c!="function"&&typeof c!="symbol"&&typeof c!="boolean"&&(l.name=c)}function lc(l,t,a){t==="number"&&Ru(l.ownerDocument)===l||l.defaultValue===""+a||(l.defaultValue=""+a)}function Va(l,t,a,e){if(l=l.options,t){t={};for(var u=0;u<a.length;u++)t["$"+a[u]]=!0;for(a=0;a<l.length;a++)u=t.hasOwnProperty("$"+l[a].value),l[a].selected!==u&&(l[a].selected=u),u&&e&&(l[a].defaultSelected=!0)}else{for(a=""+st(a),t=null,u=0;u<l.length;u++){if(l[u].value===a){l[u].selected=!0,e&&(l[u].defaultSelected=!0);return}t!==null||l[u].disabled||(t=l[u])}t!==null&&(t.selected=!0)}}function Mf(l,t,a){if(t!=null&&(t=""+st(t),t!==l.value&&(l.value=t),a==null)){l.defaultValue!==t&&(l.defaultValue=t);return}l.defaultValue=a!=null?""+st(a):""}function Df(l,t,a,e){if(t==null){if(e!=null){if(a!=null)throw Error(s(92));if(il(e)){if(1<e.length)throw Error(s(93));e=e[0]}a=e}a==null&&(a=""),t=a}a=st(t),l.defaultValue=a,e=l.textContent,e===a&&e!==""&&e!==null&&(l.value=e)}function La(l,t){if(t){var a=l.firstChild;if(a&&a===l.lastChild&&a.nodeType===3){a.nodeValue=t;return}}l.textContent=t}var Xr=new Set("animationIterationCount aspectRatio borderImageOutset borderImageSlice borderImageWidth boxFlex boxFlexGroup boxOrdinalGroup columnCount columns flex flexGrow flexPositive flexShrink flexNegative flexOrder gridArea gridRow gridRowEnd gridRowSpan gridRowStart gridColumn gridColumnEnd gridColumnSpan gridColumnStart fontWeight lineClamp lineHeight opacity order orphans scale tabSize widows zIndex zoom fillOpacity floodOpacity stopOpacity strokeDasharray strokeDashoffset strokeMiterlimit strokeOpacity strokeWidth MozAnimationIterationCount MozBoxFlex MozBoxFlexGroup MozLineClamp msAnimationIterationCount msFlex msZoom msFlexGrow msFlexNegative msFlexOrder msFlexPositive msFlexShrink msGridColumn msGridColumnSpan msGridRow msGridRowSpan WebkitAnimationIterationCount WebkitBoxFlex WebKitBoxFlexGroup WebkitBoxOrdinalGroup WebkitColumnCount WebkitColumns WebkitFlex WebkitFlexGrow WebkitFlexPositive WebkitFlexShrink WebkitLineClamp".split(" "));function Nf(l,t,a){var e=t.indexOf("--")===0;a==null||typeof a=="boolean"||a===""?e?l.setProperty(t,""):t==="float"?l.cssFloat="":l[t]="":e?l.setProperty(t,a):typeof a!="number"||a===0||Xr.has(t)?t==="float"?l.cssFloat=a:l[t]=(""+a).trim():l[t]=a+"px"}function Uf(l,t,a){if(t!=null&&typeof t!="object")throw Error(s(62));if(l=l.style,a!=null){for(var e in a)!a.hasOwnProperty(e)||t!=null&&t.hasOwnProperty(e)||(e.indexOf("--")===0?l.setProperty(e,""):e==="float"?l.cssFloat="":l[e]="");for(var u in t)e=t[u],t.hasOwnProperty(u)&&a[u]!==e&&Nf(l,u,e)}else for(var n in t)t.hasOwnProperty(n)&&Nf(l,n,t[n])}function tc(l){if(l.indexOf("-")===-1)return!1;switch(l){case"annotation-xml":case"color-profile":case"font-face":case"font-face-src":case"font-face-uri":case"font-face-format":case"font-face-name":case"missing-glyph":return!1;default:return!0}}var Qr=new Map([["acceptCharset","accept-charset"],["htmlFor","for"],["httpEquiv","http-equiv"],["crossOrigin","crossorigin"],["accentHeight","accent-height"],["alignmentBaseline","alignment-baseline"],["arabicForm","arabic-form"],["baselineShift","baseline-shift"],["capHeight","cap-height"],["clipPath","clip-path"],["clipRule","clip-rule"],["colorInterpolation","color-interpolation"],["colorInterpolationFilters","color-interpolation-filters"],["colorProfile","color-profile"],["colorRendering","color-rendering"],["dominantBaseline","dominant-baseline"],["enableBackground","enable-background"],["fillOpacity","fill-opacity"],["fillRule","fill-rule"],["floodColor","flood-color"],["floodOpacity","flood-opacity"],["fontFamily","font-family"],["fontSize","font-size"],["fontSizeAdjust","font-size-adjust"],["fontStretch","font-stretch"],["fontStyle","font-style"],["fontVariant","font-variant"],["fontWeight","font-weight"],["glyphName","glyph-name"],["glyphOrientationHorizontal","glyph-orientation-horizontal"],["glyphOrientationVertical","glyph-orientation-vertical"],["horizAdvX","horiz-adv-x"],["horizOriginX","horiz-origin-x"],["imageRendering","image-rendering"],["letterSpacing","letter-spacing"],["lightingColor","lighting-color"],["markerEnd","marker-end"],["markerMid","marker-mid"],["markerStart","marker-start"],["overlinePosition","overline-position"],["overlineThickness","overline-thickness"],["paintOrder","paint-order"],["panose-1","panose-1"],["pointerEvents","pointer-events"],["renderingIntent","rendering-intent"],["shapeRendering","shape-rendering"],["stopColor","stop-color"],["stopOpacity","stop-opacity"],["strikethroughPosition","strikethrough-position"],["strikethroughThickness","strikethrough-thickness"],["strokeDasharray","stroke-dasharray"],["strokeDashoffset","stroke-dashoffset"],["strokeLinecap","stroke-linecap"],["strokeLinejoin","stroke-linejoin"],["strokeMiterlimit","stroke-miterlimit"],["strokeOpacity","stroke-opacity"],["strokeWidth","stroke-width"],["textAnchor","text-anchor"],["textDecoration","text-decoration"],["textRendering","text-rendering"],["transformOrigin","transform-origin"],["underlinePosition","underline-position"],["underlineThickness","underline-thickness"],["unicodeBidi","unicode-bidi"],["unicodeRange","unicode-range"],["unitsPerEm","units-per-em"],["vAlphabetic","v-alphabetic"],["vHanging","v-hanging"],["vIdeographic","v-ideographic"],["vMathematical","v-mathematical"],["vectorEffect","vector-effect"],["vertAdvY","vert-adv-y"],["vertOriginX","vert-origin-x"],["vertOriginY","vert-origin-y"],["wordSpacing","word-spacing"],["writingMode","writing-mode"],["xmlnsXlink","xmlns:xlink"],["xHeight","x-height"]]),Zr=/^[\u0000-\u001F ]*j[\r\n\t]*a[\r\n\t]*v[\r\n\t]*a[\r\n\t]*s[\r\n\t]*c[\r\n\t]*r[\r\n\t]*i[\r\n\t]*p[\r\n\t]*t[\r\n\t]*:/i;function ju(l){return Zr.test(""+l)?"javascript:throw new Error('React has blocked a javascript: URL as a security precaution.')":l}var ac=null;function ec(l){return l=l.target||l.srcElement||window,l.correspondingUseElement&&(l=l.correspondingUseElement),l.nodeType===3?l.parentNode:l}var Ka=null,Ja=null;function Rf(l){var t=Ga(l);if(t&&(l=t.stateNode)){var a=l[wl]||null;l:switch(l=t.stateNode,t.type){case"input":if(In(l,a.value,a.defaultValue,a.defaultValue,a.checked,a.defaultChecked,a.type,a.name),t=a.name,a.type==="radio"&&t!=null){for(a=l;a.parentNode;)a=a.parentNode;for(a=a.querySelectorAll('input[name="'+ot(""+t)+'"][type="radio"]'),t=0;t<a.length;t++){var e=a[t];if(e!==l&&e.form===l.form){var u=e[wl]||null;if(!u)throw Error(s(90));In(e,u.value,u.defaultValue,u.defaultValue,u.checked,u.defaultChecked,u.type,u.name)}}for(t=0;t<a.length;t++)e=a[t],e.form===l.form&&xf(e)}break l;case"textarea":Mf(l,a.value,a.defaultValue);break l;case"select":t=a.value,t!=null&&Va(l,!!a.multiple,t,!1)}}}var uc=!1;function jf(l,t,a){if(uc)return l(t,a);uc=!0;try{var e=l(t);return e}finally{if(uc=!1,(Ka!==null||Ja!==null)&&(Sn(),Ka&&(t=Ka,l=Ja,Ja=Ka=null,Rf(t),l)))for(t=0;t<l.length;t++)Rf(l[t])}}function Me(l,t){var a=l.stateNode;if(a===null)return null;var e=a[wl]||null;if(e===null)return null;a=e[t];l:switch(t){case"onClick":case"onClickCapture":case"onDoubleClick":case"onDoubleClickCapture":case"onMouseDown":case"onMouseDownCapture":case"onMouseMove":case"onMouseMoveCapture":case"onMouseUp":case"onMouseUpCapture":case"onMouseEnter":(e=!e.disabled)||(l=l.type,e=!(l==="button"||l==="input"||l==="select"||l==="textarea")),l=!e;break l;default:l=!1}if(l)return null;if(a&&typeof a!="function")throw Error(s(231,t,typeof a));return a}var Ut=!(typeof window>"u"||typeof window.document>"u"||typeof window.document.createElement>"u"),nc=!1;if(Ut)try{var De={};Object.defineProperty(De,"passive",{get:function(){nc=!0}}),window.addEventListener("test",De,De),window.removeEventListener("test",De,De)}catch{nc=!1}var kt=null,cc=null,Hu=null;function Hf(){if(Hu)return Hu;var l,t=cc,a=t.length,e,u="value"in kt?kt.value:kt.textContent,n=u.length;for(l=0;l<a&&t[l]===u[l];l++);var c=a-l;for(e=1;e<=c&&t[a-e]===u[n-e];e++);return Hu=u.slice(l,1<e?1-e:void 0)}function qu(l){var t=l.keyCode;return"charCode"in l?(l=l.charCode,l===0&&t===13&&(l=13)):l=t,l===10&&(l=13),32<=l||l===13?l:0}function Yu(){return!0}function qf(){return!1}function Wl(l){function t(a,e,u,n,c){this._reactName=a,this._targetInst=u,this.type=e,this.nativeEvent=n,this.target=c,this.currentTarget=null;for(var i in l)l.hasOwnProperty(i)&&(a=l[i],this[i]=a?a(n):n[i]);return this.isDefaultPrevented=(n.defaultPrevented!=null?n.defaultPrevented:n.returnValue===!1)?Yu:qf,this.isPropagationStopped=qf,this}return D(t.prototype,{preventDefault:function(){this.defaultPrevented=!0;var a=this.nativeEvent;a&&(a.preventDefault?a.preventDefault():typeof a.returnValue!="unknown"&&(a.returnValue=!1),this.isDefaultPrevented=Yu)},stopPropagation:function(){var a=this.nativeEvent;a&&(a.stopPropagation?a.stopPropagation():typeof a.cancelBubble!="unknown"&&(a.cancelBubble=!0),this.isPropagationStopped=Yu)},persist:function(){},isPersistent:Yu}),t}var Ta={eventPhase:0,bubbles:0,cancelable:0,timeStamp:function(l){return l.timeStamp||Date.now()},defaultPrevented:0,isTrusted:0},Bu=Wl(Ta),Ne=D({},Ta,{view:0,detail:0}),Vr=Wl(Ne),ic,fc,Ue,Cu=D({},Ne,{screenX:0,screenY:0,clientX:0,clientY:0,pageX:0,pageY:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,getModifierState:oc,button:0,buttons:0,relatedTarget:function(l){return l.relatedTarget===void 0?l.fromElement===l.srcElement?l.toElement:l.fromElement:l.relatedTarget},movementX:function(l){return"movementX"in l?l.movementX:(l!==Ue&&(Ue&&l.type==="mousemove"?(ic=l.screenX-Ue.screenX,fc=l.screenY-Ue.screenY):fc=ic=0,Ue=l),ic)},movementY:function(l){return"movementY"in l?l.movementY:fc}}),Yf=Wl(Cu),Lr=D({},Cu,{dataTransfer:0}),Kr=Wl(Lr),Jr=D({},Ne,{relatedTarget:0}),sc=Wl(Jr),wr=D({},Ta,{animationName:0,elapsedTime:0,pseudoElement:0}),Wr=Wl(wr),kr=D({},Ta,{clipboardData:function(l){return"clipboardData"in l?l.clipboardData:window.clipboardData}}),$r=Wl(kr),Fr=D({},Ta,{data:0}),Bf=Wl(Fr),Pr={Esc:"Escape",Spacebar:" ",Left:"ArrowLeft",Up:"ArrowUp",Right:"ArrowRight",Down:"ArrowDown",Del:"Delete",Win:"OS",Menu:"ContextMenu",Apps:"ContextMenu",Scroll:"ScrollLock",MozPrintableKey:"Unidentified"},Ir={8:"Backspace",9:"Tab",12:"Clear",13:"Enter",16:"Shift",17:"Control",18:"Alt",19:"Pause",20:"CapsLock",27:"Escape",32:" ",33:"PageUp",34:"PageDown",35:"End",36:"Home",37:"ArrowLeft",38:"ArrowUp",39:"ArrowRight",40:"ArrowDown",45:"Insert",46:"Delete",112:"F1",113:"F2",114:"F3",115:"F4",116:"F5",117:"F6",118:"F7",119:"F8",120:"F9",121:"F10",122:"F11",123:"F12",144:"NumLock",145:"ScrollLock",224:"Meta"},l0={Alt:"altKey",Control:"ctrlKey",Meta:"metaKey",Shift:"shiftKey"};function t0(l){var t=this.nativeEvent;return t.getModifierState?t.getModifierState(l):(l=l0[l])?!!t[l]:!1}function oc(){return t0}var a0=D({},Ne,{key:function(l){if(l.key){var t=Pr[l.key]||l.key;if(t!=="Unidentified")return t}return l.type==="keypress"?(l=qu(l),l===13?"Enter":String.fromCharCode(l)):l.type==="keydown"||l.type==="keyup"?Ir[l.keyCode]||"Unidentified":""},code:0,location:0,ctrlKey:0,shiftKey:0,altKey:0,metaKey:0,repeat:0,locale:0,getModifierState:oc,charCode:function(l){return l.type==="keypress"?qu(l):0},keyCode:function(l){return l.type==="keydown"||l.type==="keyup"?l.keyCode:0},which:function(l){return l.type==="keypress"?qu(l):l.type==="keydown"||l.type==="keyup"?l.keyCode:0}}),e0=Wl(a0),u0=D({},Cu,{pointerId:0,width:0,height:0,pressure:0,tangentialPressure:0,tiltX:0,tiltY:0,twist:0,pointerType:0,isPrimary:0}),Cf=Wl(u0),n0=D({},Ne,{touches:0,targetTouches:0,changedTouches:0,altKey:0,metaKey:0,ctrlKey:0,shiftKey:0,getModifierState:oc}),c0=Wl(n0),i0=D({},Ta,{propertyName:0,elapsedTime:0,pseudoElement:0}),f0=Wl(i0),s0=D({},Cu,{deltaX:function(l){return"deltaX"in l?l.deltaX:"wheelDeltaX"in l?-l.wheelDeltaX:0},deltaY:function(l){return"deltaY"in l?l.deltaY:"wheelDeltaY"in l?-l.wheelDeltaY:"wheelDelta"in l?-l.wheelDelta:0},deltaZ:0,deltaMode:0}),o0=Wl(s0),d0=D({},Ta,{newState:0,oldState:0}),r0=Wl(d0),v0=[9,13,27,32],dc=Ut&&"CompositionEvent"in window,Re=null;Ut&&"documentMode"in document&&(Re=document.documentMode);var y0=Ut&&"TextEvent"in window&&!Re,Gf=Ut&&(!dc||Re&&8<Re&&11>=Re),Xf=" ",Qf=!1;function Zf(l,t){switch(l){case"keyup":return v0.indexOf(t.keyCode)!==-1;case"keydown":return t.keyCode!==229;case"keypress":case"mousedown":case"focusout":return!0;default:return!1}}function Vf(l){return l=l.detail,typeof l=="object"&&"data"in l?l.data:null}var wa=!1;function h0(l,t){switch(l){case"compositionend":return Vf(t);case"keypress":return t.which!==32?null:(Qf=!0,Xf);case"textInput":return l=t.data,l===Xf&&Qf?null:l;default:return null}}function m0(l,t){if(wa)return l==="compositionend"||!dc&&Zf(l,t)?(l=Hf(),Hu=cc=kt=null,wa=!1,l):null;switch(l){case"paste":return null;case"keypress":if(!(t.ctrlKey||t.altKey||t.metaKey)||t.ctrlKey&&t.altKey){if(t.char&&1<t.char.length)return t.char;if(t.which)return String.fromCharCode(t.which)}return null;case"compositionend":return Gf&&t.locale!=="ko"?null:t.data;default:return null}}var g0={color:!0,date:!0,datetime:!0,"datetime-local":!0,email:!0,month:!0,number:!0,password:!0,range:!0,search:!0,tel:!0,text:!0,time:!0,url:!0,week:!0};function Lf(l){var t=l&&l.nodeName&&l.nodeName.toLowerCase();return t==="input"?!!g0[l.type]:t==="textarea"}function Kf(l,t,a,e){Ka?Ja?Ja.push(e):Ja=[e]:Ka=e,t=pn(t,"onChange"),0<t.length&&(a=new Bu("onChange","change",null,a,e),l.push({event:a,listeners:t}))}var je=null,He=null;function b0(l){xd(l,0)}function Gu(l){var t=_e(l);if(xf(t))return l}function Jf(l,t){if(l==="change")return t}var wf=!1;if(Ut){var rc;if(Ut){var vc="oninput"in document;if(!vc){var Wf=document.createElement("div");Wf.setAttribute("oninput","return;"),vc=typeof Wf.oninput=="function"}rc=vc}else rc=!1;wf=rc&&(!document.documentMode||9<document.documentMode)}function kf(){je&&(je.detachEvent("onpropertychange",$f),He=je=null)}function $f(l){if(l.propertyName==="value"&&Gu(He)){var t=[];Kf(t,He,l,ec(l)),jf(b0,t)}}function S0(l,t,a){l==="focusin"?(kf(),je=t,He=a,je.attachEvent("onpropertychange",$f)):l==="focusout"&&kf()}function z0(l){if(l==="selectionchange"||l==="keyup"||l==="keydown")return Gu(He)}function T0(l,t){if(l==="click")return Gu(t)}function E0(l,t){if(l==="input"||l==="change")return Gu(t)}function A0(l,t){return l===t&&(l!==0||1/l===1/t)||l!==l&&t!==t}var tt=typeof Object.is=="function"?Object.is:A0;function qe(l,t){if(tt(l,t))return!0;if(typeof l!="object"||l===null||typeof t!="object"||t===null)return!1;var a=Object.keys(l),e=Object.keys(t);if(a.length!==e.length)return!1;for(e=0;e<a.length;e++){var u=a[e];if(!Zn.call(t,u)||!tt(l[u],t[u]))return!1}return!0}function Ff(l){for(;l&&l.firstChild;)l=l.firstChild;return l}function Pf(l,t){var a=Ff(l);l=0;for(var e;a;){if(a.nodeType===3){if(e=l+a.textContent.length,l<=t&&e>=t)return{node:a,offset:t-l};l=e}l:{for(;a;){if(a.nextSibling){a=a.nextSibling;break l}a=a.parentNode}a=void 0}a=Ff(a)}}function If(l,t){return l&&t?l===t?!0:l&&l.nodeType===3?!1:t&&t.nodeType===3?If(l,t.parentNode):"contains"in l?l.contains(t):l.compareDocumentPosition?!!(l.compareDocumentPosition(t)&16):!1:!1}function ls(l){l=l!=null&&l.ownerDocument!=null&&l.ownerDocument.defaultView!=null?l.ownerDocument.defaultView:window;for(var t=Ru(l.document);t instanceof l.HTMLIFrameElement;){try{var a=typeof t.contentWindow.location.href=="string"}catch{a=!1}if(a)l=t.contentWindow;else break;t=Ru(l.document)}return t}function yc(l){var t=l&&l.nodeName&&l.nodeName.toLowerCase();return t&&(t==="input"&&(l.type==="text"||l.type==="search"||l.type==="tel"||l.type==="url"||l.type==="password")||t==="textarea"||l.contentEditable==="true")}var O0=Ut&&"documentMode"in document&&11>=document.documentMode,Wa=null,hc=null,Ye=null,mc=!1;function ts(l,t,a){var e=a.window===a?a.document:a.nodeType===9?a:a.ownerDocument;mc||Wa==null||Wa!==Ru(e)||(e=Wa,"selectionStart"in e&&yc(e)?e={start:e.selectionStart,end:e.selectionEnd}:(e=(e.ownerDocument&&e.ownerDocument.defaultView||window).getSelection(),e={anchorNode:e.anchorNode,anchorOffset:e.anchorOffset,focusNode:e.focusNode,focusOffset:e.focusOffset}),Ye&&qe(Ye,e)||(Ye=e,e=pn(hc,"onSelect"),0<e.length&&(t=new Bu("onSelect","select",null,t,a),l.push({event:t,listeners:e}),t.target=Wa)))}function Ea(l,t){var a={};return a[l.toLowerCase()]=t.toLowerCase(),a["Webkit"+l]="webkit"+t,a["Moz"+l]="moz"+t,a}var ka={animationend:Ea("Animation","AnimationEnd"),animationiteration:Ea("Animation","AnimationIteration"),animationstart:Ea("Animation","AnimationStart"),transitionrun:Ea("Transition","TransitionRun"),transitionstart:Ea("Transition","TransitionStart"),transitioncancel:Ea("Transition","TransitionCancel"),transitionend:Ea("Transition","TransitionEnd")},gc={},as={};Ut&&(as=document.createElement("div").style,"AnimationEvent"in window||(delete ka.animationend.animation,delete ka.animationiteration.animation,delete ka.animationstart.animation),"TransitionEvent"in window||delete ka.transitionend.transition);function Aa(l){if(gc[l])return gc[l];if(!ka[l])return l;var t=ka[l],a;for(a in t)if(t.hasOwnProperty(a)&&a in as)return gc[l]=t[a];return l}var es=Aa("animationend"),us=Aa("animationiteration"),ns=Aa("animationstart"),p0=Aa("transitionrun"),x0=Aa("transitionstart"),_0=Aa("transitioncancel"),cs=Aa("transitionend"),is=new Map,bc="abort auxClick beforeToggle cancel canPlay canPlayThrough click close contextMenu copy cut drag dragEnd dragEnter dragExit dragLeave dragOver dragStart drop durationChange emptied encrypted ended error gotPointerCapture input invalid keyDown keyPress keyUp load loadedData loadedMetadata loadStart lostPointerCapture mouseDown mouseMove mouseOut mouseOver mouseUp paste pause play playing pointerCancel pointerDown pointerMove pointerOut pointerOver pointerUp progress rateChange reset resize seeked seeking stalled submit suspend timeUpdate touchCancel touchEnd touchStart volumeChange scroll toggle touchMove waiting wheel".split(" ");bc.push("scrollEnd");function St(l,t){is.set(l,t),za(t,[l])}var fs=new WeakMap;function dt(l,t){if(typeof l=="object"&&l!==null){var a=fs.get(l);return a!==void 0?a:(t={value:l,source:t,stack:Of(t)},fs.set(l,t),t)}return{value:l,source:t,stack:Of(t)}}var rt=[],$a=0,Sc=0;function Xu(){for(var l=$a,t=Sc=$a=0;t<l;){var a=rt[t];rt[t++]=null;var e=rt[t];rt[t++]=null;var u=rt[t];rt[t++]=null;var n=rt[t];if(rt[t++]=null,e!==null&&u!==null){var c=e.pending;c===null?u.next=u:(u.next=c.next,c.next=u),e.pending=u}n!==0&&ss(a,u,n)}}function Qu(l,t,a,e){rt[$a++]=l,rt[$a++]=t,rt[$a++]=a,rt[$a++]=e,Sc|=e,l.lanes|=e,l=l.alternate,l!==null&&(l.lanes|=e)}function zc(l,t,a,e){return Qu(l,t,a,e),Zu(l)}function Fa(l,t){return Qu(l,null,null,t),Zu(l)}function ss(l,t,a){l.lanes|=a;var e=l.alternate;e!==null&&(e.lanes|=a);for(var u=!1,n=l.return;n!==null;)n.childLanes|=a,e=n.alternate,e!==null&&(e.childLanes|=a),n.tag===22&&(l=n.stateNode,l===null||l._visibility&1||(u=!0)),l=n,n=n.return;return l.tag===3?(n=l.stateNode,u&&t!==null&&(u=31-lt(a),l=n.hiddenUpdates,e=l[u],e===null?l[u]=[t]:e.push(t),t.lane=a|536870912),n):null}function Zu(l){if(50<iu)throw iu=0,xi=null,Error(s(185));for(var t=l.return;t!==null;)l=t,t=l.return;return l.tag===3?l.stateNode:null}var Pa={};function M0(l,t,a,e){this.tag=l,this.key=a,this.sibling=this.child=this.return=this.stateNode=this.type=this.elementType=null,this.index=0,this.refCleanup=this.ref=null,this.pendingProps=t,this.dependencies=this.memoizedState=this.updateQueue=this.memoizedProps=null,this.mode=e,this.subtreeFlags=this.flags=0,this.deletions=null,this.childLanes=this.lanes=0,this.alternate=null}function at(l,t,a,e){return new M0(l,t,a,e)}function Tc(l){return l=l.prototype,!(!l||!l.isReactComponent)}function Rt(l,t){var a=l.alternate;return a===null?(a=at(l.tag,t,l.key,l.mode),a.elementType=l.elementType,a.type=l.type,a.stateNode=l.stateNode,a.alternate=l,l.alternate=a):(a.pendingProps=t,a.type=l.type,a.flags=0,a.subtreeFlags=0,a.deletions=null),a.flags=l.flags&65011712,a.childLanes=l.childLanes,a.lanes=l.lanes,a.child=l.child,a.memoizedProps=l.memoizedProps,a.memoizedState=l.memoizedState,a.updateQueue=l.updateQueue,t=l.dependencies,a.dependencies=t===null?null:{lanes:t.lanes,firstContext:t.firstContext},a.sibling=l.sibling,a.index=l.index,a.ref=l.ref,a.refCleanup=l.refCleanup,a}function os(l,t){l.flags&=65011714;var a=l.alternate;return a===null?(l.childLanes=0,l.lanes=t,l.child=null,l.subtreeFlags=0,l.memoizedProps=null,l.memoizedState=null,l.updateQueue=null,l.dependencies=null,l.stateNode=null):(l.childLanes=a.childLanes,l.lanes=a.lanes,l.child=a.child,l.subtreeFlags=0,l.deletions=null,l.memoizedProps=a.memoizedProps,l.memoizedState=a.memoizedState,l.updateQueue=a.updateQueue,l.type=a.type,t=a.dependencies,l.dependencies=t===null?null:{lanes:t.lanes,firstContext:t.firstContext}),l}function Vu(l,t,a,e,u,n){var c=0;if(e=l,typeof l=="function")Tc(l)&&(c=1);else if(typeof l=="string")c=Nv(l,a,G.current)?26:l==="html"||l==="head"||l==="body"?27:5;else l:switch(l){case ft:return l=at(31,a,t,u),l.elementType=ft,l.lanes=n,l;case ul:return Oa(a.children,u,n,t);case pl:c=8,u|=24;break;case Dl:return l=at(12,a,t,u|2),l.elementType=Dl,l.lanes=n,l;case P:return l=at(13,a,t,u),l.elementType=P,l.lanes=n,l;case Yl:return l=at(19,a,t,u),l.elementType=Yl,l.lanes=n,l;default:if(typeof l=="object"&&l!==null)switch(l.$$typeof){case hl:case dl:c=10;break l;case ql:c=9;break l;case Vl:c=11;break l;case Ll:c=14;break l;case Kl:c=16,e=null;break l}c=29,a=Error(s(130,l===null?"null":typeof l,"")),e=null}return t=at(c,a,t,u),t.elementType=l,t.type=e,t.lanes=n,t}function Oa(l,t,a,e){return l=at(7,l,e,t),l.lanes=a,l}function Ec(l,t,a){return l=at(6,l,null,t),l.lanes=a,l}function Ac(l,t,a){return t=at(4,l.children!==null?l.children:[],l.key,t),t.lanes=a,t.stateNode={containerInfo:l.containerInfo,pendingChildren:null,implementation:l.implementation},t}var Ia=[],le=0,Lu=null,Ku=0,vt=[],yt=0,pa=null,jt=1,Ht="";function xa(l,t){Ia[le++]=Ku,Ia[le++]=Lu,Lu=l,Ku=t}function ds(l,t,a){vt[yt++]=jt,vt[yt++]=Ht,vt[yt++]=pa,pa=l;var e=jt;l=Ht;var u=32-lt(e)-1;e&=~(1<<u),a+=1;var n=32-lt(t)+u;if(30<n){var c=u-u%5;n=(e&(1<<c)-1).toString(32),e>>=c,u-=c,jt=1<<32-lt(t)+u|a<<u|e,Ht=n+l}else jt=1<<n|a<<u|e,Ht=l}function Oc(l){l.return!==null&&(xa(l,1),ds(l,1,0))}function pc(l){for(;l===Lu;)Lu=Ia[--le],Ia[le]=null,Ku=Ia[--le],Ia[le]=null;for(;l===pa;)pa=vt[--yt],vt[yt]=null,Ht=vt[--yt],vt[yt]=null,jt=vt[--yt],vt[yt]=null}var Jl=null,Sl=null,el=!1,_a=null,At=!1,xc=Error(s(519));function Ma(l){var t=Error(s(418,""));throw Ge(dt(t,l)),xc}function rs(l){var t=l.stateNode,a=l.type,e=l.memoizedProps;switch(t[Xl]=l,t[wl]=e,a){case"dialog":F("cancel",t),F("close",t);break;case"iframe":case"object":case"embed":F("load",t);break;case"video":case"audio":for(a=0;a<su.length;a++)F(su[a],t);break;case"source":F("error",t);break;case"img":case"image":case"link":F("error",t),F("load",t);break;case"details":F("toggle",t);break;case"input":F("invalid",t),_f(t,e.value,e.defaultValue,e.checked,e.defaultChecked,e.type,e.name,!0),Uu(t);break;case"select":F("invalid",t);break;case"textarea":F("invalid",t),Df(t,e.value,e.defaultValue,e.children),Uu(t)}a=e.children,typeof a!="string"&&typeof a!="number"&&typeof a!="bigint"||t.textContent===""+a||e.suppressHydrationWarning===!0||Nd(t.textContent,a)?(e.popover!=null&&(F("beforetoggle",t),F("toggle",t)),e.onScroll!=null&&F("scroll",t),e.onScrollEnd!=null&&F("scrollend",t),e.onClick!=null&&(t.onclick=xn),t=!0):t=!1,t||Ma(l)}function vs(l){for(Jl=l.return;Jl;)switch(Jl.tag){case 5:case 13:At=!1;return;case 27:case 3:At=!0;return;default:Jl=Jl.return}}function Be(l){if(l!==Jl)return!1;if(!el)return vs(l),el=!0,!1;var t=l.tag,a;if((a=t!==3&&t!==27)&&((a=t===5)&&(a=l.type,a=!(a!=="form"&&a!=="button")||Zi(l.type,l.memoizedProps)),a=!a),a&&Sl&&Ma(l),vs(l),t===13){if(l=l.memoizedState,l=l!==null?l.dehydrated:null,!l)throw Error(s(317));l:{for(l=l.nextSibling,t=0;l;){if(l.nodeType===8)if(a=l.data,a==="/$"){if(t===0){Sl=Tt(l.nextSibling);break l}t--}else a!=="$"&&a!=="$!"&&a!=="$?"||t++;l=l.nextSibling}Sl=null}}else t===27?(t=Sl,da(l.type)?(l=Ji,Ji=null,Sl=l):Sl=t):Sl=Jl?Tt(l.stateNode.nextSibling):null;return!0}function Ce(){Sl=Jl=null,el=!1}function ys(){var l=_a;return l!==null&&(Fl===null?Fl=l:Fl.push.apply(Fl,l),_a=null),l}function Ge(l){_a===null?_a=[l]:_a.push(l)}var _c=p(null),Da=null,qt=null;function $t(l,t,a){U(_c,t._currentValue),t._currentValue=a}function Yt(l){l._currentValue=_c.current,H(_c)}function Mc(l,t,a){for(;l!==null;){var e=l.alternate;if((l.childLanes&t)!==t?(l.childLanes|=t,e!==null&&(e.childLanes|=t)):e!==null&&(e.childLanes&t)!==t&&(e.childLanes|=t),l===a)break;l=l.return}}function Dc(l,t,a,e){var u=l.child;for(u!==null&&(u.return=l);u!==null;){var n=u.dependencies;if(n!==null){var c=u.child;n=n.firstContext;l:for(;n!==null;){var i=n;n=u;for(var f=0;f<t.length;f++)if(i.context===t[f]){n.lanes|=a,i=n.alternate,i!==null&&(i.lanes|=a),Mc(n.return,a,l),e||(c=null);break l}n=i.next}}else if(u.tag===18){if(c=u.return,c===null)throw Error(s(341));c.lanes|=a,n=c.alternate,n!==null&&(n.lanes|=a),Mc(c,a,l),c=null}else c=u.child;if(c!==null)c.return=u;else for(c=u;c!==null;){if(c===l){c=null;break}if(u=c.sibling,u!==null){u.return=c.return,c=u;break}c=c.return}u=c}}function Xe(l,t,a,e){l=null;for(var u=t,n=!1;u!==null;){if(!n){if((u.flags&524288)!==0)n=!0;else if((u.flags&262144)!==0)break}if(u.tag===10){var c=u.alternate;if(c===null)throw Error(s(387));if(c=c.memoizedProps,c!==null){var i=u.type;tt(u.pendingProps.value,c.value)||(l!==null?l.push(i):l=[i])}}else if(u===Pl.current){if(c=u.alternate,c===null)throw Error(s(387));c.memoizedState.memoizedState!==u.memoizedState.memoizedState&&(l!==null?l.push(hu):l=[hu])}u=u.return}l!==null&&Dc(t,l,a,e),t.flags|=262144}function Ju(l){for(l=l.firstContext;l!==null;){if(!tt(l.context._currentValue,l.memoizedValue))return!0;l=l.next}return!1}function Na(l){Da=l,qt=null,l=l.dependencies,l!==null&&(l.firstContext=null)}function Ql(l){return hs(Da,l)}function wu(l,t){return Da===null&&Na(l),hs(l,t)}function hs(l,t){var a=t._currentValue;if(t={context:t,memoizedValue:a,next:null},qt===null){if(l===null)throw Error(s(308));qt=t,l.dependencies={lanes:0,firstContext:t},l.flags|=524288}else qt=qt.next=t;return a}var D0=typeof AbortController<"u"?AbortController:function(){var l=[],t=this.signal={aborted:!1,addEventListener:function(a,e){l.push(e)}};this.abort=function(){t.aborted=!0,l.forEach(function(a){return a()})}},N0=v.unstable_scheduleCallback,U0=v.unstable_NormalPriority,xl={$$typeof:dl,Consumer:null,Provider:null,_currentValue:null,_currentValue2:null,_threadCount:0};function Nc(){return{controller:new D0,data:new Map,refCount:0}}function Qe(l){l.refCount--,l.refCount===0&&N0(U0,function(){l.controller.abort()})}var Ze=null,Uc=0,te=0,ae=null;function R0(l,t){if(Ze===null){var a=Ze=[];Uc=0,te=ji(),ae={status:"pending",value:void 0,then:function(e){a.push(e)}}}return Uc++,t.then(ms,ms),t}function ms(){if(--Uc===0&&Ze!==null){ae!==null&&(ae.status="fulfilled");var l=Ze;Ze=null,te=0,ae=null;for(var t=0;t<l.length;t++)(0,l[t])()}}function j0(l,t){var a=[],e={status:"pending",value:null,reason:null,then:function(u){a.push(u)}};return l.then(function(){e.status="fulfilled",e.value=t;for(var u=0;u<a.length;u++)(0,a[u])(t)},function(u){for(e.status="rejected",e.reason=u,u=0;u<a.length;u++)(0,a[u])(void 0)}),e}var gs=S.S;S.S=function(l,t){typeof t=="object"&&t!==null&&typeof t.then=="function"&&R0(l,t),gs!==null&&gs(l,t)};var Ua=p(null);function Rc(){var l=Ua.current;return l!==null?l:yl.pooledCache}function Wu(l,t){t===null?U(Ua,Ua.current):U(Ua,t.pool)}function bs(){var l=Rc();return l===null?null:{parent:xl._currentValue,pool:l}}var Ve=Error(s(460)),Ss=Error(s(474)),ku=Error(s(542)),jc={then:function(){}};function zs(l){return l=l.status,l==="fulfilled"||l==="rejected"}function $u(){}function Ts(l,t,a){switch(a=l[a],a===void 0?l.push(t):a!==t&&(t.then($u,$u),t=a),t.status){case"fulfilled":return t.value;case"rejected":throw l=t.reason,As(l),l;default:if(typeof t.status=="string")t.then($u,$u);else{if(l=yl,l!==null&&100<l.shellSuspendCounter)throw Error(s(482));l=t,l.status="pending",l.then(function(e){if(t.status==="pending"){var u=t;u.status="fulfilled",u.value=e}},function(e){if(t.status==="pending"){var u=t;u.status="rejected",u.reason=e}})}switch(t.status){case"fulfilled":return t.value;case"rejected":throw l=t.reason,As(l),l}throw Le=t,Ve}}var Le=null;function Es(){if(Le===null)throw Error(s(459));var l=Le;return Le=null,l}function As(l){if(l===Ve||l===ku)throw Error(s(483))}var Ft=!1;function Hc(l){l.updateQueue={baseState:l.memoizedState,firstBaseUpdate:null,lastBaseUpdate:null,shared:{pending:null,lanes:0,hiddenCallbacks:null},callbacks:null}}function qc(l,t){l=l.updateQueue,t.updateQueue===l&&(t.updateQueue={baseState:l.baseState,firstBaseUpdate:l.firstBaseUpdate,lastBaseUpdate:l.lastBaseUpdate,shared:l.shared,callbacks:null})}function Pt(l){return{lane:l,tag:0,payload:null,callback:null,next:null}}function It(l,t,a){var e=l.updateQueue;if(e===null)return null;if(e=e.shared,(nl&2)!==0){var u=e.pending;return u===null?t.next=t:(t.next=u.next,u.next=t),e.pending=t,t=Zu(l),ss(l,null,a),t}return Qu(l,e,t,a),Zu(l)}function Ke(l,t,a){if(t=t.updateQueue,t!==null&&(t=t.shared,(a&4194048)!==0)){var e=t.lanes;e&=l.pendingLanes,a|=e,t.lanes=a,mf(l,a)}}function Yc(l,t){var a=l.updateQueue,e=l.alternate;if(e!==null&&(e=e.updateQueue,a===e)){var u=null,n=null;if(a=a.firstBaseUpdate,a!==null){do{var c={lane:a.lane,tag:a.tag,payload:a.payload,callback:null,next:null};n===null?u=n=c:n=n.next=c,a=a.next}while(a!==null);n===null?u=n=t:n=n.next=t}else u=n=t;a={baseState:e.baseState,firstBaseUpdate:u,lastBaseUpdate:n,shared:e.shared,callbacks:e.callbacks},l.updateQueue=a;return}l=a.lastBaseUpdate,l===null?a.firstBaseUpdate=t:l.next=t,a.lastBaseUpdate=t}var Bc=!1;function Je(){if(Bc){var l=ae;if(l!==null)throw l}}function we(l,t,a,e){Bc=!1;var u=l.updateQueue;Ft=!1;var n=u.firstBaseUpdate,c=u.lastBaseUpdate,i=u.shared.pending;if(i!==null){u.shared.pending=null;var f=i,m=f.next;f.next=null,c===null?n=m:c.next=m,c=f;var z=l.alternate;z!==null&&(z=z.updateQueue,i=z.lastBaseUpdate,i!==c&&(i===null?z.firstBaseUpdate=m:i.next=m,z.lastBaseUpdate=f))}if(n!==null){var A=u.baseState;c=0,z=m=f=null,i=n;do{var g=i.lane&-536870913,b=g!==i.lane;if(b?(I&g)===g:(e&g)===g){g!==0&&g===te&&(Bc=!0),z!==null&&(z=z.next={lane:0,tag:i.tag,payload:i.payload,callback:null,next:null});l:{var L=l,X=i;g=t;var ol=a;switch(X.tag){case 1:if(L=X.payload,typeof L=="function"){A=L.call(ol,A,g);break l}A=L;break l;case 3:L.flags=L.flags&-65537|128;case 0:if(L=X.payload,g=typeof L=="function"?L.call(ol,A,g):L,g==null)break l;A=D({},A,g);break l;case 2:Ft=!0}}g=i.callback,g!==null&&(l.flags|=64,b&&(l.flags|=8192),b=u.callbacks,b===null?u.callbacks=[g]:b.push(g))}else b={lane:g,tag:i.tag,payload:i.payload,callback:i.callback,next:null},z===null?(m=z=b,f=A):z=z.next=b,c|=g;if(i=i.next,i===null){if(i=u.shared.pending,i===null)break;b=i,i=b.next,b.next=null,u.lastBaseUpdate=b,u.shared.pending=null}}while(!0);z===null&&(f=A),u.baseState=f,u.firstBaseUpdate=m,u.lastBaseUpdate=z,n===null&&(u.shared.lanes=0),ia|=c,l.lanes=c,l.memoizedState=A}}function Os(l,t){if(typeof l!="function")throw Error(s(191,l));l.call(t)}function ps(l,t){var a=l.callbacks;if(a!==null)for(l.callbacks=null,l=0;l<a.length;l++)Os(a[l],t)}var ee=p(null),Fu=p(0);function xs(l,t){l=Vt,U(Fu,l),U(ee,t),Vt=l|t.baseLanes}function Cc(){U(Fu,Vt),U(ee,ee.current)}function Gc(){Vt=Fu.current,H(ee),H(Fu)}var la=0,W=null,fl=null,Al=null,Pu=!1,ue=!1,Ra=!1,Iu=0,We=0,ne=null,H0=0;function Tl(){throw Error(s(321))}function Xc(l,t){if(t===null)return!1;for(var a=0;a<t.length&&a<l.length;a++)if(!tt(l[a],t[a]))return!1;return!0}function Qc(l,t,a,e,u,n){return la=n,W=t,t.memoizedState=null,t.updateQueue=null,t.lanes=0,S.H=l===null||l.memoizedState===null?so:oo,Ra=!1,n=a(e,u),Ra=!1,ue&&(n=Ms(t,a,e,u)),_s(l),n}function _s(l){S.H=nn;var t=fl!==null&&fl.next!==null;if(la=0,Al=fl=W=null,Pu=!1,We=0,ne=null,t)throw Error(s(300));l===null||Rl||(l=l.dependencies,l!==null&&Ju(l)&&(Rl=!0))}function Ms(l,t,a,e){W=l;var u=0;do{if(ue&&(ne=null),We=0,ue=!1,25<=u)throw Error(s(301));if(u+=1,Al=fl=null,l.updateQueue!=null){var n=l.updateQueue;n.lastEffect=null,n.events=null,n.stores=null,n.memoCache!=null&&(n.memoCache.index=0)}S.H=Q0,n=t(a,e)}while(ue);return n}function q0(){var l=S.H,t=l.useState()[0];return t=typeof t.then=="function"?ke(t):t,l=l.useState()[0],(fl!==null?fl.memoizedState:null)!==l&&(W.flags|=1024),t}function Zc(){var l=Iu!==0;return Iu=0,l}function Vc(l,t,a){t.updateQueue=l.updateQueue,t.flags&=-2053,l.lanes&=~a}function Lc(l){if(Pu){for(l=l.memoizedState;l!==null;){var t=l.queue;t!==null&&(t.pending=null),l=l.next}Pu=!1}la=0,Al=fl=W=null,ue=!1,We=Iu=0,ne=null}function kl(){var l={memoizedState:null,baseState:null,baseQueue:null,queue:null,next:null};return Al===null?W.memoizedState=Al=l:Al=Al.next=l,Al}function Ol(){if(fl===null){var l=W.alternate;l=l!==null?l.memoizedState:null}else l=fl.next;var t=Al===null?W.memoizedState:Al.next;if(t!==null)Al=t,fl=l;else{if(l===null)throw W.alternate===null?Error(s(467)):Error(s(310));fl=l,l={memoizedState:fl.memoizedState,baseState:fl.baseState,baseQueue:fl.baseQueue,queue:fl.queue,next:null},Al===null?W.memoizedState=Al=l:Al=Al.next=l}return Al}function Kc(){return{lastEffect:null,events:null,stores:null,memoCache:null}}function ke(l){var t=We;return We+=1,ne===null&&(ne=[]),l=Ts(ne,l,t),t=W,(Al===null?t.memoizedState:Al.next)===null&&(t=t.alternate,S.H=t===null||t.memoizedState===null?so:oo),l}function ln(l){if(l!==null&&typeof l=="object"){if(typeof l.then=="function")return ke(l);if(l.$$typeof===dl)return Ql(l)}throw Error(s(438,String(l)))}function Jc(l){var t=null,a=W.updateQueue;if(a!==null&&(t=a.memoCache),t==null){var e=W.alternate;e!==null&&(e=e.updateQueue,e!==null&&(e=e.memoCache,e!=null&&(t={data:e.data.map(function(u){return u.slice()}),index:0})))}if(t==null&&(t={data:[],index:0}),a===null&&(a=Kc(),W.updateQueue=a),a.memoCache=t,a=t.data[t.index],a===void 0)for(a=t.data[t.index]=Array(l),e=0;e<l;e++)a[e]=Kt;return t.index++,a}function Bt(l,t){return typeof t=="function"?t(l):t}function tn(l){var t=Ol();return wc(t,fl,l)}function wc(l,t,a){var e=l.queue;if(e===null)throw Error(s(311));e.lastRenderedReducer=a;var u=l.baseQueue,n=e.pending;if(n!==null){if(u!==null){var c=u.next;u.next=n.next,n.next=c}t.baseQueue=u=n,e.pending=null}if(n=l.baseState,u===null)l.memoizedState=n;else{t=u.next;var i=c=null,f=null,m=t,z=!1;do{var A=m.lane&-536870913;if(A!==m.lane?(I&A)===A:(la&A)===A){var g=m.revertLane;if(g===0)f!==null&&(f=f.next={lane:0,revertLane:0,action:m.action,hasEagerState:m.hasEagerState,eagerState:m.eagerState,next:null}),A===te&&(z=!0);else if((la&g)===g){m=m.next,g===te&&(z=!0);continue}else A={lane:0,revertLane:m.revertLane,action:m.action,hasEagerState:m.hasEagerState,eagerState:m.eagerState,next:null},f===null?(i=f=A,c=n):f=f.next=A,W.lanes|=g,ia|=g;A=m.action,Ra&&a(n,A),n=m.hasEagerState?m.eagerState:a(n,A)}else g={lane:A,revertLane:m.revertLane,action:m.action,hasEagerState:m.hasEagerState,eagerState:m.eagerState,next:null},f===null?(i=f=g,c=n):f=f.next=g,W.lanes|=A,ia|=A;m=m.next}while(m!==null&&m!==t);if(f===null?c=n:f.next=i,!tt(n,l.memoizedState)&&(Rl=!0,z&&(a=ae,a!==null)))throw a;l.memoizedState=n,l.baseState=c,l.baseQueue=f,e.lastRenderedState=n}return u===null&&(e.lanes=0),[l.memoizedState,e.dispatch]}function Wc(l){var t=Ol(),a=t.queue;if(a===null)throw Error(s(311));a.lastRenderedReducer=l;var e=a.dispatch,u=a.pending,n=t.memoizedState;if(u!==null){a.pending=null;var c=u=u.next;do n=l(n,c.action),c=c.next;while(c!==u);tt(n,t.memoizedState)||(Rl=!0),t.memoizedState=n,t.baseQueue===null&&(t.baseState=n),a.lastRenderedState=n}return[n,e]}function Ds(l,t,a){var e=W,u=Ol(),n=el;if(n){if(a===void 0)throw Error(s(407));a=a()}else a=t();var c=!tt((fl||u).memoizedState,a);c&&(u.memoizedState=a,Rl=!0),u=u.queue;var i=Rs.bind(null,e,u,l);if($e(2048,8,i,[l]),u.getSnapshot!==t||c||Al!==null&&Al.memoizedState.tag&1){if(e.flags|=2048,ce(9,an(),Us.bind(null,e,u,a,t),null),yl===null)throw Error(s(349));n||(la&124)!==0||Ns(e,t,a)}return a}function Ns(l,t,a){l.flags|=16384,l={getSnapshot:t,value:a},t=W.updateQueue,t===null?(t=Kc(),W.updateQueue=t,t.stores=[l]):(a=t.stores,a===null?t.stores=[l]:a.push(l))}function Us(l,t,a,e){t.value=a,t.getSnapshot=e,js(t)&&Hs(l)}function Rs(l,t,a){return a(function(){js(t)&&Hs(l)})}function js(l){var t=l.getSnapshot;l=l.value;try{var a=t();return!tt(l,a)}catch{return!0}}function Hs(l){var t=Fa(l,2);t!==null&&it(t,l,2)}function kc(l){var t=kl();if(typeof l=="function"){var a=l;if(l=a(),Ra){wt(!0);try{a()}finally{wt(!1)}}}return t.memoizedState=t.baseState=l,t.queue={pending:null,lanes:0,dispatch:null,lastRenderedReducer:Bt,lastRenderedState:l},t}function qs(l,t,a,e){return l.baseState=a,wc(l,fl,typeof e=="function"?e:Bt)}function Y0(l,t,a,e,u){if(un(l))throw Error(s(485));if(l=t.action,l!==null){var n={payload:u,action:l,next:null,isTransition:!0,status:"pending",value:null,reason:null,listeners:[],then:function(c){n.listeners.push(c)}};S.T!==null?a(!0):n.isTransition=!1,e(n),a=t.pending,a===null?(n.next=t.pending=n,Ys(t,n)):(n.next=a.next,t.pending=a.next=n)}}function Ys(l,t){var a=t.action,e=t.payload,u=l.state;if(t.isTransition){var n=S.T,c={};S.T=c;try{var i=a(u,e),f=S.S;f!==null&&f(c,i),Bs(l,t,i)}catch(m){$c(l,t,m)}finally{S.T=n}}else try{n=a(u,e),Bs(l,t,n)}catch(m){$c(l,t,m)}}function Bs(l,t,a){a!==null&&typeof a=="object"&&typeof a.then=="function"?a.then(function(e){Cs(l,t,e)},function(e){return $c(l,t,e)}):Cs(l,t,a)}function Cs(l,t,a){t.status="fulfilled",t.value=a,Gs(t),l.state=a,t=l.pending,t!==null&&(a=t.next,a===t?l.pending=null:(a=a.next,t.next=a,Ys(l,a)))}function $c(l,t,a){var e=l.pending;if(l.pending=null,e!==null){e=e.next;do t.status="rejected",t.reason=a,Gs(t),t=t.next;while(t!==e)}l.action=null}function Gs(l){l=l.listeners;for(var t=0;t<l.length;t++)(0,l[t])()}function Xs(l,t){return t}function Qs(l,t){if(el){var a=yl.formState;if(a!==null){l:{var e=W;if(el){if(Sl){t:{for(var u=Sl,n=At;u.nodeType!==8;){if(!n){u=null;break t}if(u=Tt(u.nextSibling),u===null){u=null;break t}}n=u.data,u=n==="F!"||n==="F"?u:null}if(u){Sl=Tt(u.nextSibling),e=u.data==="F!";break l}}Ma(e)}e=!1}e&&(t=a[0])}}return a=kl(),a.memoizedState=a.baseState=t,e={pending:null,lanes:0,dispatch:null,lastRenderedReducer:Xs,lastRenderedState:t},a.queue=e,a=co.bind(null,W,e),e.dispatch=a,e=kc(!1),n=ti.bind(null,W,!1,e.queue),e=kl(),u={state:t,dispatch:null,action:l,pending:null},e.queue=u,a=Y0.bind(null,W,u,n,a),u.dispatch=a,e.memoizedState=l,[t,a,!1]}function Zs(l){var t=Ol();return Vs(t,fl,l)}function Vs(l,t,a){if(t=wc(l,t,Xs)[0],l=tn(Bt)[0],typeof t=="object"&&t!==null&&typeof t.then=="function")try{var e=ke(t)}catch(c){throw c===Ve?ku:c}else e=t;t=Ol();var u=t.queue,n=u.dispatch;return a!==t.memoizedState&&(W.flags|=2048,ce(9,an(),B0.bind(null,u,a),null)),[e,n,l]}function B0(l,t){l.action=t}function Ls(l){var t=Ol(),a=fl;if(a!==null)return Vs(t,a,l);Ol(),t=t.memoizedState,a=Ol();var e=a.queue.dispatch;return a.memoizedState=l,[t,e,!1]}function ce(l,t,a,e){return l={tag:l,create:a,deps:e,inst:t,next:null},t=W.updateQueue,t===null&&(t=Kc(),W.updateQueue=t),a=t.lastEffect,a===null?t.lastEffect=l.next=l:(e=a.next,a.next=l,l.next=e,t.lastEffect=l),l}function an(){return{destroy:void 0,resource:void 0}}function Ks(){return Ol().memoizedState}function en(l,t,a,e){var u=kl();e=e===void 0?null:e,W.flags|=l,u.memoizedState=ce(1|t,an(),a,e)}function $e(l,t,a,e){var u=Ol();e=e===void 0?null:e;var n=u.memoizedState.inst;fl!==null&&e!==null&&Xc(e,fl.memoizedState.deps)?u.memoizedState=ce(t,n,a,e):(W.flags|=l,u.memoizedState=ce(1|t,n,a,e))}function Js(l,t){en(8390656,8,l,t)}function ws(l,t){$e(2048,8,l,t)}function Ws(l,t){return $e(4,2,l,t)}function ks(l,t){return $e(4,4,l,t)}function $s(l,t){if(typeof t=="function"){l=l();var a=t(l);return function(){typeof a=="function"?a():t(null)}}if(t!=null)return l=l(),t.current=l,function(){t.current=null}}function Fs(l,t,a){a=a!=null?a.concat([l]):null,$e(4,4,$s.bind(null,t,l),a)}function Fc(){}function Ps(l,t){var a=Ol();t=t===void 0?null:t;var e=a.memoizedState;return t!==null&&Xc(t,e[1])?e[0]:(a.memoizedState=[l,t],l)}function Is(l,t){var a=Ol();t=t===void 0?null:t;var e=a.memoizedState;if(t!==null&&Xc(t,e[1]))return e[0];if(e=l(),Ra){wt(!0);try{l()}finally{wt(!1)}}return a.memoizedState=[e,t],e}function Pc(l,t,a){return a===void 0||(la&1073741824)!==0?l.memoizedState=t:(l.memoizedState=a,l=ed(),W.lanes|=l,ia|=l,a)}function lo(l,t,a,e){return tt(a,t)?a:ee.current!==null?(l=Pc(l,a,e),tt(l,t)||(Rl=!0),l):(la&42)===0?(Rl=!0,l.memoizedState=a):(l=ed(),W.lanes|=l,ia|=l,t)}function to(l,t,a,e,u){var n=N.p;N.p=n!==0&&8>n?n:8;var c=S.T,i={};S.T=i,ti(l,!1,t,a);try{var f=u(),m=S.S;if(m!==null&&m(i,f),f!==null&&typeof f=="object"&&typeof f.then=="function"){var z=j0(f,e);Fe(l,t,z,ct(l))}else Fe(l,t,e,ct(l))}catch(A){Fe(l,t,{then:function(){},status:"rejected",reason:A},ct())}finally{N.p=n,S.T=c}}function C0(){}function Ic(l,t,a,e){if(l.tag!==5)throw Error(s(476));var u=ao(l).queue;to(l,u,t,V,a===null?C0:function(){return eo(l),a(e)})}function ao(l){var t=l.memoizedState;if(t!==null)return t;t={memoizedState:V,baseState:V,baseQueue:null,queue:{pending:null,lanes:0,dispatch:null,lastRenderedReducer:Bt,lastRenderedState:V},next:null};var a={};return t.next={memoizedState:a,baseState:a,baseQueue:null,queue:{pending:null,lanes:0,dispatch:null,lastRenderedReducer:Bt,lastRenderedState:a},next:null},l.memoizedState=t,l=l.alternate,l!==null&&(l.memoizedState=t),t}function eo(l){var t=ao(l).next.queue;Fe(l,t,{},ct())}function li(){return Ql(hu)}function uo(){return Ol().memoizedState}function no(){return Ol().memoizedState}function G0(l){for(var t=l.return;t!==null;){switch(t.tag){case 24:case 3:var a=ct();l=Pt(a);var e=It(t,l,a);e!==null&&(it(e,t,a),Ke(e,t,a)),t={cache:Nc()},l.payload=t;return}t=t.return}}function X0(l,t,a){var e=ct();a={lane:e,revertLane:0,action:a,hasEagerState:!1,eagerState:null,next:null},un(l)?io(t,a):(a=zc(l,t,a,e),a!==null&&(it(a,l,e),fo(a,t,e)))}function co(l,t,a){var e=ct();Fe(l,t,a,e)}function Fe(l,t,a,e){var u={lane:e,revertLane:0,action:a,hasEagerState:!1,eagerState:null,next:null};if(un(l))io(t,u);else{var n=l.alternate;if(l.lanes===0&&(n===null||n.lanes===0)&&(n=t.lastRenderedReducer,n!==null))try{var c=t.lastRenderedState,i=n(c,a);if(u.hasEagerState=!0,u.eagerState=i,tt(i,c))return Qu(l,t,u,0),yl===null&&Xu(),!1}catch{}finally{}if(a=zc(l,t,u,e),a!==null)return it(a,l,e),fo(a,t,e),!0}return!1}function ti(l,t,a,e){if(e={lane:2,revertLane:ji(),action:e,hasEagerState:!1,eagerState:null,next:null},un(l)){if(t)throw Error(s(479))}else t=zc(l,a,e,2),t!==null&&it(t,l,2)}function un(l){var t=l.alternate;return l===W||t!==null&&t===W}function io(l,t){ue=Pu=!0;var a=l.pending;a===null?t.next=t:(t.next=a.next,a.next=t),l.pending=t}function fo(l,t,a){if((a&4194048)!==0){var e=t.lanes;e&=l.pendingLanes,a|=e,t.lanes=a,mf(l,a)}}var nn={readContext:Ql,use:ln,useCallback:Tl,useContext:Tl,useEffect:Tl,useImperativeHandle:Tl,useLayoutEffect:Tl,useInsertionEffect:Tl,useMemo:Tl,useReducer:Tl,useRef:Tl,useState:Tl,useDebugValue:Tl,useDeferredValue:Tl,useTransition:Tl,useSyncExternalStore:Tl,useId:Tl,useHostTransitionStatus:Tl,useFormState:Tl,useActionState:Tl,useOptimistic:Tl,useMemoCache:Tl,useCacheRefresh:Tl},so={readContext:Ql,use:ln,useCallback:function(l,t){return kl().memoizedState=[l,t===void 0?null:t],l},useContext:Ql,useEffect:Js,useImperativeHandle:function(l,t,a){a=a!=null?a.concat([l]):null,en(4194308,4,$s.bind(null,t,l),a)},useLayoutEffect:function(l,t){return en(4194308,4,l,t)},useInsertionEffect:function(l,t){en(4,2,l,t)},useMemo:function(l,t){var a=kl();t=t===void 0?null:t;var e=l();if(Ra){wt(!0);try{l()}finally{wt(!1)}}return a.memoizedState=[e,t],e},useReducer:function(l,t,a){var e=kl();if(a!==void 0){var u=a(t);if(Ra){wt(!0);try{a(t)}finally{wt(!1)}}}else u=t;return e.memoizedState=e.baseState=u,l={pending:null,lanes:0,dispatch:null,lastRenderedReducer:l,lastRenderedState:u},e.queue=l,l=l.dispatch=X0.bind(null,W,l),[e.memoizedState,l]},useRef:function(l){var t=kl();return l={current:l},t.memoizedState=l},useState:function(l){l=kc(l);var t=l.queue,a=co.bind(null,W,t);return t.dispatch=a,[l.memoizedState,a]},useDebugValue:Fc,useDeferredValue:function(l,t){var a=kl();return Pc(a,l,t)},useTransition:function(){var l=kc(!1);return l=to.bind(null,W,l.queue,!0,!1),kl().memoizedState=l,[!1,l]},useSyncExternalStore:function(l,t,a){var e=W,u=kl();if(el){if(a===void 0)throw Error(s(407));a=a()}else{if(a=t(),yl===null)throw Error(s(349));(I&124)!==0||Ns(e,t,a)}u.memoizedState=a;var n={value:a,getSnapshot:t};return u.queue=n,Js(Rs.bind(null,e,n,l),[l]),e.flags|=2048,ce(9,an(),Us.bind(null,e,n,a,t),null),a},useId:function(){var l=kl(),t=yl.identifierPrefix;if(el){var a=Ht,e=jt;a=(e&~(1<<32-lt(e)-1)).toString(32)+a,t="«"+t+"R"+a,a=Iu++,0<a&&(t+="H"+a.toString(32)),t+="»"}else a=H0++,t="«"+t+"r"+a.toString(32)+"»";return l.memoizedState=t},useHostTransitionStatus:li,useFormState:Qs,useActionState:Qs,useOptimistic:function(l){var t=kl();t.memoizedState=t.baseState=l;var a={pending:null,lanes:0,dispatch:null,lastRenderedReducer:null,lastRenderedState:null};return t.queue=a,t=ti.bind(null,W,!0,a),a.dispatch=t,[l,t]},useMemoCache:Jc,useCacheRefresh:function(){return kl().memoizedState=G0.bind(null,W)}},oo={readContext:Ql,use:ln,useCallback:Ps,useContext:Ql,useEffect:ws,useImperativeHandle:Fs,useInsertionEffect:Ws,useLayoutEffect:ks,useMemo:Is,useReducer:tn,useRef:Ks,useState:function(){return tn(Bt)},useDebugValue:Fc,useDeferredValue:function(l,t){var a=Ol();return lo(a,fl.memoizedState,l,t)},useTransition:function(){var l=tn(Bt)[0],t=Ol().memoizedState;return[typeof l=="boolean"?l:ke(l),t]},useSyncExternalStore:Ds,useId:uo,useHostTransitionStatus:li,useFormState:Zs,useActionState:Zs,useOptimistic:function(l,t){var a=Ol();return qs(a,fl,l,t)},useMemoCache:Jc,useCacheRefresh:no},Q0={readContext:Ql,use:ln,useCallback:Ps,useContext:Ql,useEffect:ws,useImperativeHandle:Fs,useInsertionEffect:Ws,useLayoutEffect:ks,useMemo:Is,useReducer:Wc,useRef:Ks,useState:function(){return Wc(Bt)},useDebugValue:Fc,useDeferredValue:function(l,t){var a=Ol();return fl===null?Pc(a,l,t):lo(a,fl.memoizedState,l,t)},useTransition:function(){var l=Wc(Bt)[0],t=Ol().memoizedState;return[typeof l=="boolean"?l:ke(l),t]},useSyncExternalStore:Ds,useId:uo,useHostTransitionStatus:li,useFormState:Ls,useActionState:Ls,useOptimistic:function(l,t){var a=Ol();return fl!==null?qs(a,fl,l,t):(a.baseState=l,[l,a.queue.dispatch])},useMemoCache:Jc,useCacheRefresh:no},ie=null,Pe=0;function cn(l){var t=Pe;return Pe+=1,ie===null&&(ie=[]),Ts(ie,l,t)}function Ie(l,t){t=t.props.ref,l.ref=t!==void 0?t:null}function fn(l,t){throw t.$$typeof===Q?Error(s(525)):(l=Object.prototype.toString.call(t),Error(s(31,l==="[object Object]"?"object with keys {"+Object.keys(t).join(", ")+"}":l)))}function ro(l){var t=l._init;return t(l._payload)}function vo(l){function t(r,d){if(l){var h=r.deletions;h===null?(r.deletions=[d],r.flags|=16):h.push(d)}}function a(r,d){if(!l)return null;for(;d!==null;)t(r,d),d=d.sibling;return null}function e(r){for(var d=new Map;r!==null;)r.key!==null?d.set(r.key,r):d.set(r.index,r),r=r.sibling;return d}function u(r,d){return r=Rt(r,d),r.index=0,r.sibling=null,r}function n(r,d,h){return r.index=h,l?(h=r.alternate,h!==null?(h=h.index,h<d?(r.flags|=67108866,d):h):(r.flags|=67108866,d)):(r.flags|=1048576,d)}function c(r){return l&&r.alternate===null&&(r.flags|=67108866),r}function i(r,d,h,T){return d===null||d.tag!==6?(d=Ec(h,r.mode,T),d.return=r,d):(d=u(d,h),d.return=r,d)}function f(r,d,h,T){var Y=h.type;return Y===ul?z(r,d,h.props.children,T,h.key):d!==null&&(d.elementType===Y||typeof Y=="object"&&Y!==null&&Y.$$typeof===Kl&&ro(Y)===d.type)?(d=u(d,h.props),Ie(d,h),d.return=r,d):(d=Vu(h.type,h.key,h.props,null,r.mode,T),Ie(d,h),d.return=r,d)}function m(r,d,h,T){return d===null||d.tag!==4||d.stateNode.containerInfo!==h.containerInfo||d.stateNode.implementation!==h.implementation?(d=Ac(h,r.mode,T),d.return=r,d):(d=u(d,h.children||[]),d.return=r,d)}function z(r,d,h,T,Y){return d===null||d.tag!==7?(d=Oa(h,r.mode,T,Y),d.return=r,d):(d=u(d,h),d.return=r,d)}function A(r,d,h){if(typeof d=="string"&&d!==""||typeof d=="number"||typeof d=="bigint")return d=Ec(""+d,r.mode,h),d.return=r,d;if(typeof d=="object"&&d!==null){switch(d.$$typeof){case B:return h=Vu(d.type,d.key,d.props,null,r.mode,h),Ie(h,d),h.return=r,h;case ll:return d=Ac(d,r.mode,h),d.return=r,d;case Kl:var T=d._init;return d=T(d._payload),A(r,d,h)}if(il(d)||Nl(d))return d=Oa(d,r.mode,h,null),d.return=r,d;if(typeof d.then=="function")return A(r,cn(d),h);if(d.$$typeof===dl)return A(r,wu(r,d),h);fn(r,d)}return null}function g(r,d,h,T){var Y=d!==null?d.key:null;if(typeof h=="string"&&h!==""||typeof h=="number"||typeof h=="bigint")return Y!==null?null:i(r,d,""+h,T);if(typeof h=="object"&&h!==null){switch(h.$$typeof){case B:return h.key===Y?f(r,d,h,T):null;case ll:return h.key===Y?m(r,d,h,T):null;case Kl:return Y=h._init,h=Y(h._payload),g(r,d,h,T)}if(il(h)||Nl(h))return Y!==null?null:z(r,d,h,T,null);if(typeof h.then=="function")return g(r,d,cn(h),T);if(h.$$typeof===dl)return g(r,d,wu(r,h),T);fn(r,h)}return null}function b(r,d,h,T,Y){if(typeof T=="string"&&T!==""||typeof T=="number"||typeof T=="bigint")return r=r.get(h)||null,i(d,r,""+T,Y);if(typeof T=="object"&&T!==null){switch(T.$$typeof){case B:return r=r.get(T.key===null?h:T.key)||null,f(d,r,T,Y);case ll:return r=r.get(T.key===null?h:T.key)||null,m(d,r,T,Y);case Kl:var k=T._init;return T=k(T._payload),b(r,d,h,T,Y)}if(il(T)||Nl(T))return r=r.get(h)||null,z(d,r,T,Y,null);if(typeof T.then=="function")return b(r,d,h,cn(T),Y);if(T.$$typeof===dl)return b(r,d,h,wu(d,T),Y);fn(d,T)}return null}function L(r,d,h,T){for(var Y=null,k=null,C=d,Z=d=0,Hl=null;C!==null&&Z<h.length;Z++){C.index>Z?(Hl=C,C=null):Hl=C.sibling;var al=g(r,C,h[Z],T);if(al===null){C===null&&(C=Hl);break}l&&C&&al.alternate===null&&t(r,C),d=n(al,d,Z),k===null?Y=al:k.sibling=al,k=al,C=Hl}if(Z===h.length)return a(r,C),el&&xa(r,Z),Y;if(C===null){for(;Z<h.length;Z++)C=A(r,h[Z],T),C!==null&&(d=n(C,d,Z),k===null?Y=C:k.sibling=C,k=C);return el&&xa(r,Z),Y}for(C=e(C);Z<h.length;Z++)Hl=b(C,r,Z,h[Z],T),Hl!==null&&(l&&Hl.alternate!==null&&C.delete(Hl.key===null?Z:Hl.key),d=n(Hl,d,Z),k===null?Y=Hl:k.sibling=Hl,k=Hl);return l&&C.forEach(function(ma){return t(r,ma)}),el&&xa(r,Z),Y}function X(r,d,h,T){if(h==null)throw Error(s(151));for(var Y=null,k=null,C=d,Z=d=0,Hl=null,al=h.next();C!==null&&!al.done;Z++,al=h.next()){C.index>Z?(Hl=C,C=null):Hl=C.sibling;var ma=g(r,C,al.value,T);if(ma===null){C===null&&(C=Hl);break}l&&C&&ma.alternate===null&&t(r,C),d=n(ma,d,Z),k===null?Y=ma:k.sibling=ma,k=ma,C=Hl}if(al.done)return a(r,C),el&&xa(r,Z),Y;if(C===null){for(;!al.done;Z++,al=h.next())al=A(r,al.value,T),al!==null&&(d=n(al,d,Z),k===null?Y=al:k.sibling=al,k=al);return el&&xa(r,Z),Y}for(C=e(C);!al.done;Z++,al=h.next())al=b(C,r,Z,al.value,T),al!==null&&(l&&al.alternate!==null&&C.delete(al.key===null?Z:al.key),d=n(al,d,Z),k===null?Y=al:k.sibling=al,k=al);return l&&C.forEach(function(Zv){return t(r,Zv)}),el&&xa(r,Z),Y}function ol(r,d,h,T){if(typeof h=="object"&&h!==null&&h.type===ul&&h.key===null&&(h=h.props.children),typeof h=="object"&&h!==null){switch(h.$$typeof){case B:l:{for(var Y=h.key;d!==null;){if(d.key===Y){if(Y=h.type,Y===ul){if(d.tag===7){a(r,d.sibling),T=u(d,h.props.children),T.return=r,r=T;break l}}else if(d.elementType===Y||typeof Y=="object"&&Y!==null&&Y.$$typeof===Kl&&ro(Y)===d.type){a(r,d.sibling),T=u(d,h.props),Ie(T,h),T.return=r,r=T;break l}a(r,d);break}else t(r,d);d=d.sibling}h.type===ul?(T=Oa(h.props.children,r.mode,T,h.key),T.return=r,r=T):(T=Vu(h.type,h.key,h.props,null,r.mode,T),Ie(T,h),T.return=r,r=T)}return c(r);case ll:l:{for(Y=h.key;d!==null;){if(d.key===Y)if(d.tag===4&&d.stateNode.containerInfo===h.containerInfo&&d.stateNode.implementation===h.implementation){a(r,d.sibling),T=u(d,h.children||[]),T.return=r,r=T;break l}else{a(r,d);break}else t(r,d);d=d.sibling}T=Ac(h,r.mode,T),T.return=r,r=T}return c(r);case Kl:return Y=h._init,h=Y(h._payload),ol(r,d,h,T)}if(il(h))return L(r,d,h,T);if(Nl(h)){if(Y=Nl(h),typeof Y!="function")throw Error(s(150));return h=Y.call(h),X(r,d,h,T)}if(typeof h.then=="function")return ol(r,d,cn(h),T);if(h.$$typeof===dl)return ol(r,d,wu(r,h),T);fn(r,h)}return typeof h=="string"&&h!==""||typeof h=="number"||typeof h=="bigint"?(h=""+h,d!==null&&d.tag===6?(a(r,d.sibling),T=u(d,h),T.return=r,r=T):(a(r,d),T=Ec(h,r.mode,T),T.return=r,r=T),c(r)):a(r,d)}return function(r,d,h,T){try{Pe=0;var Y=ol(r,d,h,T);return ie=null,Y}catch(C){if(C===Ve||C===ku)throw C;var k=at(29,C,null,r.mode);return k.lanes=T,k.return=r,k}finally{}}}var fe=vo(!0),yo=vo(!1),ht=p(null),Ot=null;function ta(l){var t=l.alternate;U(_l,_l.current&1),U(ht,l),Ot===null&&(t===null||ee.current!==null||t.memoizedState!==null)&&(Ot=l)}function ho(l){if(l.tag===22){if(U(_l,_l.current),U(ht,l),Ot===null){var t=l.alternate;t!==null&&t.memoizedState!==null&&(Ot=l)}}else aa()}function aa(){U(_l,_l.current),U(ht,ht.current)}function Ct(l){H(ht),Ot===l&&(Ot=null),H(_l)}var _l=p(0);function sn(l){for(var t=l;t!==null;){if(t.tag===13){var a=t.memoizedState;if(a!==null&&(a=a.dehydrated,a===null||a.data==="$?"||Ki(a)))return t}else if(t.tag===19&&t.memoizedProps.revealOrder!==void 0){if((t.flags&128)!==0)return t}else if(t.child!==null){t.child.return=t,t=t.child;continue}if(t===l)break;for(;t.sibling===null;){if(t.return===null||t.return===l)return null;t=t.return}t.sibling.return=t.return,t=t.sibling}return null}function ai(l,t,a,e){t=l.memoizedState,a=a(e,t),a=a==null?t:D({},t,a),l.memoizedState=a,l.lanes===0&&(l.updateQueue.baseState=a)}var ei={enqueueSetState:function(l,t,a){l=l._reactInternals;var e=ct(),u=Pt(e);u.payload=t,a!=null&&(u.callback=a),t=It(l,u,e),t!==null&&(it(t,l,e),Ke(t,l,e))},enqueueReplaceState:function(l,t,a){l=l._reactInternals;var e=ct(),u=Pt(e);u.tag=1,u.payload=t,a!=null&&(u.callback=a),t=It(l,u,e),t!==null&&(it(t,l,e),Ke(t,l,e))},enqueueForceUpdate:function(l,t){l=l._reactInternals;var a=ct(),e=Pt(a);e.tag=2,t!=null&&(e.callback=t),t=It(l,e,a),t!==null&&(it(t,l,a),Ke(t,l,a))}};function mo(l,t,a,e,u,n,c){return l=l.stateNode,typeof l.shouldComponentUpdate=="function"?l.shouldComponentUpdate(e,n,c):t.prototype&&t.prototype.isPureReactComponent?!qe(a,e)||!qe(u,n):!0}function go(l,t,a,e){l=t.state,typeof t.componentWillReceiveProps=="function"&&t.componentWillReceiveProps(a,e),typeof t.UNSAFE_componentWillReceiveProps=="function"&&t.UNSAFE_componentWillReceiveProps(a,e),t.state!==l&&ei.enqueueReplaceState(t,t.state,null)}function ja(l,t){var a=t;if("ref"in t){a={};for(var e in t)e!=="ref"&&(a[e]=t[e])}if(l=l.defaultProps){a===t&&(a=D({},a));for(var u in l)a[u]===void 0&&(a[u]=l[u])}return a}var on=typeof reportError=="function"?reportError:function(l){if(typeof window=="object"&&typeof window.ErrorEvent=="function"){var t=new window.ErrorEvent("error",{bubbles:!0,cancelable:!0,message:typeof l=="object"&&l!==null&&typeof l.message=="string"?String(l.message):String(l),error:l});if(!window.dispatchEvent(t))return}else if(typeof process=="object"&&typeof process.emit=="function"){process.emit("uncaughtException",l);return}console.error(l)};function bo(l){on(l)}function So(l){console.error(l)}function zo(l){on(l)}function dn(l,t){try{var a=l.onUncaughtError;a(t.value,{componentStack:t.stack})}catch(e){setTimeout(function(){throw e})}}function To(l,t,a){try{var e=l.onCaughtError;e(a.value,{componentStack:a.stack,errorBoundary:t.tag===1?t.stateNode:null})}catch(u){setTimeout(function(){throw u})}}function ui(l,t,a){return a=Pt(a),a.tag=3,a.payload={element:null},a.callback=function(){dn(l,t)},a}function Eo(l){return l=Pt(l),l.tag=3,l}function Ao(l,t,a,e){var u=a.type.getDerivedStateFromError;if(typeof u=="function"){var n=e.value;l.payload=function(){return u(n)},l.callback=function(){To(t,a,e)}}var c=a.stateNode;c!==null&&typeof c.componentDidCatch=="function"&&(l.callback=function(){To(t,a,e),typeof u!="function"&&(fa===null?fa=new Set([this]):fa.add(this));var i=e.stack;this.componentDidCatch(e.value,{componentStack:i!==null?i:""})})}function Z0(l,t,a,e,u){if(a.flags|=32768,e!==null&&typeof e=="object"&&typeof e.then=="function"){if(t=a.alternate,t!==null&&Xe(t,a,u,!0),a=ht.current,a!==null){switch(a.tag){case 13:return Ot===null?Mi():a.alternate===null&&zl===0&&(zl=3),a.flags&=-257,a.flags|=65536,a.lanes=u,e===jc?a.flags|=16384:(t=a.updateQueue,t===null?a.updateQueue=new Set([e]):t.add(e),Ni(l,e,u)),!1;case 22:return a.flags|=65536,e===jc?a.flags|=16384:(t=a.updateQueue,t===null?(t={transitions:null,markerInstances:null,retryQueue:new Set([e])},a.updateQueue=t):(a=t.retryQueue,a===null?t.retryQueue=new Set([e]):a.add(e)),Ni(l,e,u)),!1}throw Error(s(435,a.tag))}return Ni(l,e,u),Mi(),!1}if(el)return t=ht.current,t!==null?((t.flags&65536)===0&&(t.flags|=256),t.flags|=65536,t.lanes=u,e!==xc&&(l=Error(s(422),{cause:e}),Ge(dt(l,a)))):(e!==xc&&(t=Error(s(423),{cause:e}),Ge(dt(t,a))),l=l.current.alternate,l.flags|=65536,u&=-u,l.lanes|=u,e=dt(e,a),u=ui(l.stateNode,e,u),Yc(l,u),zl!==4&&(zl=2)),!1;var n=Error(s(520),{cause:e});if(n=dt(n,a),cu===null?cu=[n]:cu.push(n),zl!==4&&(zl=2),t===null)return!0;e=dt(e,a),a=t;do{switch(a.tag){case 3:return a.flags|=65536,l=u&-u,a.lanes|=l,l=ui(a.stateNode,e,l),Yc(a,l),!1;case 1:if(t=a.type,n=a.stateNode,(a.flags&128)===0&&(typeof t.getDerivedStateFromError=="function"||n!==null&&typeof n.componentDidCatch=="function"&&(fa===null||!fa.has(n))))return a.flags|=65536,u&=-u,a.lanes|=u,u=Eo(u),Ao(u,l,a,e),Yc(a,u),!1}a=a.return}while(a!==null);return!1}var Oo=Error(s(461)),Rl=!1;function Bl(l,t,a,e){t.child=l===null?yo(t,null,a,e):fe(t,l.child,a,e)}function po(l,t,a,e,u){a=a.render;var n=t.ref;if("ref"in e){var c={};for(var i in e)i!=="ref"&&(c[i]=e[i])}else c=e;return Na(t),e=Qc(l,t,a,c,n,u),i=Zc(),l!==null&&!Rl?(Vc(l,t,u),Gt(l,t,u)):(el&&i&&Oc(t),t.flags|=1,Bl(l,t,e,u),t.child)}function xo(l,t,a,e,u){if(l===null){var n=a.type;return typeof n=="function"&&!Tc(n)&&n.defaultProps===void 0&&a.compare===null?(t.tag=15,t.type=n,_o(l,t,n,e,u)):(l=Vu(a.type,null,e,t,t.mode,u),l.ref=t.ref,l.return=t,t.child=l)}if(n=l.child,!ri(l,u)){var c=n.memoizedProps;if(a=a.compare,a=a!==null?a:qe,a(c,e)&&l.ref===t.ref)return Gt(l,t,u)}return t.flags|=1,l=Rt(n,e),l.ref=t.ref,l.return=t,t.child=l}function _o(l,t,a,e,u){if(l!==null){var n=l.memoizedProps;if(qe(n,e)&&l.ref===t.ref)if(Rl=!1,t.pendingProps=e=n,ri(l,u))(l.flags&131072)!==0&&(Rl=!0);else return t.lanes=l.lanes,Gt(l,t,u)}return ni(l,t,a,e,u)}function Mo(l,t,a){var e=t.pendingProps,u=e.children,n=l!==null?l.memoizedState:null;if(e.mode==="hidden"){if((t.flags&128)!==0){if(e=n!==null?n.baseLanes|a:a,l!==null){for(u=t.child=l.child,n=0;u!==null;)n=n|u.lanes|u.childLanes,u=u.sibling;t.childLanes=n&~e}else t.childLanes=0,t.child=null;return Do(l,t,e,a)}if((a&536870912)!==0)t.memoizedState={baseLanes:0,cachePool:null},l!==null&&Wu(t,n!==null?n.cachePool:null),n!==null?xs(t,n):Cc(),ho(t);else return t.lanes=t.childLanes=536870912,Do(l,t,n!==null?n.baseLanes|a:a,a)}else n!==null?(Wu(t,n.cachePool),xs(t,n),aa(),t.memoizedState=null):(l!==null&&Wu(t,null),Cc(),aa());return Bl(l,t,u,a),t.child}function Do(l,t,a,e){var u=Rc();return u=u===null?null:{parent:xl._currentValue,pool:u},t.memoizedState={baseLanes:a,cachePool:u},l!==null&&Wu(t,null),Cc(),ho(t),l!==null&&Xe(l,t,e,!0),null}function rn(l,t){var a=t.ref;if(a===null)l!==null&&l.ref!==null&&(t.flags|=4194816);else{if(typeof a!="function"&&typeof a!="object")throw Error(s(284));(l===null||l.ref!==a)&&(t.flags|=4194816)}}function ni(l,t,a,e,u){return Na(t),a=Qc(l,t,a,e,void 0,u),e=Zc(),l!==null&&!Rl?(Vc(l,t,u),Gt(l,t,u)):(el&&e&&Oc(t),t.flags|=1,Bl(l,t,a,u),t.child)}function No(l,t,a,e,u,n){return Na(t),t.updateQueue=null,a=Ms(t,e,a,u),_s(l),e=Zc(),l!==null&&!Rl?(Vc(l,t,n),Gt(l,t,n)):(el&&e&&Oc(t),t.flags|=1,Bl(l,t,a,n),t.child)}function Uo(l,t,a,e,u){if(Na(t),t.stateNode===null){var n=Pa,c=a.contextType;typeof c=="object"&&c!==null&&(n=Ql(c)),n=new a(e,n),t.memoizedState=n.state!==null&&n.state!==void 0?n.state:null,n.updater=ei,t.stateNode=n,n._reactInternals=t,n=t.stateNode,n.props=e,n.state=t.memoizedState,n.refs={},Hc(t),c=a.contextType,n.context=typeof c=="object"&&c!==null?Ql(c):Pa,n.state=t.memoizedState,c=a.getDerivedStateFromProps,typeof c=="function"&&(ai(t,a,c,e),n.state=t.memoizedState),typeof a.getDerivedStateFromProps=="function"||typeof n.getSnapshotBeforeUpdate=="function"||typeof n.UNSAFE_componentWillMount!="function"&&typeof n.componentWillMount!="function"||(c=n.state,typeof n.componentWillMount=="function"&&n.componentWillMount(),typeof n.UNSAFE_componentWillMount=="function"&&n.UNSAFE_componentWillMount(),c!==n.state&&ei.enqueueReplaceState(n,n.state,null),we(t,e,n,u),Je(),n.state=t.memoizedState),typeof n.componentDidMount=="function"&&(t.flags|=4194308),e=!0}else if(l===null){n=t.stateNode;var i=t.memoizedProps,f=ja(a,i);n.props=f;var m=n.context,z=a.contextType;c=Pa,typeof z=="object"&&z!==null&&(c=Ql(z));var A=a.getDerivedStateFromProps;z=typeof A=="function"||typeof n.getSnapshotBeforeUpdate=="function",i=t.pendingProps!==i,z||typeof n.UNSAFE_componentWillReceiveProps!="function"&&typeof n.componentWillReceiveProps!="function"||(i||m!==c)&&go(t,n,e,c),Ft=!1;var g=t.memoizedState;n.state=g,we(t,e,n,u),Je(),m=t.memoizedState,i||g!==m||Ft?(typeof A=="function"&&(ai(t,a,A,e),m=t.memoizedState),(f=Ft||mo(t,a,f,e,g,m,c))?(z||typeof n.UNSAFE_componentWillMount!="function"&&typeof n.componentWillMount!="function"||(typeof n.componentWillMount=="function"&&n.componentWillMount(),typeof n.UNSAFE_componentWillMount=="function"&&n.UNSAFE_componentWillMount()),typeof n.componentDidMount=="function"&&(t.flags|=4194308)):(typeof n.componentDidMount=="function"&&(t.flags|=4194308),t.memoizedProps=e,t.memoizedState=m),n.props=e,n.state=m,n.context=c,e=f):(typeof n.componentDidMount=="function"&&(t.flags|=4194308),e=!1)}else{n=t.stateNode,qc(l,t),c=t.memoizedProps,z=ja(a,c),n.props=z,A=t.pendingProps,g=n.context,m=a.contextType,f=Pa,typeof m=="object"&&m!==null&&(f=Ql(m)),i=a.getDerivedStateFromProps,(m=typeof i=="function"||typeof n.getSnapshotBeforeUpdate=="function")||typeof n.UNSAFE_componentWillReceiveProps!="function"&&typeof n.componentWillReceiveProps!="function"||(c!==A||g!==f)&&go(t,n,e,f),Ft=!1,g=t.memoizedState,n.state=g,we(t,e,n,u),Je();var b=t.memoizedState;c!==A||g!==b||Ft||l!==null&&l.dependencies!==null&&Ju(l.dependencies)?(typeof i=="function"&&(ai(t,a,i,e),b=t.memoizedState),(z=Ft||mo(t,a,z,e,g,b,f)||l!==null&&l.dependencies!==null&&Ju(l.dependencies))?(m||typeof n.UNSAFE_componentWillUpdate!="function"&&typeof n.componentWillUpdate!="function"||(typeof n.componentWillUpdate=="function"&&n.componentWillUpdate(e,b,f),typeof n.UNSAFE_componentWillUpdate=="function"&&n.UNSAFE_componentWillUpdate(e,b,f)),typeof n.componentDidUpdate=="function"&&(t.flags|=4),typeof n.getSnapshotBeforeUpdate=="function"&&(t.flags|=1024)):(typeof n.componentDidUpdate!="function"||c===l.memoizedProps&&g===l.memoizedState||(t.flags|=4),typeof n.getSnapshotBeforeUpdate!="function"||c===l.memoizedProps&&g===l.memoizedState||(t.flags|=1024),t.memoizedProps=e,t.memoizedState=b),n.props=e,n.state=b,n.context=f,e=z):(typeof n.componentDidUpdate!="function"||c===l.memoizedProps&&g===l.memoizedState||(t.flags|=4),typeof n.getSnapshotBeforeUpdate!="function"||c===l.memoizedProps&&g===l.memoizedState||(t.flags|=1024),e=!1)}return n=e,rn(l,t),e=(t.flags&128)!==0,n||e?(n=t.stateNode,a=e&&typeof a.getDerivedStateFromError!="function"?null:n.render(),t.flags|=1,l!==null&&e?(t.child=fe(t,l.child,null,u),t.child=fe(t,null,a,u)):Bl(l,t,a,u),t.memoizedState=n.state,l=t.child):l=Gt(l,t,u),l}function Ro(l,t,a,e){return Ce(),t.flags|=256,Bl(l,t,a,e),t.child}var ci={dehydrated:null,treeContext:null,retryLane:0,hydrationErrors:null};function ii(l){return{baseLanes:l,cachePool:bs()}}function fi(l,t,a){return l=l!==null?l.childLanes&~a:0,t&&(l|=mt),l}function jo(l,t,a){var e=t.pendingProps,u=!1,n=(t.flags&128)!==0,c;if((c=n)||(c=l!==null&&l.memoizedState===null?!1:(_l.current&2)!==0),c&&(u=!0,t.flags&=-129),c=(t.flags&32)!==0,t.flags&=-33,l===null){if(el){if(u?ta(t):aa(),el){var i=Sl,f;if(f=i){l:{for(f=i,i=At;f.nodeType!==8;){if(!i){i=null;break l}if(f=Tt(f.nextSibling),f===null){i=null;break l}}i=f}i!==null?(t.memoizedState={dehydrated:i,treeContext:pa!==null?{id:jt,overflow:Ht}:null,retryLane:536870912,hydrationErrors:null},f=at(18,null,null,0),f.stateNode=i,f.return=t,t.child=f,Jl=t,Sl=null,f=!0):f=!1}f||Ma(t)}if(i=t.memoizedState,i!==null&&(i=i.dehydrated,i!==null))return Ki(i)?t.lanes=32:t.lanes=536870912,null;Ct(t)}return i=e.children,e=e.fallback,u?(aa(),u=t.mode,i=vn({mode:"hidden",children:i},u),e=Oa(e,u,a,null),i.return=t,e.return=t,i.sibling=e,t.child=i,u=t.child,u.memoizedState=ii(a),u.childLanes=fi(l,c,a),t.memoizedState=ci,e):(ta(t),si(t,i))}if(f=l.memoizedState,f!==null&&(i=f.dehydrated,i!==null)){if(n)t.flags&256?(ta(t),t.flags&=-257,t=oi(l,t,a)):t.memoizedState!==null?(aa(),t.child=l.child,t.flags|=128,t=null):(aa(),u=e.fallback,i=t.mode,e=vn({mode:"visible",children:e.children},i),u=Oa(u,i,a,null),u.flags|=2,e.return=t,u.return=t,e.sibling=u,t.child=e,fe(t,l.child,null,a),e=t.child,e.memoizedState=ii(a),e.childLanes=fi(l,c,a),t.memoizedState=ci,t=u);else if(ta(t),Ki(i)){if(c=i.nextSibling&&i.nextSibling.dataset,c)var m=c.dgst;c=m,e=Error(s(419)),e.stack="",e.digest=c,Ge({value:e,source:null,stack:null}),t=oi(l,t,a)}else if(Rl||Xe(l,t,a,!1),c=(a&l.childLanes)!==0,Rl||c){if(c=yl,c!==null&&(e=a&-a,e=(e&42)!==0?1:Jn(e),e=(e&(c.suspendedLanes|a))!==0?0:e,e!==0&&e!==f.retryLane))throw f.retryLane=e,Fa(l,e),it(c,l,e),Oo;i.data==="$?"||Mi(),t=oi(l,t,a)}else i.data==="$?"?(t.flags|=192,t.child=l.child,t=null):(l=f.treeContext,Sl=Tt(i.nextSibling),Jl=t,el=!0,_a=null,At=!1,l!==null&&(vt[yt++]=jt,vt[yt++]=Ht,vt[yt++]=pa,jt=l.id,Ht=l.overflow,pa=t),t=si(t,e.children),t.flags|=4096);return t}return u?(aa(),u=e.fallback,i=t.mode,f=l.child,m=f.sibling,e=Rt(f,{mode:"hidden",children:e.children}),e.subtreeFlags=f.subtreeFlags&65011712,m!==null?u=Rt(m,u):(u=Oa(u,i,a,null),u.flags|=2),u.return=t,e.return=t,e.sibling=u,t.child=e,e=u,u=t.child,i=l.child.memoizedState,i===null?i=ii(a):(f=i.cachePool,f!==null?(m=xl._currentValue,f=f.parent!==m?{parent:m,pool:m}:f):f=bs(),i={baseLanes:i.baseLanes|a,cachePool:f}),u.memoizedState=i,u.childLanes=fi(l,c,a),t.memoizedState=ci,e):(ta(t),a=l.child,l=a.sibling,a=Rt(a,{mode:"visible",children:e.children}),a.return=t,a.sibling=null,l!==null&&(c=t.deletions,c===null?(t.deletions=[l],t.flags|=16):c.push(l)),t.child=a,t.memoizedState=null,a)}function si(l,t){return t=vn({mode:"visible",children:t},l.mode),t.return=l,l.child=t}function vn(l,t){return l=at(22,l,null,t),l.lanes=0,l.stateNode={_visibility:1,_pendingMarkers:null,_retryCache:null,_transitions:null},l}function oi(l,t,a){return fe(t,l.child,null,a),l=si(t,t.pendingProps.children),l.flags|=2,t.memoizedState=null,l}function Ho(l,t,a){l.lanes|=t;var e=l.alternate;e!==null&&(e.lanes|=t),Mc(l.return,t,a)}function di(l,t,a,e,u){var n=l.memoizedState;n===null?l.memoizedState={isBackwards:t,rendering:null,renderingStartTime:0,last:e,tail:a,tailMode:u}:(n.isBackwards=t,n.rendering=null,n.renderingStartTime=0,n.last=e,n.tail=a,n.tailMode=u)}function qo(l,t,a){var e=t.pendingProps,u=e.revealOrder,n=e.tail;if(Bl(l,t,e.children,a),e=_l.current,(e&2)!==0)e=e&1|2,t.flags|=128;else{if(l!==null&&(l.flags&128)!==0)l:for(l=t.child;l!==null;){if(l.tag===13)l.memoizedState!==null&&Ho(l,a,t);else if(l.tag===19)Ho(l,a,t);else if(l.child!==null){l.child.return=l,l=l.child;continue}if(l===t)break l;for(;l.sibling===null;){if(l.return===null||l.return===t)break l;l=l.return}l.sibling.return=l.return,l=l.sibling}e&=1}switch(U(_l,e),u){case"forwards":for(a=t.child,u=null;a!==null;)l=a.alternate,l!==null&&sn(l)===null&&(u=a),a=a.sibling;a=u,a===null?(u=t.child,t.child=null):(u=a.sibling,a.sibling=null),di(t,!1,u,a,n);break;case"backwards":for(a=null,u=t.child,t.child=null;u!==null;){if(l=u.alternate,l!==null&&sn(l)===null){t.child=u;break}l=u.sibling,u.sibling=a,a=u,u=l}di(t,!0,a,null,n);break;case"together":di(t,!1,null,null,void 0);break;default:t.memoizedState=null}return t.child}function Gt(l,t,a){if(l!==null&&(t.dependencies=l.dependencies),ia|=t.lanes,(a&t.childLanes)===0)if(l!==null){if(Xe(l,t,a,!1),(a&t.childLanes)===0)return null}else return null;if(l!==null&&t.child!==l.child)throw Error(s(153));if(t.child!==null){for(l=t.child,a=Rt(l,l.pendingProps),t.child=a,a.return=t;l.sibling!==null;)l=l.sibling,a=a.sibling=Rt(l,l.pendingProps),a.return=t;a.sibling=null}return t.child}function ri(l,t){return(l.lanes&t)!==0?!0:(l=l.dependencies,!!(l!==null&&Ju(l)))}function V0(l,t,a){switch(t.tag){case 3:ml(t,t.stateNode.containerInfo),$t(t,xl,l.memoizedState.cache),Ce();break;case 27:case 5:Qn(t);break;case 4:ml(t,t.stateNode.containerInfo);break;case 10:$t(t,t.type,t.memoizedProps.value);break;case 13:var e=t.memoizedState;if(e!==null)return e.dehydrated!==null?(ta(t),t.flags|=128,null):(a&t.child.childLanes)!==0?jo(l,t,a):(ta(t),l=Gt(l,t,a),l!==null?l.sibling:null);ta(t);break;case 19:var u=(l.flags&128)!==0;if(e=(a&t.childLanes)!==0,e||(Xe(l,t,a,!1),e=(a&t.childLanes)!==0),u){if(e)return qo(l,t,a);t.flags|=128}if(u=t.memoizedState,u!==null&&(u.rendering=null,u.tail=null,u.lastEffect=null),U(_l,_l.current),e)break;return null;case 22:case 23:return t.lanes=0,Mo(l,t,a);case 24:$t(t,xl,l.memoizedState.cache)}return Gt(l,t,a)}function Yo(l,t,a){if(l!==null)if(l.memoizedProps!==t.pendingProps)Rl=!0;else{if(!ri(l,a)&&(t.flags&128)===0)return Rl=!1,V0(l,t,a);Rl=(l.flags&131072)!==0}else Rl=!1,el&&(t.flags&1048576)!==0&&ds(t,Ku,t.index);switch(t.lanes=0,t.tag){case 16:l:{l=t.pendingProps;var e=t.elementType,u=e._init;if(e=u(e._payload),t.type=e,typeof e=="function")Tc(e)?(l=ja(e,l),t.tag=1,t=Uo(null,t,e,l,a)):(t.tag=0,t=ni(null,t,e,l,a));else{if(e!=null){if(u=e.$$typeof,u===Vl){t.tag=11,t=po(null,t,e,l,a);break l}else if(u===Ll){t.tag=14,t=xo(null,t,e,l,a);break l}}throw t=K(e)||e,Error(s(306,t,""))}}return t;case 0:return ni(l,t,t.type,t.pendingProps,a);case 1:return e=t.type,u=ja(e,t.pendingProps),Uo(l,t,e,u,a);case 3:l:{if(ml(t,t.stateNode.containerInfo),l===null)throw Error(s(387));e=t.pendingProps;var n=t.memoizedState;u=n.element,qc(l,t),we(t,e,null,a);var c=t.memoizedState;if(e=c.cache,$t(t,xl,e),e!==n.cache&&Dc(t,[xl],a,!0),Je(),e=c.element,n.isDehydrated)if(n={element:e,isDehydrated:!1,cache:c.cache},t.updateQueue.baseState=n,t.memoizedState=n,t.flags&256){t=Ro(l,t,e,a);break l}else if(e!==u){u=dt(Error(s(424)),t),Ge(u),t=Ro(l,t,e,a);break l}else{switch(l=t.stateNode.containerInfo,l.nodeType){case 9:l=l.body;break;default:l=l.nodeName==="HTML"?l.ownerDocument.body:l}for(Sl=Tt(l.firstChild),Jl=t,el=!0,_a=null,At=!0,a=yo(t,null,e,a),t.child=a;a;)a.flags=a.flags&-3|4096,a=a.sibling}else{if(Ce(),e===u){t=Gt(l,t,a);break l}Bl(l,t,e,a)}t=t.child}return t;case 26:return rn(l,t),l===null?(a=Xd(t.type,null,t.pendingProps,null))?t.memoizedState=a:el||(a=t.type,l=t.pendingProps,e=_n(J.current).createElement(a),e[Xl]=t,e[wl]=l,Gl(e,a,l),Ul(e),t.stateNode=e):t.memoizedState=Xd(t.type,l.memoizedProps,t.pendingProps,l.memoizedState),null;case 27:return Qn(t),l===null&&el&&(e=t.stateNode=Bd(t.type,t.pendingProps,J.current),Jl=t,At=!0,u=Sl,da(t.type)?(Ji=u,Sl=Tt(e.firstChild)):Sl=u),Bl(l,t,t.pendingProps.children,a),rn(l,t),l===null&&(t.flags|=4194304),t.child;case 5:return l===null&&el&&((u=e=Sl)&&(e=gv(e,t.type,t.pendingProps,At),e!==null?(t.stateNode=e,Jl=t,Sl=Tt(e.firstChild),At=!1,u=!0):u=!1),u||Ma(t)),Qn(t),u=t.type,n=t.pendingProps,c=l!==null?l.memoizedProps:null,e=n.children,Zi(u,n)?e=null:c!==null&&Zi(u,c)&&(t.flags|=32),t.memoizedState!==null&&(u=Qc(l,t,q0,null,null,a),hu._currentValue=u),rn(l,t),Bl(l,t,e,a),t.child;case 6:return l===null&&el&&((l=a=Sl)&&(a=bv(a,t.pendingProps,At),a!==null?(t.stateNode=a,Jl=t,Sl=null,l=!0):l=!1),l||Ma(t)),null;case 13:return jo(l,t,a);case 4:return ml(t,t.stateNode.containerInfo),e=t.pendingProps,l===null?t.child=fe(t,null,e,a):Bl(l,t,e,a),t.child;case 11:return po(l,t,t.type,t.pendingProps,a);case 7:return Bl(l,t,t.pendingProps,a),t.child;case 8:return Bl(l,t,t.pendingProps.children,a),t.child;case 12:return Bl(l,t,t.pendingProps.children,a),t.child;case 10:return e=t.pendingProps,$t(t,t.type,e.value),Bl(l,t,e.children,a),t.child;case 9:return u=t.type._context,e=t.pendingProps.children,Na(t),u=Ql(u),e=e(u),t.flags|=1,Bl(l,t,e,a),t.child;case 14:return xo(l,t,t.type,t.pendingProps,a);case 15:return _o(l,t,t.type,t.pendingProps,a);case 19:return qo(l,t,a);case 31:return e=t.pendingProps,a=t.mode,e={mode:e.mode,children:e.children},l===null?(a=vn(e,a),a.ref=t.ref,t.child=a,a.return=t,t=a):(a=Rt(l.child,e),a.ref=t.ref,t.child=a,a.return=t,t=a),t;case 22:return Mo(l,t,a);case 24:return Na(t),e=Ql(xl),l===null?(u=Rc(),u===null&&(u=yl,n=Nc(),u.pooledCache=n,n.refCount++,n!==null&&(u.pooledCacheLanes|=a),u=n),t.memoizedState={parent:e,cache:u},Hc(t),$t(t,xl,u)):((l.lanes&a)!==0&&(qc(l,t),we(t,null,null,a),Je()),u=l.memoizedState,n=t.memoizedState,u.parent!==e?(u={parent:e,cache:e},t.memoizedState=u,t.lanes===0&&(t.memoizedState=t.updateQueue.baseState=u),$t(t,xl,e)):(e=n.cache,$t(t,xl,e),e!==u.cache&&Dc(t,[xl],a,!0))),Bl(l,t,t.pendingProps.children,a),t.child;case 29:throw t.pendingProps}throw Error(s(156,t.tag))}function Xt(l){l.flags|=4}function Bo(l,t){if(t.type!=="stylesheet"||(t.state.loading&4)!==0)l.flags&=-16777217;else if(l.flags|=16777216,!Kd(t)){if(t=ht.current,t!==null&&((I&4194048)===I?Ot!==null:(I&62914560)!==I&&(I&536870912)===0||t!==Ot))throw Le=jc,Ss;l.flags|=8192}}function yn(l,t){t!==null&&(l.flags|=4),l.flags&16384&&(t=l.tag!==22?yf():536870912,l.lanes|=t,re|=t)}function lu(l,t){if(!el)switch(l.tailMode){case"hidden":t=l.tail;for(var a=null;t!==null;)t.alternate!==null&&(a=t),t=t.sibling;a===null?l.tail=null:a.sibling=null;break;case"collapsed":a=l.tail;for(var e=null;a!==null;)a.alternate!==null&&(e=a),a=a.sibling;e===null?t||l.tail===null?l.tail=null:l.tail.sibling=null:e.sibling=null}}function bl(l){var t=l.alternate!==null&&l.alternate.child===l.child,a=0,e=0;if(t)for(var u=l.child;u!==null;)a|=u.lanes|u.childLanes,e|=u.subtreeFlags&65011712,e|=u.flags&65011712,u.return=l,u=u.sibling;else for(u=l.child;u!==null;)a|=u.lanes|u.childLanes,e|=u.subtreeFlags,e|=u.flags,u.return=l,u=u.sibling;return l.subtreeFlags|=e,l.childLanes=a,t}function L0(l,t,a){var e=t.pendingProps;switch(pc(t),t.tag){case 31:case 16:case 15:case 0:case 11:case 7:case 8:case 12:case 9:case 14:return bl(t),null;case 1:return bl(t),null;case 3:return a=t.stateNode,e=null,l!==null&&(e=l.memoizedState.cache),t.memoizedState.cache!==e&&(t.flags|=2048),Yt(xl),Jt(),a.pendingContext&&(a.context=a.pendingContext,a.pendingContext=null),(l===null||l.child===null)&&(Be(t)?Xt(t):l===null||l.memoizedState.isDehydrated&&(t.flags&256)===0||(t.flags|=1024,ys())),bl(t),null;case 26:return a=t.memoizedState,l===null?(Xt(t),a!==null?(bl(t),Bo(t,a)):(bl(t),t.flags&=-16777217)):a?a!==l.memoizedState?(Xt(t),bl(t),Bo(t,a)):(bl(t),t.flags&=-16777217):(l.memoizedProps!==e&&Xt(t),bl(t),t.flags&=-16777217),null;case 27:Ou(t),a=J.current;var u=t.type;if(l!==null&&t.stateNode!=null)l.memoizedProps!==e&&Xt(t);else{if(!e){if(t.stateNode===null)throw Error(s(166));return bl(t),null}l=G.current,Be(t)?rs(t):(l=Bd(u,e,a),t.stateNode=l,Xt(t))}return bl(t),null;case 5:if(Ou(t),a=t.type,l!==null&&t.stateNode!=null)l.memoizedProps!==e&&Xt(t);else{if(!e){if(t.stateNode===null)throw Error(s(166));return bl(t),null}if(l=G.current,Be(t))rs(t);else{switch(u=_n(J.current),l){case 1:l=u.createElementNS("http://www.w3.org/2000/svg",a);break;case 2:l=u.createElementNS("http://www.w3.org/1998/Math/MathML",a);break;default:switch(a){case"svg":l=u.createElementNS("http://www.w3.org/2000/svg",a);break;case"math":l=u.createElementNS("http://www.w3.org/1998/Math/MathML",a);break;case"script":l=u.createElement("div"),l.innerHTML="<script><\/script>",l=l.removeChild(l.firstChild);break;case"select":l=typeof e.is=="string"?u.createElement("select",{is:e.is}):u.createElement("select"),e.multiple?l.multiple=!0:e.size&&(l.size=e.size);break;default:l=typeof e.is=="string"?u.createElement(a,{is:e.is}):u.createElement(a)}}l[Xl]=t,l[wl]=e;l:for(u=t.child;u!==null;){if(u.tag===5||u.tag===6)l.appendChild(u.stateNode);else if(u.tag!==4&&u.tag!==27&&u.child!==null){u.child.return=u,u=u.child;continue}if(u===t)break l;for(;u.sibling===null;){if(u.return===null||u.return===t)break l;u=u.return}u.sibling.return=u.return,u=u.sibling}t.stateNode=l;l:switch(Gl(l,a,e),a){case"button":case"input":case"select":case"textarea":l=!!e.autoFocus;break l;case"img":l=!0;break l;default:l=!1}l&&Xt(t)}}return bl(t),t.flags&=-16777217,null;case 6:if(l&&t.stateNode!=null)l.memoizedProps!==e&&Xt(t);else{if(typeof e!="string"&&t.stateNode===null)throw Error(s(166));if(l=J.current,Be(t)){if(l=t.stateNode,a=t.memoizedProps,e=null,u=Jl,u!==null)switch(u.tag){case 27:case 5:e=u.memoizedProps}l[Xl]=t,l=!!(l.nodeValue===a||e!==null&&e.suppressHydrationWarning===!0||Nd(l.nodeValue,a)),l||Ma(t)}else l=_n(l).createTextNode(e),l[Xl]=t,t.stateNode=l}return bl(t),null;case 13:if(e=t.memoizedState,l===null||l.memoizedState!==null&&l.memoizedState.dehydrated!==null){if(u=Be(t),e!==null&&e.dehydrated!==null){if(l===null){if(!u)throw Error(s(318));if(u=t.memoizedState,u=u!==null?u.dehydrated:null,!u)throw Error(s(317));u[Xl]=t}else Ce(),(t.flags&128)===0&&(t.memoizedState=null),t.flags|=4;bl(t),u=!1}else u=ys(),l!==null&&l.memoizedState!==null&&(l.memoizedState.hydrationErrors=u),u=!0;if(!u)return t.flags&256?(Ct(t),t):(Ct(t),null)}if(Ct(t),(t.flags&128)!==0)return t.lanes=a,t;if(a=e!==null,l=l!==null&&l.memoizedState!==null,a){e=t.child,u=null,e.alternate!==null&&e.alternate.memoizedState!==null&&e.alternate.memoizedState.cachePool!==null&&(u=e.alternate.memoizedState.cachePool.pool);var n=null;e.memoizedState!==null&&e.memoizedState.cachePool!==null&&(n=e.memoizedState.cachePool.pool),n!==u&&(e.flags|=2048)}return a!==l&&a&&(t.child.flags|=8192),yn(t,t.updateQueue),bl(t),null;case 4:return Jt(),l===null&&Bi(t.stateNode.containerInfo),bl(t),null;case 10:return Yt(t.type),bl(t),null;case 19:if(H(_l),u=t.memoizedState,u===null)return bl(t),null;if(e=(t.flags&128)!==0,n=u.rendering,n===null)if(e)lu(u,!1);else{if(zl!==0||l!==null&&(l.flags&128)!==0)for(l=t.child;l!==null;){if(n=sn(l),n!==null){for(t.flags|=128,lu(u,!1),l=n.updateQueue,t.updateQueue=l,yn(t,l),t.subtreeFlags=0,l=a,a=t.child;a!==null;)os(a,l),a=a.sibling;return U(_l,_l.current&1|2),t.child}l=l.sibling}u.tail!==null&&Et()>gn&&(t.flags|=128,e=!0,lu(u,!1),t.lanes=4194304)}else{if(!e)if(l=sn(n),l!==null){if(t.flags|=128,e=!0,l=l.updateQueue,t.updateQueue=l,yn(t,l),lu(u,!0),u.tail===null&&u.tailMode==="hidden"&&!n.alternate&&!el)return bl(t),null}else 2*Et()-u.renderingStartTime>gn&&a!==536870912&&(t.flags|=128,e=!0,lu(u,!1),t.lanes=4194304);u.isBackwards?(n.sibling=t.child,t.child=n):(l=u.last,l!==null?l.sibling=n:t.child=n,u.last=n)}return u.tail!==null?(t=u.tail,u.rendering=t,u.tail=t.sibling,u.renderingStartTime=Et(),t.sibling=null,l=_l.current,U(_l,e?l&1|2:l&1),t):(bl(t),null);case 22:case 23:return Ct(t),Gc(),e=t.memoizedState!==null,l!==null?l.memoizedState!==null!==e&&(t.flags|=8192):e&&(t.flags|=8192),e?(a&536870912)!==0&&(t.flags&128)===0&&(bl(t),t.subtreeFlags&6&&(t.flags|=8192)):bl(t),a=t.updateQueue,a!==null&&yn(t,a.retryQueue),a=null,l!==null&&l.memoizedState!==null&&l.memoizedState.cachePool!==null&&(a=l.memoizedState.cachePool.pool),e=null,t.memoizedState!==null&&t.memoizedState.cachePool!==null&&(e=t.memoizedState.cachePool.pool),e!==a&&(t.flags|=2048),l!==null&&H(Ua),null;case 24:return a=null,l!==null&&(a=l.memoizedState.cache),t.memoizedState.cache!==a&&(t.flags|=2048),Yt(xl),bl(t),null;case 25:return null;case 30:return null}throw Error(s(156,t.tag))}function K0(l,t){switch(pc(t),t.tag){case 1:return l=t.flags,l&65536?(t.flags=l&-65537|128,t):null;case 3:return Yt(xl),Jt(),l=t.flags,(l&65536)!==0&&(l&128)===0?(t.flags=l&-65537|128,t):null;case 26:case 27:case 5:return Ou(t),null;case 13:if(Ct(t),l=t.memoizedState,l!==null&&l.dehydrated!==null){if(t.alternate===null)throw Error(s(340));Ce()}return l=t.flags,l&65536?(t.flags=l&-65537|128,t):null;case 19:return H(_l),null;case 4:return Jt(),null;case 10:return Yt(t.type),null;case 22:case 23:return Ct(t),Gc(),l!==null&&H(Ua),l=t.flags,l&65536?(t.flags=l&-65537|128,t):null;case 24:return Yt(xl),null;case 25:return null;default:return null}}function Co(l,t){switch(pc(t),t.tag){case 3:Yt(xl),Jt();break;case 26:case 27:case 5:Ou(t);break;case 4:Jt();break;case 13:Ct(t);break;case 19:H(_l);break;case 10:Yt(t.type);break;case 22:case 23:Ct(t),Gc(),l!==null&&H(Ua);break;case 24:Yt(xl)}}function tu(l,t){try{var a=t.updateQueue,e=a!==null?a.lastEffect:null;if(e!==null){var u=e.next;a=u;do{if((a.tag&l)===l){e=void 0;var n=a.create,c=a.inst;e=n(),c.destroy=e}a=a.next}while(a!==u)}}catch(i){vl(t,t.return,i)}}function ea(l,t,a){try{var e=t.updateQueue,u=e!==null?e.lastEffect:null;if(u!==null){var n=u.next;e=n;do{if((e.tag&l)===l){var c=e.inst,i=c.destroy;if(i!==void 0){c.destroy=void 0,u=t;var f=a,m=i;try{m()}catch(z){vl(u,f,z)}}}e=e.next}while(e!==n)}}catch(z){vl(t,t.return,z)}}function Go(l){var t=l.updateQueue;if(t!==null){var a=l.stateNode;try{ps(t,a)}catch(e){vl(l,l.return,e)}}}function Xo(l,t,a){a.props=ja(l.type,l.memoizedProps),a.state=l.memoizedState;try{a.componentWillUnmount()}catch(e){vl(l,t,e)}}function au(l,t){try{var a=l.ref;if(a!==null){switch(l.tag){case 26:case 27:case 5:var e=l.stateNode;break;case 30:e=l.stateNode;break;default:e=l.stateNode}typeof a=="function"?l.refCleanup=a(e):a.current=e}}catch(u){vl(l,t,u)}}function pt(l,t){var a=l.ref,e=l.refCleanup;if(a!==null)if(typeof e=="function")try{e()}catch(u){vl(l,t,u)}finally{l.refCleanup=null,l=l.alternate,l!=null&&(l.refCleanup=null)}else if(typeof a=="function")try{a(null)}catch(u){vl(l,t,u)}else a.current=null}function Qo(l){var t=l.type,a=l.memoizedProps,e=l.stateNode;try{l:switch(t){case"button":case"input":case"select":case"textarea":a.autoFocus&&e.focus();break l;case"img":a.src?e.src=a.src:a.srcSet&&(e.srcset=a.srcSet)}}catch(u){vl(l,l.return,u)}}function vi(l,t,a){try{var e=l.stateNode;rv(e,l.type,a,t),e[wl]=t}catch(u){vl(l,l.return,u)}}function Zo(l){return l.tag===5||l.tag===3||l.tag===26||l.tag===27&&da(l.type)||l.tag===4}function yi(l){l:for(;;){for(;l.sibling===null;){if(l.return===null||Zo(l.return))return null;l=l.return}for(l.sibling.return=l.return,l=l.sibling;l.tag!==5&&l.tag!==6&&l.tag!==18;){if(l.tag===27&&da(l.type)||l.flags&2||l.child===null||l.tag===4)continue l;l.child.return=l,l=l.child}if(!(l.flags&2))return l.stateNode}}function hi(l,t,a){var e=l.tag;if(e===5||e===6)l=l.stateNode,t?(a.nodeType===9?a.body:a.nodeName==="HTML"?a.ownerDocument.body:a).insertBefore(l,t):(t=a.nodeType===9?a.body:a.nodeName==="HTML"?a.ownerDocument.body:a,t.appendChild(l),a=a._reactRootContainer,a!=null||t.onclick!==null||(t.onclick=xn));else if(e!==4&&(e===27&&da(l.type)&&(a=l.stateNode,t=null),l=l.child,l!==null))for(hi(l,t,a),l=l.sibling;l!==null;)hi(l,t,a),l=l.sibling}function hn(l,t,a){var e=l.tag;if(e===5||e===6)l=l.stateNode,t?a.insertBefore(l,t):a.appendChild(l);else if(e!==4&&(e===27&&da(l.type)&&(a=l.stateNode),l=l.child,l!==null))for(hn(l,t,a),l=l.sibling;l!==null;)hn(l,t,a),l=l.sibling}function Vo(l){var t=l.stateNode,a=l.memoizedProps;try{for(var e=l.type,u=t.attributes;u.length;)t.removeAttributeNode(u[0]);Gl(t,e,a),t[Xl]=l,t[wl]=a}catch(n){vl(l,l.return,n)}}var Qt=!1,El=!1,mi=!1,Lo=typeof WeakSet=="function"?WeakSet:Set,jl=null;function J0(l,t){if(l=l.containerInfo,Xi=jn,l=ls(l),yc(l)){if("selectionStart"in l)var a={start:l.selectionStart,end:l.selectionEnd};else l:{a=(a=l.ownerDocument)&&a.defaultView||window;var e=a.getSelection&&a.getSelection();if(e&&e.rangeCount!==0){a=e.anchorNode;var u=e.anchorOffset,n=e.focusNode;e=e.focusOffset;try{a.nodeType,n.nodeType}catch{a=null;break l}var c=0,i=-1,f=-1,m=0,z=0,A=l,g=null;t:for(;;){for(var b;A!==a||u!==0&&A.nodeType!==3||(i=c+u),A!==n||e!==0&&A.nodeType!==3||(f=c+e),A.nodeType===3&&(c+=A.nodeValue.length),(b=A.firstChild)!==null;)g=A,A=b;for(;;){if(A===l)break t;if(g===a&&++m===u&&(i=c),g===n&&++z===e&&(f=c),(b=A.nextSibling)!==null)break;A=g,g=A.parentNode}A=b}a=i===-1||f===-1?null:{start:i,end:f}}else a=null}a=a||{start:0,end:0}}else a=null;for(Qi={focusedElem:l,selectionRange:a},jn=!1,jl=t;jl!==null;)if(t=jl,l=t.child,(t.subtreeFlags&1024)!==0&&l!==null)l.return=t,jl=l;else for(;jl!==null;){switch(t=jl,n=t.alternate,l=t.flags,t.tag){case 0:break;case 11:case 15:break;case 1:if((l&1024)!==0&&n!==null){l=void 0,a=t,u=n.memoizedProps,n=n.memoizedState,e=a.stateNode;try{var L=ja(a.type,u,a.elementType===a.type);l=e.getSnapshotBeforeUpdate(L,n),e.__reactInternalSnapshotBeforeUpdate=l}catch(X){vl(a,a.return,X)}}break;case 3:if((l&1024)!==0){if(l=t.stateNode.containerInfo,a=l.nodeType,a===9)Li(l);else if(a===1)switch(l.nodeName){case"HEAD":case"HTML":case"BODY":Li(l);break;default:l.textContent=""}}break;case 5:case 26:case 27:case 6:case 4:case 17:break;default:if((l&1024)!==0)throw Error(s(163))}if(l=t.sibling,l!==null){l.return=t.return,jl=l;break}jl=t.return}}function Ko(l,t,a){var e=a.flags;switch(a.tag){case 0:case 11:case 15:ua(l,a),e&4&&tu(5,a);break;case 1:if(ua(l,a),e&4)if(l=a.stateNode,t===null)try{l.componentDidMount()}catch(c){vl(a,a.return,c)}else{var u=ja(a.type,t.memoizedProps);t=t.memoizedState;try{l.componentDidUpdate(u,t,l.__reactInternalSnapshotBeforeUpdate)}catch(c){vl(a,a.return,c)}}e&64&&Go(a),e&512&&au(a,a.return);break;case 3:if(ua(l,a),e&64&&(l=a.updateQueue,l!==null)){if(t=null,a.child!==null)switch(a.child.tag){case 27:case 5:t=a.child.stateNode;break;case 1:t=a.child.stateNode}try{ps(l,t)}catch(c){vl(a,a.return,c)}}break;case 27:t===null&&e&4&&Vo(a);case 26:case 5:ua(l,a),t===null&&e&4&&Qo(a),e&512&&au(a,a.return);break;case 12:ua(l,a);break;case 13:ua(l,a),e&4&&Wo(l,a),e&64&&(l=a.memoizedState,l!==null&&(l=l.dehydrated,l!==null&&(a=tv.bind(null,a),Sv(l,a))));break;case 22:if(e=a.memoizedState!==null||Qt,!e){t=t!==null&&t.memoizedState!==null||El,u=Qt;var n=El;Qt=e,(El=t)&&!n?na(l,a,(a.subtreeFlags&8772)!==0):ua(l,a),Qt=u,El=n}break;case 30:break;default:ua(l,a)}}function Jo(l){var t=l.alternate;t!==null&&(l.alternate=null,Jo(t)),l.child=null,l.deletions=null,l.sibling=null,l.tag===5&&(t=l.stateNode,t!==null&&kn(t)),l.stateNode=null,l.return=null,l.dependencies=null,l.memoizedProps=null,l.memoizedState=null,l.pendingProps=null,l.stateNode=null,l.updateQueue=null}var gl=null,$l=!1;function Zt(l,t,a){for(a=a.child;a!==null;)wo(l,t,a),a=a.sibling}function wo(l,t,a){if(Il&&typeof Il.onCommitFiberUnmount=="function")try{Il.onCommitFiberUnmount(Ae,a)}catch{}switch(a.tag){case 26:El||pt(a,t),Zt(l,t,a),a.memoizedState?a.memoizedState.count--:a.stateNode&&(a=a.stateNode,a.parentNode.removeChild(a));break;case 27:El||pt(a,t);var e=gl,u=$l;da(a.type)&&(gl=a.stateNode,$l=!1),Zt(l,t,a),du(a.stateNode),gl=e,$l=u;break;case 5:El||pt(a,t);case 6:if(e=gl,u=$l,gl=null,Zt(l,t,a),gl=e,$l=u,gl!==null)if($l)try{(gl.nodeType===9?gl.body:gl.nodeName==="HTML"?gl.ownerDocument.body:gl).removeChild(a.stateNode)}catch(n){vl(a,t,n)}else try{gl.removeChild(a.stateNode)}catch(n){vl(a,t,n)}break;case 18:gl!==null&&($l?(l=gl,qd(l.nodeType===9?l.body:l.nodeName==="HTML"?l.ownerDocument.body:l,a.stateNode),Su(l)):qd(gl,a.stateNode));break;case 4:e=gl,u=$l,gl=a.stateNode.containerInfo,$l=!0,Zt(l,t,a),gl=e,$l=u;break;case 0:case 11:case 14:case 15:El||ea(2,a,t),El||ea(4,a,t),Zt(l,t,a);break;case 1:El||(pt(a,t),e=a.stateNode,typeof e.componentWillUnmount=="function"&&Xo(a,t,e)),Zt(l,t,a);break;case 21:Zt(l,t,a);break;case 22:El=(e=El)||a.memoizedState!==null,Zt(l,t,a),El=e;break;default:Zt(l,t,a)}}function Wo(l,t){if(t.memoizedState===null&&(l=t.alternate,l!==null&&(l=l.memoizedState,l!==null&&(l=l.dehydrated,l!==null))))try{Su(l)}catch(a){vl(t,t.return,a)}}function w0(l){switch(l.tag){case 13:case 19:var t=l.stateNode;return t===null&&(t=l.stateNode=new Lo),t;case 22:return l=l.stateNode,t=l._retryCache,t===null&&(t=l._retryCache=new Lo),t;default:throw Error(s(435,l.tag))}}function gi(l,t){var a=w0(l);t.forEach(function(e){var u=av.bind(null,l,e);a.has(e)||(a.add(e),e.then(u,u))})}function et(l,t){var a=t.deletions;if(a!==null)for(var e=0;e<a.length;e++){var u=a[e],n=l,c=t,i=c;l:for(;i!==null;){switch(i.tag){case 27:if(da(i.type)){gl=i.stateNode,$l=!1;break l}break;case 5:gl=i.stateNode,$l=!1;break l;case 3:case 4:gl=i.stateNode.containerInfo,$l=!0;break l}i=i.return}if(gl===null)throw Error(s(160));wo(n,c,u),gl=null,$l=!1,n=u.alternate,n!==null&&(n.return=null),u.return=null}if(t.subtreeFlags&13878)for(t=t.child;t!==null;)ko(t,l),t=t.sibling}var zt=null;function ko(l,t){var a=l.alternate,e=l.flags;switch(l.tag){case 0:case 11:case 14:case 15:et(t,l),ut(l),e&4&&(ea(3,l,l.return),tu(3,l),ea(5,l,l.return));break;case 1:et(t,l),ut(l),e&512&&(El||a===null||pt(a,a.return)),e&64&&Qt&&(l=l.updateQueue,l!==null&&(e=l.callbacks,e!==null&&(a=l.shared.hiddenCallbacks,l.shared.hiddenCallbacks=a===null?e:a.concat(e))));break;case 26:var u=zt;if(et(t,l),ut(l),e&512&&(El||a===null||pt(a,a.return)),e&4){var n=a!==null?a.memoizedState:null;if(e=l.memoizedState,a===null)if(e===null)if(l.stateNode===null){l:{e=l.type,a=l.memoizedProps,u=u.ownerDocument||u;t:switch(e){case"title":n=u.getElementsByTagName("title")[0],(!n||n[xe]||n[Xl]||n.namespaceURI==="http://www.w3.org/2000/svg"||n.hasAttribute("itemprop"))&&(n=u.createElement(e),u.head.insertBefore(n,u.querySelector("head > title"))),Gl(n,e,a),n[Xl]=l,Ul(n),e=n;break l;case"link":var c=Vd("link","href",u).get(e+(a.href||""));if(c){for(var i=0;i<c.length;i++)if(n=c[i],n.getAttribute("href")===(a.href==null||a.href===""?null:a.href)&&n.getAttribute("rel")===(a.rel==null?null:a.rel)&&n.getAttribute("title")===(a.title==null?null:a.title)&&n.getAttribute("crossorigin")===(a.crossOrigin==null?null:a.crossOrigin)){c.splice(i,1);break t}}n=u.createElement(e),Gl(n,e,a),u.head.appendChild(n);break;case"meta":if(c=Vd("meta","content",u).get(e+(a.content||""))){for(i=0;i<c.length;i++)if(n=c[i],n.getAttribute("content")===(a.content==null?null:""+a.content)&&n.getAttribute("name")===(a.name==null?null:a.name)&&n.getAttribute("property")===(a.property==null?null:a.property)&&n.getAttribute("http-equiv")===(a.httpEquiv==null?null:a.httpEquiv)&&n.getAttribute("charset")===(a.charSet==null?null:a.charSet)){c.splice(i,1);break t}}n=u.createElement(e),Gl(n,e,a),u.head.appendChild(n);break;default:throw Error(s(468,e))}n[Xl]=l,Ul(n),e=n}l.stateNode=e}else Ld(u,l.type,l.stateNode);else l.stateNode=Zd(u,e,l.memoizedProps);else n!==e?(n===null?a.stateNode!==null&&(a=a.stateNode,a.parentNode.removeChild(a)):n.count--,e===null?Ld(u,l.type,l.stateNode):Zd(u,e,l.memoizedProps)):e===null&&l.stateNode!==null&&vi(l,l.memoizedProps,a.memoizedProps)}break;case 27:et(t,l),ut(l),e&512&&(El||a===null||pt(a,a.return)),a!==null&&e&4&&vi(l,l.memoizedProps,a.memoizedProps);break;case 5:if(et(t,l),ut(l),e&512&&(El||a===null||pt(a,a.return)),l.flags&32){u=l.stateNode;try{La(u,"")}catch(b){vl(l,l.return,b)}}e&4&&l.stateNode!=null&&(u=l.memoizedProps,vi(l,u,a!==null?a.memoizedProps:u)),e&1024&&(mi=!0);break;case 6:if(et(t,l),ut(l),e&4){if(l.stateNode===null)throw Error(s(162));e=l.memoizedProps,a=l.stateNode;try{a.nodeValue=e}catch(b){vl(l,l.return,b)}}break;case 3:if(Nn=null,u=zt,zt=Mn(t.containerInfo),et(t,l),zt=u,ut(l),e&4&&a!==null&&a.memoizedState.isDehydrated)try{Su(t.containerInfo)}catch(b){vl(l,l.return,b)}mi&&(mi=!1,$o(l));break;case 4:e=zt,zt=Mn(l.stateNode.containerInfo),et(t,l),ut(l),zt=e;break;case 12:et(t,l),ut(l);break;case 13:et(t,l),ut(l),l.child.flags&8192&&l.memoizedState!==null!=(a!==null&&a.memoizedState!==null)&&(Ai=Et()),e&4&&(e=l.updateQueue,e!==null&&(l.updateQueue=null,gi(l,e)));break;case 22:u=l.memoizedState!==null;var f=a!==null&&a.memoizedState!==null,m=Qt,z=El;if(Qt=m||u,El=z||f,et(t,l),El=z,Qt=m,ut(l),e&8192)l:for(t=l.stateNode,t._visibility=u?t._visibility&-2:t._visibility|1,u&&(a===null||f||Qt||El||Ha(l)),a=null,t=l;;){if(t.tag===5||t.tag===26){if(a===null){f=a=t;try{if(n=f.stateNode,u)c=n.style,typeof c.setProperty=="function"?c.setProperty("display","none","important"):c.display="none";else{i=f.stateNode;var A=f.memoizedProps.style,g=A!=null&&A.hasOwnProperty("display")?A.display:null;i.style.display=g==null||typeof g=="boolean"?"":(""+g).trim()}}catch(b){vl(f,f.return,b)}}}else if(t.tag===6){if(a===null){f=t;try{f.stateNode.nodeValue=u?"":f.memoizedProps}catch(b){vl(f,f.return,b)}}}else if((t.tag!==22&&t.tag!==23||t.memoizedState===null||t===l)&&t.child!==null){t.child.return=t,t=t.child;continue}if(t===l)break l;for(;t.sibling===null;){if(t.return===null||t.return===l)break l;a===t&&(a=null),t=t.return}a===t&&(a=null),t.sibling.return=t.return,t=t.sibling}e&4&&(e=l.updateQueue,e!==null&&(a=e.retryQueue,a!==null&&(e.retryQueue=null,gi(l,a))));break;case 19:et(t,l),ut(l),e&4&&(e=l.updateQueue,e!==null&&(l.updateQueue=null,gi(l,e)));break;case 30:break;case 21:break;default:et(t,l),ut(l)}}function ut(l){var t=l.flags;if(t&2){try{for(var a,e=l.return;e!==null;){if(Zo(e)){a=e;break}e=e.return}if(a==null)throw Error(s(160));switch(a.tag){case 27:var u=a.stateNode,n=yi(l);hn(l,n,u);break;case 5:var c=a.stateNode;a.flags&32&&(La(c,""),a.flags&=-33);var i=yi(l);hn(l,i,c);break;case 3:case 4:var f=a.stateNode.containerInfo,m=yi(l);hi(l,m,f);break;default:throw Error(s(161))}}catch(z){vl(l,l.return,z)}l.flags&=-3}t&4096&&(l.flags&=-4097)}function $o(l){if(l.subtreeFlags&1024)for(l=l.child;l!==null;){var t=l;$o(t),t.tag===5&&t.flags&1024&&t.stateNode.reset(),l=l.sibling}}function ua(l,t){if(t.subtreeFlags&8772)for(t=t.child;t!==null;)Ko(l,t.alternate,t),t=t.sibling}function Ha(l){for(l=l.child;l!==null;){var t=l;switch(t.tag){case 0:case 11:case 14:case 15:ea(4,t,t.return),Ha(t);break;case 1:pt(t,t.return);var a=t.stateNode;typeof a.componentWillUnmount=="function"&&Xo(t,t.return,a),Ha(t);break;case 27:du(t.stateNode);case 26:case 5:pt(t,t.return),Ha(t);break;case 22:t.memoizedState===null&&Ha(t);break;case 30:Ha(t);break;default:Ha(t)}l=l.sibling}}function na(l,t,a){for(a=a&&(t.subtreeFlags&8772)!==0,t=t.child;t!==null;){var e=t.alternate,u=l,n=t,c=n.flags;switch(n.tag){case 0:case 11:case 15:na(u,n,a),tu(4,n);break;case 1:if(na(u,n,a),e=n,u=e.stateNode,typeof u.componentDidMount=="function")try{u.componentDidMount()}catch(m){vl(e,e.return,m)}if(e=n,u=e.updateQueue,u!==null){var i=e.stateNode;try{var f=u.shared.hiddenCallbacks;if(f!==null)for(u.shared.hiddenCallbacks=null,u=0;u<f.length;u++)Os(f[u],i)}catch(m){vl(e,e.return,m)}}a&&c&64&&Go(n),au(n,n.return);break;case 27:Vo(n);case 26:case 5:na(u,n,a),a&&e===null&&c&4&&Qo(n),au(n,n.return);break;case 12:na(u,n,a);break;case 13:na(u,n,a),a&&c&4&&Wo(u,n);break;case 22:n.memoizedState===null&&na(u,n,a),au(n,n.return);break;case 30:break;default:na(u,n,a)}t=t.sibling}}function bi(l,t){var a=null;l!==null&&l.memoizedState!==null&&l.memoizedState.cachePool!==null&&(a=l.memoizedState.cachePool.pool),l=null,t.memoizedState!==null&&t.memoizedState.cachePool!==null&&(l=t.memoizedState.cachePool.pool),l!==a&&(l!=null&&l.refCount++,a!=null&&Qe(a))}function Si(l,t){l=null,t.alternate!==null&&(l=t.alternate.memoizedState.cache),t=t.memoizedState.cache,t!==l&&(t.refCount++,l!=null&&Qe(l))}function xt(l,t,a,e){if(t.subtreeFlags&10256)for(t=t.child;t!==null;)Fo(l,t,a,e),t=t.sibling}function Fo(l,t,a,e){var u=t.flags;switch(t.tag){case 0:case 11:case 15:xt(l,t,a,e),u&2048&&tu(9,t);break;case 1:xt(l,t,a,e);break;case 3:xt(l,t,a,e),u&2048&&(l=null,t.alternate!==null&&(l=t.alternate.memoizedState.cache),t=t.memoizedState.cache,t!==l&&(t.refCount++,l!=null&&Qe(l)));break;case 12:if(u&2048){xt(l,t,a,e),l=t.stateNode;try{var n=t.memoizedProps,c=n.id,i=n.onPostCommit;typeof i=="function"&&i(c,t.alternate===null?"mount":"update",l.passiveEffectDuration,-0)}catch(f){vl(t,t.return,f)}}else xt(l,t,a,e);break;case 13:xt(l,t,a,e);break;case 23:break;case 22:n=t.stateNode,c=t.alternate,t.memoizedState!==null?n._visibility&2?xt(l,t,a,e):eu(l,t):n._visibility&2?xt(l,t,a,e):(n._visibility|=2,se(l,t,a,e,(t.subtreeFlags&10256)!==0)),u&2048&&bi(c,t);break;case 24:xt(l,t,a,e),u&2048&&Si(t.alternate,t);break;default:xt(l,t,a,e)}}function se(l,t,a,e,u){for(u=u&&(t.subtreeFlags&10256)!==0,t=t.child;t!==null;){var n=l,c=t,i=a,f=e,m=c.flags;switch(c.tag){case 0:case 11:case 15:se(n,c,i,f,u),tu(8,c);break;case 23:break;case 22:var z=c.stateNode;c.memoizedState!==null?z._visibility&2?se(n,c,i,f,u):eu(n,c):(z._visibility|=2,se(n,c,i,f,u)),u&&m&2048&&bi(c.alternate,c);break;case 24:se(n,c,i,f,u),u&&m&2048&&Si(c.alternate,c);break;default:se(n,c,i,f,u)}t=t.sibling}}function eu(l,t){if(t.subtreeFlags&10256)for(t=t.child;t!==null;){var a=l,e=t,u=e.flags;switch(e.tag){case 22:eu(a,e),u&2048&&bi(e.alternate,e);break;case 24:eu(a,e),u&2048&&Si(e.alternate,e);break;default:eu(a,e)}t=t.sibling}}var uu=8192;function oe(l){if(l.subtreeFlags&uu)for(l=l.child;l!==null;)Po(l),l=l.sibling}function Po(l){switch(l.tag){case 26:oe(l),l.flags&uu&&l.memoizedState!==null&&Rv(zt,l.memoizedState,l.memoizedProps);break;case 5:oe(l);break;case 3:case 4:var t=zt;zt=Mn(l.stateNode.containerInfo),oe(l),zt=t;break;case 22:l.memoizedState===null&&(t=l.alternate,t!==null&&t.memoizedState!==null?(t=uu,uu=16777216,oe(l),uu=t):oe(l));break;default:oe(l)}}function Io(l){var t=l.alternate;if(t!==null&&(l=t.child,l!==null)){t.child=null;do t=l.sibling,l.sibling=null,l=t;while(l!==null)}}function nu(l){var t=l.deletions;if((l.flags&16)!==0){if(t!==null)for(var a=0;a<t.length;a++){var e=t[a];jl=e,td(e,l)}Io(l)}if(l.subtreeFlags&10256)for(l=l.child;l!==null;)ld(l),l=l.sibling}function ld(l){switch(l.tag){case 0:case 11:case 15:nu(l),l.flags&2048&&ea(9,l,l.return);break;case 3:nu(l);break;case 12:nu(l);break;case 22:var t=l.stateNode;l.memoizedState!==null&&t._visibility&2&&(l.return===null||l.return.tag!==13)?(t._visibility&=-3,mn(l)):nu(l);break;default:nu(l)}}function mn(l){var t=l.deletions;if((l.flags&16)!==0){if(t!==null)for(var a=0;a<t.length;a++){var e=t[a];jl=e,td(e,l)}Io(l)}for(l=l.child;l!==null;){switch(t=l,t.tag){case 0:case 11:case 15:ea(8,t,t.return),mn(t);break;case 22:a=t.stateNode,a._visibility&2&&(a._visibility&=-3,mn(t));break;default:mn(t)}l=l.sibling}}function td(l,t){for(;jl!==null;){var a=jl;switch(a.tag){case 0:case 11:case 15:ea(8,a,t);break;case 23:case 22:if(a.memoizedState!==null&&a.memoizedState.cachePool!==null){var e=a.memoizedState.cachePool.pool;e!=null&&e.refCount++}break;case 24:Qe(a.memoizedState.cache)}if(e=a.child,e!==null)e.return=a,jl=e;else l:for(a=l;jl!==null;){e=jl;var u=e.sibling,n=e.return;if(Jo(e),e===a){jl=null;break l}if(u!==null){u.return=n,jl=u;break l}jl=n}}}var W0={getCacheForType:function(l){var t=Ql(xl),a=t.data.get(l);return a===void 0&&(a=l(),t.data.set(l,a)),a}},k0=typeof WeakMap=="function"?WeakMap:Map,nl=0,yl=null,$=null,I=0,cl=0,nt=null,ca=!1,de=!1,zi=!1,Vt=0,zl=0,ia=0,qa=0,Ti=0,mt=0,re=0,cu=null,Fl=null,Ei=!1,Ai=0,gn=1/0,bn=null,fa=null,Cl=0,sa=null,ve=null,ye=0,Oi=0,pi=null,ad=null,iu=0,xi=null;function ct(){if((nl&2)!==0&&I!==0)return I&-I;if(S.T!==null){var l=te;return l!==0?l:ji()}return gf()}function ed(){mt===0&&(mt=(I&536870912)===0||el?vf():536870912);var l=ht.current;return l!==null&&(l.flags|=32),mt}function it(l,t,a){(l===yl&&(cl===2||cl===9)||l.cancelPendingCommit!==null)&&(he(l,0),oa(l,I,mt,!1)),pe(l,a),((nl&2)===0||l!==yl)&&(l===yl&&((nl&2)===0&&(qa|=a),zl===4&&oa(l,I,mt,!1)),_t(l))}function ud(l,t,a){if((nl&6)!==0)throw Error(s(327));var e=!a&&(t&124)===0&&(t&l.expiredLanes)===0||Oe(l,t),u=e?P0(l,t):Di(l,t,!0),n=e;do{if(u===0){de&&!e&&oa(l,t,0,!1);break}else{if(a=l.current.alternate,n&&!$0(a)){u=Di(l,t,!1),n=!1;continue}if(u===2){if(n=t,l.errorRecoveryDisabledLanes&n)var c=0;else c=l.pendingLanes&-536870913,c=c!==0?c:c&536870912?536870912:0;if(c!==0){t=c;l:{var i=l;u=cu;var f=i.current.memoizedState.isDehydrated;if(f&&(he(i,c).flags|=256),c=Di(i,c,!1),c!==2){if(zi&&!f){i.errorRecoveryDisabledLanes|=n,qa|=n,u=4;break l}n=Fl,Fl=u,n!==null&&(Fl===null?Fl=n:Fl.push.apply(Fl,n))}u=c}if(n=!1,u!==2)continue}}if(u===1){he(l,0),oa(l,t,0,!0);break}l:{switch(e=l,n=u,n){case 0:case 1:throw Error(s(345));case 4:if((t&4194048)!==t)break;case 6:oa(e,t,mt,!ca);break l;case 2:Fl=null;break;case 3:case 5:break;default:throw Error(s(329))}if((t&62914560)===t&&(u=Ai+300-Et(),10<u)){if(oa(e,t,mt,!ca),Mu(e,0,!0)!==0)break l;e.timeoutHandle=jd(nd.bind(null,e,a,Fl,bn,Ei,t,mt,qa,re,ca,n,2,-0,0),u);break l}nd(e,a,Fl,bn,Ei,t,mt,qa,re,ca,n,0,-0,0)}}break}while(!0);_t(l)}function nd(l,t,a,e,u,n,c,i,f,m,z,A,g,b){if(l.timeoutHandle=-1,A=t.subtreeFlags,(A&8192||(A&16785408)===16785408)&&(yu={stylesheets:null,count:0,unsuspend:Uv},Po(t),A=jv(),A!==null)){l.cancelPendingCommit=A(rd.bind(null,l,t,n,a,e,u,c,i,f,z,1,g,b)),oa(l,n,c,!m);return}rd(l,t,n,a,e,u,c,i,f)}function $0(l){for(var t=l;;){var a=t.tag;if((a===0||a===11||a===15)&&t.flags&16384&&(a=t.updateQueue,a!==null&&(a=a.stores,a!==null)))for(var e=0;e<a.length;e++){var u=a[e],n=u.getSnapshot;u=u.value;try{if(!tt(n(),u))return!1}catch{return!1}}if(a=t.child,t.subtreeFlags&16384&&a!==null)a.return=t,t=a;else{if(t===l)break;for(;t.sibling===null;){if(t.return===null||t.return===l)return!0;t=t.return}t.sibling.return=t.return,t=t.sibling}}return!0}function oa(l,t,a,e){t&=~Ti,t&=~qa,l.suspendedLanes|=t,l.pingedLanes&=~t,e&&(l.warmLanes|=t),e=l.expirationTimes;for(var u=t;0<u;){var n=31-lt(u),c=1<<n;e[n]=-1,u&=~c}a!==0&&hf(l,a,t)}function Sn(){return(nl&6)===0?(fu(0),!1):!0}function _i(){if($!==null){if(cl===0)var l=$.return;else l=$,qt=Da=null,Lc(l),ie=null,Pe=0,l=$;for(;l!==null;)Co(l.alternate,l),l=l.return;$=null}}function he(l,t){var a=l.timeoutHandle;a!==-1&&(l.timeoutHandle=-1,yv(a)),a=l.cancelPendingCommit,a!==null&&(l.cancelPendingCommit=null,a()),_i(),yl=l,$=a=Rt(l.current,null),I=t,cl=0,nt=null,ca=!1,de=Oe(l,t),zi=!1,re=mt=Ti=qa=ia=zl=0,Fl=cu=null,Ei=!1,(t&8)!==0&&(t|=t&32);var e=l.entangledLanes;if(e!==0)for(l=l.entanglements,e&=t;0<e;){var u=31-lt(e),n=1<<u;t|=l[u],e&=~n}return Vt=t,Xu(),a}function cd(l,t){W=null,S.H=nn,t===Ve||t===ku?(t=Es(),cl=3):t===Ss?(t=Es(),cl=4):cl=t===Oo?8:t!==null&&typeof t=="object"&&typeof t.then=="function"?6:1,nt=t,$===null&&(zl=1,dn(l,dt(t,l.current)))}function id(){var l=S.H;return S.H=nn,l===null?nn:l}function fd(){var l=S.A;return S.A=W0,l}function Mi(){zl=4,ca||(I&4194048)!==I&&ht.current!==null||(de=!0),(ia&134217727)===0&&(qa&134217727)===0||yl===null||oa(yl,I,mt,!1)}function Di(l,t,a){var e=nl;nl|=2;var u=id(),n=fd();(yl!==l||I!==t)&&(bn=null,he(l,t)),t=!1;var c=zl;l:do try{if(cl!==0&&$!==null){var i=$,f=nt;switch(cl){case 8:_i(),c=6;break l;case 3:case 2:case 9:case 6:ht.current===null&&(t=!0);var m=cl;if(cl=0,nt=null,me(l,i,f,m),a&&de){c=0;break l}break;default:m=cl,cl=0,nt=null,me(l,i,f,m)}}F0(),c=zl;break}catch(z){cd(l,z)}while(!0);return t&&l.shellSuspendCounter++,qt=Da=null,nl=e,S.H=u,S.A=n,$===null&&(yl=null,I=0,Xu()),c}function F0(){for(;$!==null;)sd($)}function P0(l,t){var a=nl;nl|=2;var e=id(),u=fd();yl!==l||I!==t?(bn=null,gn=Et()+500,he(l,t)):de=Oe(l,t);l:do try{if(cl!==0&&$!==null){t=$;var n=nt;t:switch(cl){case 1:cl=0,nt=null,me(l,t,n,1);break;case 2:case 9:if(zs(n)){cl=0,nt=null,od(t);break}t=function(){cl!==2&&cl!==9||yl!==l||(cl=7),_t(l)},n.then(t,t);break l;case 3:cl=7;break l;case 4:cl=5;break l;case 7:zs(n)?(cl=0,nt=null,od(t)):(cl=0,nt=null,me(l,t,n,7));break;case 5:var c=null;switch($.tag){case 26:c=$.memoizedState;case 5:case 27:var i=$;if(!c||Kd(c)){cl=0,nt=null;var f=i.sibling;if(f!==null)$=f;else{var m=i.return;m!==null?($=m,zn(m)):$=null}break t}}cl=0,nt=null,me(l,t,n,5);break;case 6:cl=0,nt=null,me(l,t,n,6);break;case 8:_i(),zl=6;break l;default:throw Error(s(462))}}I0();break}catch(z){cd(l,z)}while(!0);return qt=Da=null,S.H=e,S.A=u,nl=a,$!==null?0:(yl=null,I=0,Xu(),zl)}function I0(){for(;$!==null&&!Tr();)sd($)}function sd(l){var t=Yo(l.alternate,l,Vt);l.memoizedProps=l.pendingProps,t===null?zn(l):$=t}function od(l){var t=l,a=t.alternate;switch(t.tag){case 15:case 0:t=No(a,t,t.pendingProps,t.type,void 0,I);break;case 11:t=No(a,t,t.pendingProps,t.type.render,t.ref,I);break;case 5:Lc(t);default:Co(a,t),t=$=os(t,Vt),t=Yo(a,t,Vt)}l.memoizedProps=l.pendingProps,t===null?zn(l):$=t}function me(l,t,a,e){qt=Da=null,Lc(t),ie=null,Pe=0;var u=t.return;try{if(Z0(l,u,t,a,I)){zl=1,dn(l,dt(a,l.current)),$=null;return}}catch(n){if(u!==null)throw $=u,n;zl=1,dn(l,dt(a,l.current)),$=null;return}t.flags&32768?(el||e===1?l=!0:de||(I&536870912)!==0?l=!1:(ca=l=!0,(e===2||e===9||e===3||e===6)&&(e=ht.current,e!==null&&e.tag===13&&(e.flags|=16384))),dd(t,l)):zn(t)}function zn(l){var t=l;do{if((t.flags&32768)!==0){dd(t,ca);return}l=t.return;var a=L0(t.alternate,t,Vt);if(a!==null){$=a;return}if(t=t.sibling,t!==null){$=t;return}$=t=l}while(t!==null);zl===0&&(zl=5)}function dd(l,t){do{var a=K0(l.alternate,l);if(a!==null){a.flags&=32767,$=a;return}if(a=l.return,a!==null&&(a.flags|=32768,a.subtreeFlags=0,a.deletions=null),!t&&(l=l.sibling,l!==null)){$=l;return}$=l=a}while(l!==null);zl=6,$=null}function rd(l,t,a,e,u,n,c,i,f){l.cancelPendingCommit=null;do Tn();while(Cl!==0);if((nl&6)!==0)throw Error(s(327));if(t!==null){if(t===l.current)throw Error(s(177));if(n=t.lanes|t.childLanes,n|=Sc,Ur(l,a,n,c,i,f),l===yl&&($=yl=null,I=0),ve=t,sa=l,ye=a,Oi=n,pi=u,ad=e,(t.subtreeFlags&10256)!==0||(t.flags&10256)!==0?(l.callbackNode=null,l.callbackPriority=0,ev(pu,function(){return gd(),null})):(l.callbackNode=null,l.callbackPriority=0),e=(t.flags&13878)!==0,(t.subtreeFlags&13878)!==0||e){e=S.T,S.T=null,u=N.p,N.p=2,c=nl,nl|=4;try{J0(l,t,a)}finally{nl=c,N.p=u,S.T=e}}Cl=1,vd(),yd(),hd()}}function vd(){if(Cl===1){Cl=0;var l=sa,t=ve,a=(t.flags&13878)!==0;if((t.subtreeFlags&13878)!==0||a){a=S.T,S.T=null;var e=N.p;N.p=2;var u=nl;nl|=4;try{ko(t,l);var n=Qi,c=ls(l.containerInfo),i=n.focusedElem,f=n.selectionRange;if(c!==i&&i&&i.ownerDocument&&If(i.ownerDocument.documentElement,i)){if(f!==null&&yc(i)){var m=f.start,z=f.end;if(z===void 0&&(z=m),"selectionStart"in i)i.selectionStart=m,i.selectionEnd=Math.min(z,i.value.length);else{var A=i.ownerDocument||document,g=A&&A.defaultView||window;if(g.getSelection){var b=g.getSelection(),L=i.textContent.length,X=Math.min(f.start,L),ol=f.end===void 0?X:Math.min(f.end,L);!b.extend&&X>ol&&(c=ol,ol=X,X=c);var r=Pf(i,X),d=Pf(i,ol);if(r&&d&&(b.rangeCount!==1||b.anchorNode!==r.node||b.anchorOffset!==r.offset||b.focusNode!==d.node||b.focusOffset!==d.offset)){var h=A.createRange();h.setStart(r.node,r.offset),b.removeAllRanges(),X>ol?(b.addRange(h),b.extend(d.node,d.offset)):(h.setEnd(d.node,d.offset),b.addRange(h))}}}}for(A=[],b=i;b=b.parentNode;)b.nodeType===1&&A.push({element:b,left:b.scrollLeft,top:b.scrollTop});for(typeof i.focus=="function"&&i.focus(),i=0;i<A.length;i++){var T=A[i];T.element.scrollLeft=T.left,T.element.scrollTop=T.top}}jn=!!Xi,Qi=Xi=null}finally{nl=u,N.p=e,S.T=a}}l.current=t,Cl=2}}function yd(){if(Cl===2){Cl=0;var l=sa,t=ve,a=(t.flags&8772)!==0;if((t.subtreeFlags&8772)!==0||a){a=S.T,S.T=null;var e=N.p;N.p=2;var u=nl;nl|=4;try{Ko(l,t.alternate,t)}finally{nl=u,N.p=e,S.T=a}}Cl=3}}function hd(){if(Cl===4||Cl===3){Cl=0,Er();var l=sa,t=ve,a=ye,e=ad;(t.subtreeFlags&10256)!==0||(t.flags&10256)!==0?Cl=5:(Cl=0,ve=sa=null,md(l,l.pendingLanes));var u=l.pendingLanes;if(u===0&&(fa=null),wn(a),t=t.stateNode,Il&&typeof Il.onCommitFiberRoot=="function")try{Il.onCommitFiberRoot(Ae,t,void 0,(t.current.flags&128)===128)}catch{}if(e!==null){t=S.T,u=N.p,N.p=2,S.T=null;try{for(var n=l.onRecoverableError,c=0;c<e.length;c++){var i=e[c];n(i.value,{componentStack:i.stack})}}finally{S.T=t,N.p=u}}(ye&3)!==0&&Tn(),_t(l),u=l.pendingLanes,(a&4194090)!==0&&(u&42)!==0?l===xi?iu++:(iu=0,xi=l):iu=0,fu(0)}}function md(l,t){(l.pooledCacheLanes&=t)===0&&(t=l.pooledCache,t!=null&&(l.pooledCache=null,Qe(t)))}function Tn(l){return vd(),yd(),hd(),gd()}function gd(){if(Cl!==5)return!1;var l=sa,t=Oi;Oi=0;var a=wn(ye),e=S.T,u=N.p;try{N.p=32>a?32:a,S.T=null,a=pi,pi=null;var n=sa,c=ye;if(Cl=0,ve=sa=null,ye=0,(nl&6)!==0)throw Error(s(331));var i=nl;if(nl|=4,ld(n.current),Fo(n,n.current,c,a),nl=i,fu(0,!1),Il&&typeof Il.onPostCommitFiberRoot=="function")try{Il.onPostCommitFiberRoot(Ae,n)}catch{}return!0}finally{N.p=u,S.T=e,md(l,t)}}function bd(l,t,a){t=dt(a,t),t=ui(l.stateNode,t,2),l=It(l,t,2),l!==null&&(pe(l,2),_t(l))}function vl(l,t,a){if(l.tag===3)bd(l,l,a);else for(;t!==null;){if(t.tag===3){bd(t,l,a);break}else if(t.tag===1){var e=t.stateNode;if(typeof t.type.getDerivedStateFromError=="function"||typeof e.componentDidCatch=="function"&&(fa===null||!fa.has(e))){l=dt(a,l),a=Eo(2),e=It(t,a,2),e!==null&&(Ao(a,e,t,l),pe(e,2),_t(e));break}}t=t.return}}function Ni(l,t,a){var e=l.pingCache;if(e===null){e=l.pingCache=new k0;var u=new Set;e.set(t,u)}else u=e.get(t),u===void 0&&(u=new Set,e.set(t,u));u.has(a)||(zi=!0,u.add(a),l=lv.bind(null,l,t,a),t.then(l,l))}function lv(l,t,a){var e=l.pingCache;e!==null&&e.delete(t),l.pingedLanes|=l.suspendedLanes&a,l.warmLanes&=~a,yl===l&&(I&a)===a&&(zl===4||zl===3&&(I&62914560)===I&&300>Et()-Ai?(nl&2)===0&&he(l,0):Ti|=a,re===I&&(re=0)),_t(l)}function Sd(l,t){t===0&&(t=yf()),l=Fa(l,t),l!==null&&(pe(l,t),_t(l))}function tv(l){var t=l.memoizedState,a=0;t!==null&&(a=t.retryLane),Sd(l,a)}function av(l,t){var a=0;switch(l.tag){case 13:var e=l.stateNode,u=l.memoizedState;u!==null&&(a=u.retryLane);break;case 19:e=l.stateNode;break;case 22:e=l.stateNode._retryCache;break;default:throw Error(s(314))}e!==null&&e.delete(t),Sd(l,a)}function ev(l,t){return Vn(l,t)}var En=null,ge=null,Ui=!1,An=!1,Ri=!1,Ya=0;function _t(l){l!==ge&&l.next===null&&(ge===null?En=ge=l:ge=ge.next=l),An=!0,Ui||(Ui=!0,nv())}function fu(l,t){if(!Ri&&An){Ri=!0;do for(var a=!1,e=En;e!==null;){if(l!==0){var u=e.pendingLanes;if(u===0)var n=0;else{var c=e.suspendedLanes,i=e.pingedLanes;n=(1<<31-lt(42|l)+1)-1,n&=u&~(c&~i),n=n&201326741?n&201326741|1:n?n|2:0}n!==0&&(a=!0,Ad(e,n))}else n=I,n=Mu(e,e===yl?n:0,e.cancelPendingCommit!==null||e.timeoutHandle!==-1),(n&3)===0||Oe(e,n)||(a=!0,Ad(e,n));e=e.next}while(a);Ri=!1}}function uv(){zd()}function zd(){An=Ui=!1;var l=0;Ya!==0&&(vv()&&(l=Ya),Ya=0);for(var t=Et(),a=null,e=En;e!==null;){var u=e.next,n=Td(e,t);n===0?(e.next=null,a===null?En=u:a.next=u,u===null&&(ge=a)):(a=e,(l!==0||(n&3)!==0)&&(An=!0)),e=u}fu(l)}function Td(l,t){for(var a=l.suspendedLanes,e=l.pingedLanes,u=l.expirationTimes,n=l.pendingLanes&-62914561;0<n;){var c=31-lt(n),i=1<<c,f=u[c];f===-1?((i&a)===0||(i&e)!==0)&&(u[c]=Nr(i,t)):f<=t&&(l.expiredLanes|=i),n&=~i}if(t=yl,a=I,a=Mu(l,l===t?a:0,l.cancelPendingCommit!==null||l.timeoutHandle!==-1),e=l.callbackNode,a===0||l===t&&(cl===2||cl===9)||l.cancelPendingCommit!==null)return e!==null&&e!==null&&Ln(e),l.callbackNode=null,l.callbackPriority=0;if((a&3)===0||Oe(l,a)){if(t=a&-a,t===l.callbackPriority)return t;switch(e!==null&&Ln(e),wn(a)){case 2:case 8:a=df;break;case 32:a=pu;break;case 268435456:a=rf;break;default:a=pu}return e=Ed.bind(null,l),a=Vn(a,e),l.callbackPriority=t,l.callbackNode=a,t}return e!==null&&e!==null&&Ln(e),l.callbackPriority=2,l.callbackNode=null,2}function Ed(l,t){if(Cl!==0&&Cl!==5)return l.callbackNode=null,l.callbackPriority=0,null;var a=l.callbackNode;if(Tn()&&l.callbackNode!==a)return null;var e=I;return e=Mu(l,l===yl?e:0,l.cancelPendingCommit!==null||l.timeoutHandle!==-1),e===0?null:(ud(l,e,t),Td(l,Et()),l.callbackNode!=null&&l.callbackNode===a?Ed.bind(null,l):null)}function Ad(l,t){if(Tn())return null;ud(l,t,!0)}function nv(){hv(function(){(nl&6)!==0?Vn(of,uv):zd()})}function ji(){return Ya===0&&(Ya=vf()),Ya}function Od(l){return l==null||typeof l=="symbol"||typeof l=="boolean"?null:typeof l=="function"?l:ju(""+l)}function pd(l,t){var a=t.ownerDocument.createElement("input");return a.name=t.name,a.value=t.value,l.id&&a.setAttribute("form",l.id),t.parentNode.insertBefore(a,t),l=new FormData(l),a.parentNode.removeChild(a),l}function cv(l,t,a,e,u){if(t==="submit"&&a&&a.stateNode===u){var n=Od((u[wl]||null).action),c=e.submitter;c&&(t=(t=c[wl]||null)?Od(t.formAction):c.getAttribute("formAction"),t!==null&&(n=t,c=null));var i=new Bu("action","action",null,e,u);l.push({event:i,listeners:[{instance:null,listener:function(){if(e.defaultPrevented){if(Ya!==0){var f=c?pd(u,c):new FormData(u);Ic(a,{pending:!0,data:f,method:u.method,action:n},null,f)}}else typeof n=="function"&&(i.preventDefault(),f=c?pd(u,c):new FormData(u),Ic(a,{pending:!0,data:f,method:u.method,action:n},n,f))},currentTarget:u}]})}}for(var Hi=0;Hi<bc.length;Hi++){var qi=bc[Hi],iv=qi.toLowerCase(),fv=qi[0].toUpperCase()+qi.slice(1);St(iv,"on"+fv)}St(es,"onAnimationEnd"),St(us,"onAnimationIteration"),St(ns,"onAnimationStart"),St("dblclick","onDoubleClick"),St("focusin","onFocus"),St("focusout","onBlur"),St(p0,"onTransitionRun"),St(x0,"onTransitionStart"),St(_0,"onTransitionCancel"),St(cs,"onTransitionEnd"),Qa("onMouseEnter",["mouseout","mouseover"]),Qa("onMouseLeave",["mouseout","mouseover"]),Qa("onPointerEnter",["pointerout","pointerover"]),Qa("onPointerLeave",["pointerout","pointerover"]),za("onChange","change click focusin focusout input keydown keyup selectionchange".split(" ")),za("onSelect","focusout contextmenu dragend focusin keydown keyup mousedown mouseup selectionchange".split(" ")),za("onBeforeInput",["compositionend","keypress","textInput","paste"]),za("onCompositionEnd","compositionend focusout keydown keypress keyup mousedown".split(" ")),za("onCompositionStart","compositionstart focusout keydown keypress keyup mousedown".split(" ")),za("onCompositionUpdate","compositionupdate focusout keydown keypress keyup mousedown".split(" "));var su="abort canplay canplaythrough durationchange emptied encrypted ended error loadeddata loadedmetadata loadstart pause play playing progress ratechange resize seeked seeking stalled suspend timeupdate volumechange waiting".split(" "),sv=new Set("beforetoggle cancel close invalid load scroll scrollend toggle".split(" ").concat(su));function xd(l,t){t=(t&4)!==0;for(var a=0;a<l.length;a++){var e=l[a],u=e.event;e=e.listeners;l:{var n=void 0;if(t)for(var c=e.length-1;0<=c;c--){var i=e[c],f=i.instance,m=i.currentTarget;if(i=i.listener,f!==n&&u.isPropagationStopped())break l;n=i,u.currentTarget=m;try{n(u)}catch(z){on(z)}u.currentTarget=null,n=f}else for(c=0;c<e.length;c++){if(i=e[c],f=i.instance,m=i.currentTarget,i=i.listener,f!==n&&u.isPropagationStopped())break l;n=i,u.currentTarget=m;try{n(u)}catch(z){on(z)}u.currentTarget=null,n=f}}}}function F(l,t){var a=t[Wn];a===void 0&&(a=t[Wn]=new Set);var e=l+"__bubble";a.has(e)||(_d(t,l,2,!1),a.add(e))}function Yi(l,t,a){var e=0;t&&(e|=4),_d(a,l,e,t)}var On="_reactListening"+Math.random().toString(36).slice(2);function Bi(l){if(!l[On]){l[On]=!0,Sf.forEach(function(a){a!=="selectionchange"&&(sv.has(a)||Yi(a,!1,l),Yi(a,!0,l))});var t=l.nodeType===9?l:l.ownerDocument;t===null||t[On]||(t[On]=!0,Yi("selectionchange",!1,t))}}function _d(l,t,a,e){switch(Fd(t)){case 2:var u=Yv;break;case 8:u=Bv;break;default:u=Fi}a=u.bind(null,t,a,l),u=void 0,!nc||t!=="touchstart"&&t!=="touchmove"&&t!=="wheel"||(u=!0),e?u!==void 0?l.addEventListener(t,a,{capture:!0,passive:u}):l.addEventListener(t,a,!0):u!==void 0?l.addEventListener(t,a,{passive:u}):l.addEventListener(t,a,!1)}function Ci(l,t,a,e,u){var n=e;if((t&1)===0&&(t&2)===0&&e!==null)l:for(;;){if(e===null)return;var c=e.tag;if(c===3||c===4){var i=e.stateNode.containerInfo;if(i===u)break;if(c===4)for(c=e.return;c!==null;){var f=c.tag;if((f===3||f===4)&&c.stateNode.containerInfo===u)return;c=c.return}for(;i!==null;){if(c=Ca(i),c===null)return;if(f=c.tag,f===5||f===6||f===26||f===27){e=n=c;continue l}i=i.parentNode}}e=e.return}jf(function(){var m=n,z=ec(a),A=[];l:{var g=is.get(l);if(g!==void 0){var b=Bu,L=l;switch(l){case"keypress":if(qu(a)===0)break l;case"keydown":case"keyup":b=e0;break;case"focusin":L="focus",b=sc;break;case"focusout":L="blur",b=sc;break;case"beforeblur":case"afterblur":b=sc;break;case"click":if(a.button===2)break l;case"auxclick":case"dblclick":case"mousedown":case"mousemove":case"mouseup":case"mouseout":case"mouseover":case"contextmenu":b=Yf;break;case"drag":case"dragend":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"dragstart":case"drop":b=Kr;break;case"touchcancel":case"touchend":case"touchmove":case"touchstart":b=c0;break;case es:case us:case ns:b=Wr;break;case cs:b=f0;break;case"scroll":case"scrollend":b=Vr;break;case"wheel":b=o0;break;case"copy":case"cut":case"paste":b=$r;break;case"gotpointercapture":case"lostpointercapture":case"pointercancel":case"pointerdown":case"pointermove":case"pointerout":case"pointerover":case"pointerup":b=Cf;break;case"toggle":case"beforetoggle":b=r0}var X=(t&4)!==0,ol=!X&&(l==="scroll"||l==="scrollend"),r=X?g!==null?g+"Capture":null:g;X=[];for(var d=m,h;d!==null;){var T=d;if(h=T.stateNode,T=T.tag,T!==5&&T!==26&&T!==27||h===null||r===null||(T=Me(d,r),T!=null&&X.push(ou(d,T,h))),ol)break;d=d.return}0<X.length&&(g=new b(g,L,null,a,z),A.push({event:g,listeners:X}))}}if((t&7)===0){l:{if(g=l==="mouseover"||l==="pointerover",b=l==="mouseout"||l==="pointerout",g&&a!==ac&&(L=a.relatedTarget||a.fromElement)&&(Ca(L)||L[Ba]))break l;if((b||g)&&(g=z.window===z?z:(g=z.ownerDocument)?g.defaultView||g.parentWindow:window,b?(L=a.relatedTarget||a.toElement,b=m,L=L?Ca(L):null,L!==null&&(ol=q(L),X=L.tag,L!==ol||X!==5&&X!==27&&X!==6)&&(L=null)):(b=null,L=m),b!==L)){if(X=Yf,T="onMouseLeave",r="onMouseEnter",d="mouse",(l==="pointerout"||l==="pointerover")&&(X=Cf,T="onPointerLeave",r="onPointerEnter",d="pointer"),ol=b==null?g:_e(b),h=L==null?g:_e(L),g=new X(T,d+"leave",b,a,z),g.target=ol,g.relatedTarget=h,T=null,Ca(z)===m&&(X=new X(r,d+"enter",L,a,z),X.target=h,X.relatedTarget=ol,T=X),ol=T,b&&L)t:{for(X=b,r=L,d=0,h=X;h;h=be(h))d++;for(h=0,T=r;T;T=be(T))h++;for(;0<d-h;)X=be(X),d--;for(;0<h-d;)r=be(r),h--;for(;d--;){if(X===r||r!==null&&X===r.alternate)break t;X=be(X),r=be(r)}X=null}else X=null;b!==null&&Md(A,g,b,X,!1),L!==null&&ol!==null&&Md(A,ol,L,X,!0)}}l:{if(g=m?_e(m):window,b=g.nodeName&&g.nodeName.toLowerCase(),b==="select"||b==="input"&&g.type==="file")var Y=Jf;else if(Lf(g))if(wf)Y=E0;else{Y=z0;var k=S0}else b=g.nodeName,!b||b.toLowerCase()!=="input"||g.type!=="checkbox"&&g.type!=="radio"?m&&tc(m.elementType)&&(Y=Jf):Y=T0;if(Y&&(Y=Y(l,m))){Kf(A,Y,a,z);break l}k&&k(l,g,m),l==="focusout"&&m&&g.type==="number"&&m.memoizedProps.value!=null&&lc(g,"number",g.value)}switch(k=m?_e(m):window,l){case"focusin":(Lf(k)||k.contentEditable==="true")&&(Wa=k,hc=m,Ye=null);break;case"focusout":Ye=hc=Wa=null;break;case"mousedown":mc=!0;break;case"contextmenu":case"mouseup":case"dragend":mc=!1,ts(A,a,z);break;case"selectionchange":if(O0)break;case"keydown":case"keyup":ts(A,a,z)}var C;if(dc)l:{switch(l){case"compositionstart":var Z="onCompositionStart";break l;case"compositionend":Z="onCompositionEnd";break l;case"compositionupdate":Z="onCompositionUpdate";break l}Z=void 0}else wa?Zf(l,a)&&(Z="onCompositionEnd"):l==="keydown"&&a.keyCode===229&&(Z="onCompositionStart");Z&&(Gf&&a.locale!=="ko"&&(wa||Z!=="onCompositionStart"?Z==="onCompositionEnd"&&wa&&(C=Hf()):(kt=z,cc="value"in kt?kt.value:kt.textContent,wa=!0)),k=pn(m,Z),0<k.length&&(Z=new Bf(Z,l,null,a,z),A.push({event:Z,listeners:k}),C?Z.data=C:(C=Vf(a),C!==null&&(Z.data=C)))),(C=y0?h0(l,a):m0(l,a))&&(Z=pn(m,"onBeforeInput"),0<Z.length&&(k=new Bf("onBeforeInput","beforeinput",null,a,z),A.push({event:k,listeners:Z}),k.data=C)),cv(A,l,m,a,z)}xd(A,t)})}function ou(l,t,a){return{instance:l,listener:t,currentTarget:a}}function pn(l,t){for(var a=t+"Capture",e=[];l!==null;){var u=l,n=u.stateNode;if(u=u.tag,u!==5&&u!==26&&u!==27||n===null||(u=Me(l,a),u!=null&&e.unshift(ou(l,u,n)),u=Me(l,t),u!=null&&e.push(ou(l,u,n))),l.tag===3)return e;l=l.return}return[]}function be(l){if(l===null)return null;do l=l.return;while(l&&l.tag!==5&&l.tag!==27);return l||null}function Md(l,t,a,e,u){for(var n=t._reactName,c=[];a!==null&&a!==e;){var i=a,f=i.alternate,m=i.stateNode;if(i=i.tag,f!==null&&f===e)break;i!==5&&i!==26&&i!==27||m===null||(f=m,u?(m=Me(a,n),m!=null&&c.unshift(ou(a,m,f))):u||(m=Me(a,n),m!=null&&c.push(ou(a,m,f)))),a=a.return}c.length!==0&&l.push({event:t,listeners:c})}var ov=/\r\n?/g,dv=/\u0000|\uFFFD/g;function Dd(l){return(typeof l=="string"?l:""+l).replace(ov,`
-`).replace(dv,"")}function Nd(l,t){return t=Dd(t),Dd(l)===t}function xn(){}function sl(l,t,a,e,u,n){switch(a){case"children":typeof e=="string"?t==="body"||t==="textarea"&&e===""||La(l,e):(typeof e=="number"||typeof e=="bigint")&&t!=="body"&&La(l,""+e);break;case"className":Nu(l,"class",e);break;case"tabIndex":Nu(l,"tabindex",e);break;case"dir":case"role":case"viewBox":case"width":case"height":Nu(l,a,e);break;case"style":Uf(l,e,n);break;case"data":if(t!=="object"){Nu(l,"data",e);break}case"src":case"href":if(e===""&&(t!=="a"||a!=="href")){l.removeAttribute(a);break}if(e==null||typeof e=="function"||typeof e=="symbol"||typeof e=="boolean"){l.removeAttribute(a);break}e=ju(""+e),l.setAttribute(a,e);break;case"action":case"formAction":if(typeof e=="function"){l.setAttribute(a,"javascript:throw new Error('A React form was unexpectedly submitted. If you called form.submit() manually, consider using form.requestSubmit() instead. If you\\'re trying to use event.stopPropagation() in a submit event handler, consider also calling event.preventDefault().')");break}else typeof n=="function"&&(a==="formAction"?(t!=="input"&&sl(l,t,"name",u.name,u,null),sl(l,t,"formEncType",u.formEncType,u,null),sl(l,t,"formMethod",u.formMethod,u,null),sl(l,t,"formTarget",u.formTarget,u,null)):(sl(l,t,"encType",u.encType,u,null),sl(l,t,"method",u.method,u,null),sl(l,t,"target",u.target,u,null)));if(e==null||typeof e=="symbol"||typeof e=="boolean"){l.removeAttribute(a);break}e=ju(""+e),l.setAttribute(a,e);break;case"onClick":e!=null&&(l.onclick=xn);break;case"onScroll":e!=null&&F("scroll",l);break;case"onScrollEnd":e!=null&&F("scrollend",l);break;case"dangerouslySetInnerHTML":if(e!=null){if(typeof e!="object"||!("__html"in e))throw Error(s(61));if(a=e.__html,a!=null){if(u.children!=null)throw Error(s(60));l.innerHTML=a}}break;case"multiple":l.multiple=e&&typeof e!="function"&&typeof e!="symbol";break;case"muted":l.muted=e&&typeof e!="function"&&typeof e!="symbol";break;case"suppressContentEditableWarning":case"suppressHydrationWarning":case"defaultValue":case"defaultChecked":case"innerHTML":case"ref":break;case"autoFocus":break;case"xlinkHref":if(e==null||typeof e=="function"||typeof e=="boolean"||typeof e=="symbol"){l.removeAttribute("xlink:href");break}a=ju(""+e),l.setAttributeNS("http://www.w3.org/1999/xlink","xlink:href",a);break;case"contentEditable":case"spellCheck":case"draggable":case"value":case"autoReverse":case"externalResourcesRequired":case"focusable":case"preserveAlpha":e!=null&&typeof e!="function"&&typeof e!="symbol"?l.setAttribute(a,""+e):l.removeAttribute(a);break;case"inert":case"allowFullScreen":case"async":case"autoPlay":case"controls":case"default":case"defer":case"disabled":case"disablePictureInPicture":case"disableRemotePlayback":case"formNoValidate":case"hidden":case"loop":case"noModule":case"noValidate":case"open":case"playsInline":case"readOnly":case"required":case"reversed":case"scoped":case"seamless":case"itemScope":e&&typeof e!="function"&&typeof e!="symbol"?l.setAttribute(a,""):l.removeAttribute(a);break;case"capture":case"download":e===!0?l.setAttribute(a,""):e!==!1&&e!=null&&typeof e!="function"&&typeof e!="symbol"?l.setAttribute(a,e):l.removeAttribute(a);break;case"cols":case"rows":case"size":case"span":e!=null&&typeof e!="function"&&typeof e!="symbol"&&!isNaN(e)&&1<=e?l.setAttribute(a,e):l.removeAttribute(a);break;case"rowSpan":case"start":e==null||typeof e=="function"||typeof e=="symbol"||isNaN(e)?l.removeAttribute(a):l.setAttribute(a,e);break;case"popover":F("beforetoggle",l),F("toggle",l),Du(l,"popover",e);break;case"xlinkActuate":Nt(l,"http://www.w3.org/1999/xlink","xlink:actuate",e);break;case"xlinkArcrole":Nt(l,"http://www.w3.org/1999/xlink","xlink:arcrole",e);break;case"xlinkRole":Nt(l,"http://www.w3.org/1999/xlink","xlink:role",e);break;case"xlinkShow":Nt(l,"http://www.w3.org/1999/xlink","xlink:show",e);break;case"xlinkTitle":Nt(l,"http://www.w3.org/1999/xlink","xlink:title",e);break;case"xlinkType":Nt(l,"http://www.w3.org/1999/xlink","xlink:type",e);break;case"xmlBase":Nt(l,"http://www.w3.org/XML/1998/namespace","xml:base",e);break;case"xmlLang":Nt(l,"http://www.w3.org/XML/1998/namespace","xml:lang",e);break;case"xmlSpace":Nt(l,"http://www.w3.org/XML/1998/namespace","xml:space",e);break;case"is":Du(l,"is",e);break;case"innerText":case"textContent":break;default:(!(2<a.length)||a[0]!=="o"&&a[0]!=="O"||a[1]!=="n"&&a[1]!=="N")&&(a=Qr.get(a)||a,Du(l,a,e))}}function Gi(l,t,a,e,u,n){switch(a){case"style":Uf(l,e,n);break;case"dangerouslySetInnerHTML":if(e!=null){if(typeof e!="object"||!("__html"in e))throw Error(s(61));if(a=e.__html,a!=null){if(u.children!=null)throw Error(s(60));l.innerHTML=a}}break;case"children":typeof e=="string"?La(l,e):(typeof e=="number"||typeof e=="bigint")&&La(l,""+e);break;case"onScroll":e!=null&&F("scroll",l);break;case"onScrollEnd":e!=null&&F("scrollend",l);break;case"onClick":e!=null&&(l.onclick=xn);break;case"suppressContentEditableWarning":case"suppressHydrationWarning":case"innerHTML":case"ref":break;case"innerText":case"textContent":break;default:if(!zf.hasOwnProperty(a))l:{if(a[0]==="o"&&a[1]==="n"&&(u=a.endsWith("Capture"),t=a.slice(2,u?a.length-7:void 0),n=l[wl]||null,n=n!=null?n[a]:null,typeof n=="function"&&l.removeEventListener(t,n,u),typeof e=="function")){typeof n!="function"&&n!==null&&(a in l?l[a]=null:l.hasAttribute(a)&&l.removeAttribute(a)),l.addEventListener(t,e,u);break l}a in l?l[a]=e:e===!0?l.setAttribute(a,""):Du(l,a,e)}}}function Gl(l,t,a){switch(t){case"div":case"span":case"svg":case"path":case"a":case"g":case"p":case"li":break;case"img":F("error",l),F("load",l);var e=!1,u=!1,n;for(n in a)if(a.hasOwnProperty(n)){var c=a[n];if(c!=null)switch(n){case"src":e=!0;break;case"srcSet":u=!0;break;case"children":case"dangerouslySetInnerHTML":throw Error(s(137,t));default:sl(l,t,n,c,a,null)}}u&&sl(l,t,"srcSet",a.srcSet,a,null),e&&sl(l,t,"src",a.src,a,null);return;case"input":F("invalid",l);var i=n=c=u=null,f=null,m=null;for(e in a)if(a.hasOwnProperty(e)){var z=a[e];if(z!=null)switch(e){case"name":u=z;break;case"type":c=z;break;case"checked":f=z;break;case"defaultChecked":m=z;break;case"value":n=z;break;case"defaultValue":i=z;break;case"children":case"dangerouslySetInnerHTML":if(z!=null)throw Error(s(137,t));break;default:sl(l,t,e,z,a,null)}}_f(l,n,i,f,m,c,u,!1),Uu(l);return;case"select":F("invalid",l),e=c=n=null;for(u in a)if(a.hasOwnProperty(u)&&(i=a[u],i!=null))switch(u){case"value":n=i;break;case"defaultValue":c=i;break;case"multiple":e=i;default:sl(l,t,u,i,a,null)}t=n,a=c,l.multiple=!!e,t!=null?Va(l,!!e,t,!1):a!=null&&Va(l,!!e,a,!0);return;case"textarea":F("invalid",l),n=u=e=null;for(c in a)if(a.hasOwnProperty(c)&&(i=a[c],i!=null))switch(c){case"value":e=i;break;case"defaultValue":u=i;break;case"children":n=i;break;case"dangerouslySetInnerHTML":if(i!=null)throw Error(s(91));break;default:sl(l,t,c,i,a,null)}Df(l,e,u,n),Uu(l);return;case"option":for(f in a)if(a.hasOwnProperty(f)&&(e=a[f],e!=null))switch(f){case"selected":l.selected=e&&typeof e!="function"&&typeof e!="symbol";break;default:sl(l,t,f,e,a,null)}return;case"dialog":F("beforetoggle",l),F("toggle",l),F("cancel",l),F("close",l);break;case"iframe":case"object":F("load",l);break;case"video":case"audio":for(e=0;e<su.length;e++)F(su[e],l);break;case"image":F("error",l),F("load",l);break;case"details":F("toggle",l);break;case"embed":case"source":case"link":F("error",l),F("load",l);case"area":case"base":case"br":case"col":case"hr":case"keygen":case"meta":case"param":case"track":case"wbr":case"menuitem":for(m in a)if(a.hasOwnProperty(m)&&(e=a[m],e!=null))switch(m){case"children":case"dangerouslySetInnerHTML":throw Error(s(137,t));default:sl(l,t,m,e,a,null)}return;default:if(tc(t)){for(z in a)a.hasOwnProperty(z)&&(e=a[z],e!==void 0&&Gi(l,t,z,e,a,void 0));return}}for(i in a)a.hasOwnProperty(i)&&(e=a[i],e!=null&&sl(l,t,i,e,a,null))}function rv(l,t,a,e){switch(t){case"div":case"span":case"svg":case"path":case"a":case"g":case"p":case"li":break;case"input":var u=null,n=null,c=null,i=null,f=null,m=null,z=null;for(b in a){var A=a[b];if(a.hasOwnProperty(b)&&A!=null)switch(b){case"checked":break;case"value":break;case"defaultValue":f=A;default:e.hasOwnProperty(b)||sl(l,t,b,null,e,A)}}for(var g in e){var b=e[g];if(A=a[g],e.hasOwnProperty(g)&&(b!=null||A!=null))switch(g){case"type":n=b;break;case"name":u=b;break;case"checked":m=b;break;case"defaultChecked":z=b;break;case"value":c=b;break;case"defaultValue":i=b;break;case"children":case"dangerouslySetInnerHTML":if(b!=null)throw Error(s(137,t));break;default:b!==A&&sl(l,t,g,b,e,A)}}In(l,c,i,f,m,z,n,u);return;case"select":b=c=i=g=null;for(n in a)if(f=a[n],a.hasOwnProperty(n)&&f!=null)switch(n){case"value":break;case"multiple":b=f;default:e.hasOwnProperty(n)||sl(l,t,n,null,e,f)}for(u in e)if(n=e[u],f=a[u],e.hasOwnProperty(u)&&(n!=null||f!=null))switch(u){case"value":g=n;break;case"defaultValue":i=n;break;case"multiple":c=n;default:n!==f&&sl(l,t,u,n,e,f)}t=i,a=c,e=b,g!=null?Va(l,!!a,g,!1):!!e!=!!a&&(t!=null?Va(l,!!a,t,!0):Va(l,!!a,a?[]:"",!1));return;case"textarea":b=g=null;for(i in a)if(u=a[i],a.hasOwnProperty(i)&&u!=null&&!e.hasOwnProperty(i))switch(i){case"value":break;case"children":break;default:sl(l,t,i,null,e,u)}for(c in e)if(u=e[c],n=a[c],e.hasOwnProperty(c)&&(u!=null||n!=null))switch(c){case"value":g=u;break;case"defaultValue":b=u;break;case"children":break;case"dangerouslySetInnerHTML":if(u!=null)throw Error(s(91));break;default:u!==n&&sl(l,t,c,u,e,n)}Mf(l,g,b);return;case"option":for(var L in a)if(g=a[L],a.hasOwnProperty(L)&&g!=null&&!e.hasOwnProperty(L))switch(L){case"selected":l.selected=!1;break;default:sl(l,t,L,null,e,g)}for(f in e)if(g=e[f],b=a[f],e.hasOwnProperty(f)&&g!==b&&(g!=null||b!=null))switch(f){case"selected":l.selected=g&&typeof g!="function"&&typeof g!="symbol";break;default:sl(l,t,f,g,e,b)}return;case"img":case"link":case"area":case"base":case"br":case"col":case"embed":case"hr":case"keygen":case"meta":case"param":case"source":case"track":case"wbr":case"menuitem":for(var X in a)g=a[X],a.hasOwnProperty(X)&&g!=null&&!e.hasOwnProperty(X)&&sl(l,t,X,null,e,g);for(m in e)if(g=e[m],b=a[m],e.hasOwnProperty(m)&&g!==b&&(g!=null||b!=null))switch(m){case"children":case"dangerouslySetInnerHTML":if(g!=null)throw Error(s(137,t));break;default:sl(l,t,m,g,e,b)}return;default:if(tc(t)){for(var ol in a)g=a[ol],a.hasOwnProperty(ol)&&g!==void 0&&!e.hasOwnProperty(ol)&&Gi(l,t,ol,void 0,e,g);for(z in e)g=e[z],b=a[z],!e.hasOwnProperty(z)||g===b||g===void 0&&b===void 0||Gi(l,t,z,g,e,b);return}}for(var r in a)g=a[r],a.hasOwnProperty(r)&&g!=null&&!e.hasOwnProperty(r)&&sl(l,t,r,null,e,g);for(A in e)g=e[A],b=a[A],!e.hasOwnProperty(A)||g===b||g==null&&b==null||sl(l,t,A,g,e,b)}var Xi=null,Qi=null;function _n(l){return l.nodeType===9?l:l.ownerDocument}function Ud(l){switch(l){case"http://www.w3.org/2000/svg":return 1;case"http://www.w3.org/1998/Math/MathML":return 2;default:return 0}}function Rd(l,t){if(l===0)switch(t){case"svg":return 1;case"math":return 2;default:return 0}return l===1&&t==="foreignObject"?0:l}function Zi(l,t){return l==="textarea"||l==="noscript"||typeof t.children=="string"||typeof t.children=="number"||typeof t.children=="bigint"||typeof t.dangerouslySetInnerHTML=="object"&&t.dangerouslySetInnerHTML!==null&&t.dangerouslySetInnerHTML.__html!=null}var Vi=null;function vv(){var l=window.event;return l&&l.type==="popstate"?l===Vi?!1:(Vi=l,!0):(Vi=null,!1)}var jd=typeof setTimeout=="function"?setTimeout:void 0,yv=typeof clearTimeout=="function"?clearTimeout:void 0,Hd=typeof Promise=="function"?Promise:void 0,hv=typeof queueMicrotask=="function"?queueMicrotask:typeof Hd<"u"?function(l){return Hd.resolve(null).then(l).catch(mv)}:jd;function mv(l){setTimeout(function(){throw l})}function da(l){return l==="head"}function qd(l,t){var a=t,e=0,u=0;do{var n=a.nextSibling;if(l.removeChild(a),n&&n.nodeType===8)if(a=n.data,a==="/$"){if(0<e&&8>e){a=e;var c=l.ownerDocument;if(a&1&&du(c.documentElement),a&2&&du(c.body),a&4)for(a=c.head,du(a),c=a.firstChild;c;){var i=c.nextSibling,f=c.nodeName;c[xe]||f==="SCRIPT"||f==="STYLE"||f==="LINK"&&c.rel.toLowerCase()==="stylesheet"||a.removeChild(c),c=i}}if(u===0){l.removeChild(n),Su(t);return}u--}else a==="$"||a==="$?"||a==="$!"?u++:e=a.charCodeAt(0)-48;else e=0;a=n}while(a);Su(t)}function Li(l){var t=l.firstChild;for(t&&t.nodeType===10&&(t=t.nextSibling);t;){var a=t;switch(t=t.nextSibling,a.nodeName){case"HTML":case"HEAD":case"BODY":Li(a),kn(a);continue;case"SCRIPT":case"STYLE":continue;case"LINK":if(a.rel.toLowerCase()==="stylesheet")continue}l.removeChild(a)}}function gv(l,t,a,e){for(;l.nodeType===1;){var u=a;if(l.nodeName.toLowerCase()!==t.toLowerCase()){if(!e&&(l.nodeName!=="INPUT"||l.type!=="hidden"))break}else if(e){if(!l[xe])switch(t){case"meta":if(!l.hasAttribute("itemprop"))break;return l;case"link":if(n=l.getAttribute("rel"),n==="stylesheet"&&l.hasAttribute("data-precedence"))break;if(n!==u.rel||l.getAttribute("href")!==(u.href==null||u.href===""?null:u.href)||l.getAttribute("crossorigin")!==(u.crossOrigin==null?null:u.crossOrigin)||l.getAttribute("title")!==(u.title==null?null:u.title))break;return l;case"style":if(l.hasAttribute("data-precedence"))break;return l;case"script":if(n=l.getAttribute("src"),(n!==(u.src==null?null:u.src)||l.getAttribute("type")!==(u.type==null?null:u.type)||l.getAttribute("crossorigin")!==(u.crossOrigin==null?null:u.crossOrigin))&&n&&l.hasAttribute("async")&&!l.hasAttribute("itemprop"))break;return l;default:return l}}else if(t==="input"&&l.type==="hidden"){var n=u.name==null?null:""+u.name;if(u.type==="hidden"&&l.getAttribute("name")===n)return l}else return l;if(l=Tt(l.nextSibling),l===null)break}return null}function bv(l,t,a){if(t==="")return null;for(;l.nodeType!==3;)if((l.nodeType!==1||l.nodeName!=="INPUT"||l.type!=="hidden")&&!a||(l=Tt(l.nextSibling),l===null))return null;return l}function Ki(l){return l.data==="$!"||l.data==="$?"&&l.ownerDocument.readyState==="complete"}function Sv(l,t){var a=l.ownerDocument;if(l.data!=="$?"||a.readyState==="complete")t();else{var e=function(){t(),a.removeEventListener("DOMContentLoaded",e)};a.addEventListener("DOMContentLoaded",e),l._reactRetry=e}}function Tt(l){for(;l!=null;l=l.nextSibling){var t=l.nodeType;if(t===1||t===3)break;if(t===8){if(t=l.data,t==="$"||t==="$!"||t==="$?"||t==="F!"||t==="F")break;if(t==="/$")return null}}return l}var Ji=null;function Yd(l){l=l.previousSibling;for(var t=0;l;){if(l.nodeType===8){var a=l.data;if(a==="$"||a==="$!"||a==="$?"){if(t===0)return l;t--}else a==="/$"&&t++}l=l.previousSibling}return null}function Bd(l,t,a){switch(t=_n(a),l){case"html":if(l=t.documentElement,!l)throw Error(s(452));return l;case"head":if(l=t.head,!l)throw Error(s(453));return l;case"body":if(l=t.body,!l)throw Error(s(454));return l;default:throw Error(s(451))}}function du(l){for(var t=l.attributes;t.length;)l.removeAttributeNode(t[0]);kn(l)}var gt=new Map,Cd=new Set;function Mn(l){return typeof l.getRootNode=="function"?l.getRootNode():l.nodeType===9?l:l.ownerDocument}var Lt=N.d;N.d={f:zv,r:Tv,D:Ev,C:Av,L:Ov,m:pv,X:_v,S:xv,M:Mv};function zv(){var l=Lt.f(),t=Sn();return l||t}function Tv(l){var t=Ga(l);t!==null&&t.tag===5&&t.type==="form"?eo(t):Lt.r(l)}var Se=typeof document>"u"?null:document;function Gd(l,t,a){var e=Se;if(e&&typeof t=="string"&&t){var u=ot(t);u='link[rel="'+l+'"][href="'+u+'"]',typeof a=="string"&&(u+='[crossorigin="'+a+'"]'),Cd.has(u)||(Cd.add(u),l={rel:l,crossOrigin:a,href:t},e.querySelector(u)===null&&(t=e.createElement("link"),Gl(t,"link",l),Ul(t),e.head.appendChild(t)))}}function Ev(l){Lt.D(l),Gd("dns-prefetch",l,null)}function Av(l,t){Lt.C(l,t),Gd("preconnect",l,t)}function Ov(l,t,a){Lt.L(l,t,a);var e=Se;if(e&&l&&t){var u='link[rel="preload"][as="'+ot(t)+'"]';t==="image"&&a&&a.imageSrcSet?(u+='[imagesrcset="'+ot(a.imageSrcSet)+'"]',typeof a.imageSizes=="string"&&(u+='[imagesizes="'+ot(a.imageSizes)+'"]')):u+='[href="'+ot(l)+'"]';var n=u;switch(t){case"style":n=ze(l);break;case"script":n=Te(l)}gt.has(n)||(l=D({rel:"preload",href:t==="image"&&a&&a.imageSrcSet?void 0:l,as:t},a),gt.set(n,l),e.querySelector(u)!==null||t==="style"&&e.querySelector(ru(n))||t==="script"&&e.querySelector(vu(n))||(t=e.createElement("link"),Gl(t,"link",l),Ul(t),e.head.appendChild(t)))}}function pv(l,t){Lt.m(l,t);var a=Se;if(a&&l){var e=t&&typeof t.as=="string"?t.as:"script",u='link[rel="modulepreload"][as="'+ot(e)+'"][href="'+ot(l)+'"]',n=u;switch(e){case"audioworklet":case"paintworklet":case"serviceworker":case"sharedworker":case"worker":case"script":n=Te(l)}if(!gt.has(n)&&(l=D({rel:"modulepreload",href:l},t),gt.set(n,l),a.querySelector(u)===null)){switch(e){case"audioworklet":case"paintworklet":case"serviceworker":case"sharedworker":case"worker":case"script":if(a.querySelector(vu(n)))return}e=a.createElement("link"),Gl(e,"link",l),Ul(e),a.head.appendChild(e)}}}function xv(l,t,a){Lt.S(l,t,a);var e=Se;if(e&&l){var u=Xa(e).hoistableStyles,n=ze(l);t=t||"default";var c=u.get(n);if(!c){var i={loading:0,preload:null};if(c=e.querySelector(ru(n)))i.loading=5;else{l=D({rel:"stylesheet",href:l,"data-precedence":t},a),(a=gt.get(n))&&wi(l,a);var f=c=e.createElement("link");Ul(f),Gl(f,"link",l),f._p=new Promise(function(m,z){f.onload=m,f.onerror=z}),f.addEventListener("load",function(){i.loading|=1}),f.addEventListener("error",function(){i.loading|=2}),i.loading|=4,Dn(c,t,e)}c={type:"stylesheet",instance:c,count:1,state:i},u.set(n,c)}}}function _v(l,t){Lt.X(l,t);var a=Se;if(a&&l){var e=Xa(a).hoistableScripts,u=Te(l),n=e.get(u);n||(n=a.querySelector(vu(u)),n||(l=D({src:l,async:!0},t),(t=gt.get(u))&&Wi(l,t),n=a.createElement("script"),Ul(n),Gl(n,"link",l),a.head.appendChild(n)),n={type:"script",instance:n,count:1,state:null},e.set(u,n))}}function Mv(l,t){Lt.M(l,t);var a=Se;if(a&&l){var e=Xa(a).hoistableScripts,u=Te(l),n=e.get(u);n||(n=a.querySelector(vu(u)),n||(l=D({src:l,async:!0,type:"module"},t),(t=gt.get(u))&&Wi(l,t),n=a.createElement("script"),Ul(n),Gl(n,"link",l),a.head.appendChild(n)),n={type:"script",instance:n,count:1,state:null},e.set(u,n))}}function Xd(l,t,a,e){var u=(u=J.current)?Mn(u):null;if(!u)throw Error(s(446));switch(l){case"meta":case"title":return null;case"style":return typeof a.precedence=="string"&&typeof a.href=="string"?(t=ze(a.href),a=Xa(u).hoistableStyles,e=a.get(t),e||(e={type:"style",instance:null,count:0,state:null},a.set(t,e)),e):{type:"void",instance:null,count:0,state:null};case"link":if(a.rel==="stylesheet"&&typeof a.href=="string"&&typeof a.precedence=="string"){l=ze(a.href);var n=Xa(u).hoistableStyles,c=n.get(l);if(c||(u=u.ownerDocument||u,c={type:"stylesheet",instance:null,count:0,state:{loading:0,preload:null}},n.set(l,c),(n=u.querySelector(ru(l)))&&!n._p&&(c.instance=n,c.state.loading=5),gt.has(l)||(a={rel:"preload",as:"style",href:a.href,crossOrigin:a.crossOrigin,integrity:a.integrity,media:a.media,hrefLang:a.hrefLang,referrerPolicy:a.referrerPolicy},gt.set(l,a),n||Dv(u,l,a,c.state))),t&&e===null)throw Error(s(528,""));return c}if(t&&e!==null)throw Error(s(529,""));return null;case"script":return t=a.async,a=a.src,typeof a=="string"&&t&&typeof t!="function"&&typeof t!="symbol"?(t=Te(a),a=Xa(u).hoistableScripts,e=a.get(t),e||(e={type:"script",instance:null,count:0,state:null},a.set(t,e)),e):{type:"void",instance:null,count:0,state:null};default:throw Error(s(444,l))}}function ze(l){return'href="'+ot(l)+'"'}function ru(l){return'link[rel="stylesheet"]['+l+"]"}function Qd(l){return D({},l,{"data-precedence":l.precedence,precedence:null})}function Dv(l,t,a,e){l.querySelector('link[rel="preload"][as="style"]['+t+"]")?e.loading=1:(t=l.createElement("link"),e.preload=t,t.addEventListener("load",function(){return e.loading|=1}),t.addEventListener("error",function(){return e.loading|=2}),Gl(t,"link",a),Ul(t),l.head.appendChild(t))}function Te(l){return'[src="'+ot(l)+'"]'}function vu(l){return"script[async]"+l}function Zd(l,t,a){if(t.count++,t.instance===null)switch(t.type){case"style":var e=l.querySelector('style[data-href~="'+ot(a.href)+'"]');if(e)return t.instance=e,Ul(e),e;var u=D({},a,{"data-href":a.href,"data-precedence":a.precedence,href:null,precedence:null});return e=(l.ownerDocument||l).createElement("style"),Ul(e),Gl(e,"style",u),Dn(e,a.precedence,l),t.instance=e;case"stylesheet":u=ze(a.href);var n=l.querySelector(ru(u));if(n)return t.state.loading|=4,t.instance=n,Ul(n),n;e=Qd(a),(u=gt.get(u))&&wi(e,u),n=(l.ownerDocument||l).createElement("link"),Ul(n);var c=n;return c._p=new Promise(function(i,f){c.onload=i,c.onerror=f}),Gl(n,"link",e),t.state.loading|=4,Dn(n,a.precedence,l),t.instance=n;case"script":return n=Te(a.src),(u=l.querySelector(vu(n)))?(t.instance=u,Ul(u),u):(e=a,(u=gt.get(n))&&(e=D({},a),Wi(e,u)),l=l.ownerDocument||l,u=l.createElement("script"),Ul(u),Gl(u,"link",e),l.head.appendChild(u),t.instance=u);case"void":return null;default:throw Error(s(443,t.type))}else t.type==="stylesheet"&&(t.state.loading&4)===0&&(e=t.instance,t.state.loading|=4,Dn(e,a.precedence,l));return t.instance}function Dn(l,t,a){for(var e=a.querySelectorAll('link[rel="stylesheet"][data-precedence],style[data-precedence]'),u=e.length?e[e.length-1]:null,n=u,c=0;c<e.length;c++){var i=e[c];if(i.dataset.precedence===t)n=i;else if(n!==u)break}n?n.parentNode.insertBefore(l,n.nextSibling):(t=a.nodeType===9?a.head:a,t.insertBefore(l,t.firstChild))}function wi(l,t){l.crossOrigin==null&&(l.crossOrigin=t.crossOrigin),l.referrerPolicy==null&&(l.referrerPolicy=t.referrerPolicy),l.title==null&&(l.title=t.title)}function Wi(l,t){l.crossOrigin==null&&(l.crossOrigin=t.crossOrigin),l.referrerPolicy==null&&(l.referrerPolicy=t.referrerPolicy),l.integrity==null&&(l.integrity=t.integrity)}var Nn=null;function Vd(l,t,a){if(Nn===null){var e=new Map,u=Nn=new Map;u.set(a,e)}else u=Nn,e=u.get(a),e||(e=new Map,u.set(a,e));if(e.has(l))return e;for(e.set(l,null),a=a.getElementsByTagName(l),u=0;u<a.length;u++){var n=a[u];if(!(n[xe]||n[Xl]||l==="link"&&n.getAttribute("rel")==="stylesheet")&&n.namespaceURI!=="http://www.w3.org/2000/svg"){var c=n.getAttribute(t)||"";c=l+c;var i=e.get(c);i?i.push(n):e.set(c,[n])}}return e}function Ld(l,t,a){l=l.ownerDocument||l,l.head.insertBefore(a,t==="title"?l.querySelector("head > title"):null)}function Nv(l,t,a){if(a===1||t.itemProp!=null)return!1;switch(l){case"meta":case"title":return!0;case"style":if(typeof t.precedence!="string"||typeof t.href!="string"||t.href==="")break;return!0;case"link":if(typeof t.rel!="string"||typeof t.href!="string"||t.href===""||t.onLoad||t.onError)break;switch(t.rel){case"stylesheet":return l=t.disabled,typeof t.precedence=="string"&&l==null;default:return!0}case"script":if(t.async&&typeof t.async!="function"&&typeof t.async!="symbol"&&!t.onLoad&&!t.onError&&t.src&&typeof t.src=="string")return!0}return!1}function Kd(l){return!(l.type==="stylesheet"&&(l.state.loading&3)===0)}var yu=null;function Uv(){}function Rv(l,t,a){if(yu===null)throw Error(s(475));var e=yu;if(t.type==="stylesheet"&&(typeof a.media!="string"||matchMedia(a.media).matches!==!1)&&(t.state.loading&4)===0){if(t.instance===null){var u=ze(a.href),n=l.querySelector(ru(u));if(n){l=n._p,l!==null&&typeof l=="object"&&typeof l.then=="function"&&(e.count++,e=Un.bind(e),l.then(e,e)),t.state.loading|=4,t.instance=n,Ul(n);return}n=l.ownerDocument||l,a=Qd(a),(u=gt.get(u))&&wi(a,u),n=n.createElement("link"),Ul(n);var c=n;c._p=new Promise(function(i,f){c.onload=i,c.onerror=f}),Gl(n,"link",a),t.instance=n}e.stylesheets===null&&(e.stylesheets=new Map),e.stylesheets.set(t,l),(l=t.state.preload)&&(t.state.loading&3)===0&&(e.count++,t=Un.bind(e),l.addEventListener("load",t),l.addEventListener("error",t))}}function jv(){if(yu===null)throw Error(s(475));var l=yu;return l.stylesheets&&l.count===0&&ki(l,l.stylesheets),0<l.count?function(t){var a=setTimeout(function(){if(l.stylesheets&&ki(l,l.stylesheets),l.unsuspend){var e=l.unsuspend;l.unsuspend=null,e()}},6e4);return l.unsuspend=t,function(){l.unsuspend=null,clearTimeout(a)}}:null}function Un(){if(this.count--,this.count===0){if(this.stylesheets)ki(this,this.stylesheets);else if(this.unsuspend){var l=this.unsuspend;this.unsuspend=null,l()}}}var Rn=null;function ki(l,t){l.stylesheets=null,l.unsuspend!==null&&(l.count++,Rn=new Map,t.forEach(Hv,l),Rn=null,Un.call(l))}function Hv(l,t){if(!(t.state.loading&4)){var a=Rn.get(l);if(a)var e=a.get(null);else{a=new Map,Rn.set(l,a);for(var u=l.querySelectorAll("link[data-precedence],style[data-precedence]"),n=0;n<u.length;n++){var c=u[n];(c.nodeName==="LINK"||c.getAttribute("media")!=="not all")&&(a.set(c.dataset.precedence,c),e=c)}e&&a.set(null,e)}u=t.instance,c=u.getAttribute("data-precedence"),n=a.get(c)||e,n===e&&a.set(null,u),a.set(c,u),this.count++,e=Un.bind(this),u.addEventListener("load",e),u.addEventListener("error",e),n?n.parentNode.insertBefore(u,n.nextSibling):(l=l.nodeType===9?l.head:l,l.insertBefore(u,l.firstChild)),t.state.loading|=4}}var hu={$$typeof:dl,Provider:null,Consumer:null,_currentValue:V,_currentValue2:V,_threadCount:0};function qv(l,t,a,e,u,n,c,i){this.tag=1,this.containerInfo=l,this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.next=this.pendingContext=this.context=this.cancelPendingCommit=null,this.callbackPriority=0,this.expirationTimes=Kn(-1),this.entangledLanes=this.shellSuspendCounter=this.errorRecoveryDisabledLanes=this.expiredLanes=this.warmLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=Kn(0),this.hiddenUpdates=Kn(null),this.identifierPrefix=e,this.onUncaughtError=u,this.onCaughtError=n,this.onRecoverableError=c,this.pooledCache=null,this.pooledCacheLanes=0,this.formState=i,this.incompleteTransitions=new Map}function Jd(l,t,a,e,u,n,c,i,f,m,z,A){return l=new qv(l,t,a,c,i,f,m,A),t=1,n===!0&&(t|=24),n=at(3,null,null,t),l.current=n,n.stateNode=l,t=Nc(),t.refCount++,l.pooledCache=t,t.refCount++,n.memoizedState={element:e,isDehydrated:a,cache:t},Hc(n),l}function wd(l){return l?(l=Pa,l):Pa}function Wd(l,t,a,e,u,n){u=wd(u),e.context===null?e.context=u:e.pendingContext=u,e=Pt(t),e.payload={element:a},n=n===void 0?null:n,n!==null&&(e.callback=n),a=It(l,e,t),a!==null&&(it(a,l,t),Ke(a,l,t))}function kd(l,t){if(l=l.memoizedState,l!==null&&l.dehydrated!==null){var a=l.retryLane;l.retryLane=a!==0&&a<t?a:t}}function $i(l,t){kd(l,t),(l=l.alternate)&&kd(l,t)}function $d(l){if(l.tag===13){var t=Fa(l,67108864);t!==null&&it(t,l,67108864),$i(l,67108864)}}var jn=!0;function Yv(l,t,a,e){var u=S.T;S.T=null;var n=N.p;try{N.p=2,Fi(l,t,a,e)}finally{N.p=n,S.T=u}}function Bv(l,t,a,e){var u=S.T;S.T=null;var n=N.p;try{N.p=8,Fi(l,t,a,e)}finally{N.p=n,S.T=u}}function Fi(l,t,a,e){if(jn){var u=Pi(e);if(u===null)Ci(l,t,e,Hn,a),Pd(l,e);else if(Gv(u,l,t,a,e))e.stopPropagation();else if(Pd(l,e),t&4&&-1<Cv.indexOf(l)){for(;u!==null;){var n=Ga(u);if(n!==null)switch(n.tag){case 3:if(n=n.stateNode,n.current.memoizedState.isDehydrated){var c=Sa(n.pendingLanes);if(c!==0){var i=n;for(i.pendingLanes|=2,i.entangledLanes|=2;c;){var f=1<<31-lt(c);i.entanglements[1]|=f,c&=~f}_t(n),(nl&6)===0&&(gn=Et()+500,fu(0))}}break;case 13:i=Fa(n,2),i!==null&&it(i,n,2),Sn(),$i(n,2)}if(n=Pi(e),n===null&&Ci(l,t,e,Hn,a),n===u)break;u=n}u!==null&&e.stopPropagation()}else Ci(l,t,e,null,a)}}function Pi(l){return l=ec(l),Ii(l)}var Hn=null;function Ii(l){if(Hn=null,l=Ca(l),l!==null){var t=q(l);if(t===null)l=null;else{var a=t.tag;if(a===13){if(l=R(t),l!==null)return l;l=null}else if(a===3){if(t.stateNode.current.memoizedState.isDehydrated)return t.tag===3?t.stateNode.containerInfo:null;l=null}else t!==l&&(l=null)}}return Hn=l,null}function Fd(l){switch(l){case"beforetoggle":case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"toggle":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 2;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 8;case"message":switch(Ar()){case of:return 2;case df:return 8;case pu:case Or:return 32;case rf:return 268435456;default:return 32}default:return 32}}var lf=!1,ra=null,va=null,ya=null,mu=new Map,gu=new Map,ha=[],Cv="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset".split(" ");function Pd(l,t){switch(l){case"focusin":case"focusout":ra=null;break;case"dragenter":case"dragleave":va=null;break;case"mouseover":case"mouseout":ya=null;break;case"pointerover":case"pointerout":mu.delete(t.pointerId);break;case"gotpointercapture":case"lostpointercapture":gu.delete(t.pointerId)}}function bu(l,t,a,e,u,n){return l===null||l.nativeEvent!==n?(l={blockedOn:t,domEventName:a,eventSystemFlags:e,nativeEvent:n,targetContainers:[u]},t!==null&&(t=Ga(t),t!==null&&$d(t)),l):(l.eventSystemFlags|=e,t=l.targetContainers,u!==null&&t.indexOf(u)===-1&&t.push(u),l)}function Gv(l,t,a,e,u){switch(t){case"focusin":return ra=bu(ra,l,t,a,e,u),!0;case"dragenter":return va=bu(va,l,t,a,e,u),!0;case"mouseover":return ya=bu(ya,l,t,a,e,u),!0;case"pointerover":var n=u.pointerId;return mu.set(n,bu(mu.get(n)||null,l,t,a,e,u)),!0;case"gotpointercapture":return n=u.pointerId,gu.set(n,bu(gu.get(n)||null,l,t,a,e,u)),!0}return!1}function Id(l){var t=Ca(l.target);if(t!==null){var a=q(t);if(a!==null){if(t=a.tag,t===13){if(t=R(a),t!==null){l.blockedOn=t,Rr(l.priority,function(){if(a.tag===13){var e=ct();e=Jn(e);var u=Fa(a,e);u!==null&&it(u,a,e),$i(a,e)}});return}}else if(t===3&&a.stateNode.current.memoizedState.isDehydrated){l.blockedOn=a.tag===3?a.stateNode.containerInfo:null;return}}}l.blockedOn=null}function qn(l){if(l.blockedOn!==null)return!1;for(var t=l.targetContainers;0<t.length;){var a=Pi(l.nativeEvent);if(a===null){a=l.nativeEvent;var e=new a.constructor(a.type,a);ac=e,a.target.dispatchEvent(e),ac=null}else return t=Ga(a),t!==null&&$d(t),l.blockedOn=a,!1;t.shift()}return!0}function lr(l,t,a){qn(l)&&a.delete(t)}function Xv(){lf=!1,ra!==null&&qn(ra)&&(ra=null),va!==null&&qn(va)&&(va=null),ya!==null&&qn(ya)&&(ya=null),mu.forEach(lr),gu.forEach(lr)}function Yn(l,t){l.blockedOn===t&&(l.blockedOn=null,lf||(lf=!0,v.unstable_scheduleCallback(v.unstable_NormalPriority,Xv)))}var Bn=null;function tr(l){Bn!==l&&(Bn=l,v.unstable_scheduleCallback(v.unstable_NormalPriority,function(){Bn===l&&(Bn=null);for(var t=0;t<l.length;t+=3){var a=l[t],e=l[t+1],u=l[t+2];if(typeof e!="function"){if(Ii(e||a)===null)continue;break}var n=Ga(a);n!==null&&(l.splice(t,3),t-=3,Ic(n,{pending:!0,data:u,method:a.method,action:e},e,u))}}))}function Su(l){function t(f){return Yn(f,l)}ra!==null&&Yn(ra,l),va!==null&&Yn(va,l),ya!==null&&Yn(ya,l),mu.forEach(t),gu.forEach(t);for(var a=0;a<ha.length;a++){var e=ha[a];e.blockedOn===l&&(e.blockedOn=null)}for(;0<ha.length&&(a=ha[0],a.blockedOn===null);)Id(a),a.blockedOn===null&&ha.shift();if(a=(l.ownerDocument||l).$$reactFormReplay,a!=null)for(e=0;e<a.length;e+=3){var u=a[e],n=a[e+1],c=u[wl]||null;if(typeof n=="function")c||tr(a);else if(c){var i=null;if(n&&n.hasAttribute("formAction")){if(u=n,c=n[wl]||null)i=c.formAction;else if(Ii(u)!==null)continue}else i=c.action;typeof i=="function"?a[e+1]=i:(a.splice(e,3),e-=3),tr(a)}}}function tf(l){this._internalRoot=l}Cn.prototype.render=tf.prototype.render=function(l){var t=this._internalRoot;if(t===null)throw Error(s(409));var a=t.current,e=ct();Wd(a,e,l,t,null,null)},Cn.prototype.unmount=tf.prototype.unmount=function(){var l=this._internalRoot;if(l!==null){this._internalRoot=null;var t=l.containerInfo;Wd(l.current,2,null,l,null,null),Sn(),t[Ba]=null}};function Cn(l){this._internalRoot=l}Cn.prototype.unstable_scheduleHydration=function(l){if(l){var t=gf();l={blockedOn:null,target:l,priority:t};for(var a=0;a<ha.length&&t!==0&&t<ha[a].priority;a++);ha.splice(a,0,l),a===0&&Id(l)}};var ar=_.version;if(ar!=="19.1.1")throw Error(s(527,ar,"19.1.1"));N.findDOMNode=function(l){var t=l._reactInternals;if(t===void 0)throw typeof l.render=="function"?Error(s(188)):(l=Object.keys(l).join(","),Error(s(268,l)));return l=M(t),l=l!==null?y(l):null,l=l===null?null:l.stateNode,l};var Qv={bundleType:0,version:"19.1.1",rendererPackageName:"react-dom",currentDispatcherRef:S,reconcilerVersion:"19.1.1"};if(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__<"u"){var Gn=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!Gn.isDisabled&&Gn.supportsFiber)try{Ae=Gn.inject(Qv),Il=Gn}catch{}}return Tu.createRoot=function(l,t){if(!x(l))throw Error(s(299));var a=!1,e="",u=bo,n=So,c=zo,i=null;return t!=null&&(t.unstable_strictMode===!0&&(a=!0),t.identifierPrefix!==void 0&&(e=t.identifierPrefix),t.onUncaughtError!==void 0&&(u=t.onUncaughtError),t.onCaughtError!==void 0&&(n=t.onCaughtError),t.onRecoverableError!==void 0&&(c=t.onRecoverableError),t.unstable_transitionCallbacks!==void 0&&(i=t.unstable_transitionCallbacks)),t=Jd(l,1,!1,null,null,a,e,u,n,c,i,null),l[Ba]=t.current,Bi(l),new tf(t)},Tu.hydrateRoot=function(l,t,a){if(!x(l))throw Error(s(299));var e=!1,u="",n=bo,c=So,i=zo,f=null,m=null;return a!=null&&(a.unstable_strictMode===!0&&(e=!0),a.identifierPrefix!==void 0&&(u=a.identifierPrefix),a.onUncaughtError!==void 0&&(n=a.onUncaughtError),a.onCaughtError!==void 0&&(c=a.onCaughtError),a.onRecoverableError!==void 0&&(i=a.onRecoverableError),a.unstable_transitionCallbacks!==void 0&&(f=a.unstable_transitionCallbacks),a.formState!==void 0&&(m=a.formState)),t=Jd(l,1,!0,t,a??null,e,u,n,c,i,f,m),t.context=wd(null),a=t.current,e=ct(),e=Jn(e),u=Pt(e),u.callback=null,It(a,u,e),a=e,t.current.lanes=a,pe(t,a),_t(t),l[Ba]=t.current,Bi(l),new Cn(t)},Tu.version="19.1.1",Tu}var rr;function Pv(){if(rr)return uf.exports;rr=1;function v(){if(!(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__>"u"||typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE!="function"))try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(v)}catch(_){console.error(_)}}return v(),uf.exports=Fv(),uf.exports}var Iv=Pv();const ly=yr(Iv),ty="portfolio-tracker-db",Ee="handles";function hr(){return new Promise((v,_)=>{const E=indexedDB.open(ty,1);E.onupgradeneeded=()=>{const s=E.result;s.objectStoreNames.contains(Ee)||s.createObjectStore(Ee)},E.onsuccess=()=>v(E.result),E.onerror=()=>_(E.error)})}async function vr(v,_){const E=await hr();await new Promise((s,x)=>{const j=E.transaction(Ee,"readwrite").objectStore(Ee).put(_,v);j.onsuccess=()=>s(),j.onerror=()=>x(j.error)})}async function ay(v){const _=await hr();return await new Promise((E,s)=>{const R=_.transaction(Ee,"readonly").objectStore(Ee).get(v);R.onsuccess=()=>E(R.result),R.onerror=()=>s(R.error)})}const mr=new TextEncoder().encode("PTv1.enc");async function gr(v,_){const E=new TextEncoder,s=await crypto.subtle.importKey("raw",E.encode(v),{name:"PBKDF2"},!1,["deriveKey"]);return await crypto.subtle.deriveKey({name:"PBKDF2",salt:_,iterations:15e4,hash:"SHA-256"},s,{name:"AES-GCM",length:256},!1,["encrypt","decrypt"])}function ey(...v){const _=v.reduce((x,q)=>x+q.length,0),E=new Uint8Array(_);let s=0;for(const x of v)E.set(x,s),s+=x.length;return E}async function uy(v,_){const E=new TextEncoder().encode(JSON.stringify(v,null,2)),s=crypto.getRandomValues(new Uint8Array(16)),x=crypto.getRandomValues(new Uint8Array(12)),q=await gr(_,s),R=new Uint8Array(await crypto.subtle.encrypt({name:"AES-GCM",iv:x},q,E));return ey(new Uint8Array(mr),s,x,R)}async function ny(v,_){const E=new Uint8Array(v),s=E.slice(0,8);if(!cy(s,mr))throw new Error("Invalid file format");const x=E.slice(8,24),q=E.slice(24,36),R=E.slice(36),j=await gr(_,x),M=await crypto.subtle.decrypt({name:"AES-GCM",iv:q},j,R),y=new TextDecoder().decode(new Uint8Array(M));return JSON.parse(y)}function cy(v,_){if(v.length!==_.length)return!1;for(let E=0;E<v.length;E++)if(v[E]!==_[E])return!1;return!0}async function iy(){try{const[v]=await window.showOpenFilePicker({types:[{description:"Portfolio",accept:{"application/octet-stream":[".enc"]}}]});return await vr("fileHandle",v),v}catch{const _=await window.showSaveFilePicker({suggestedName:"portfolio.enc",types:[{description:"Portfolio",accept:{"application/octet-stream":[".enc"]}}]});return await vr("fileHandle",_),_}}async function fy(){return await ay("fileHandle")}async function sy(v,_){const E=await v.getFile();if(E.size===0)return{version:1,assetTypes:{},allocation:{},snapshots:[]};const s=await E.arrayBuffer();return await ny(s,_)}async function oy(v,_,E){const s=await uy(E,_),x=await v.createWritable();await x.write(s),await x.close()}function Xn(v){return(v||[]).reduce((_,E)=>_+(Number(E.value)||0),0)}function dy(v){const _={};for(const E of v||[]){const s=E.type;_[s]=(_[s]||0)+(Number(E.value)||0)}return _}function ry(v){const _=Object.values(v).reduce((s,x)=>s+(Number(x)||0),0)||1,E={};for(const s of Object.keys(v))E[s]=(Number(v[s])||0)/_;return E}function vy(v,_,E){const s=Xn(v),x=dy(v),q=ry(_),R=Array.from(new Set([...Object.keys(x),...Object.keys(q)])),j=s+(Number(E)||0),M={};for(const B of R)M[B]=(q[B]||0)*j;const y={};for(const B of R)y[B]=Math.max(0,(M[B]||0)-(x[B]||0));const D=Object.values(y).reduce((B,ll)=>B+ll,0)||1,Q={};for(const B of R)Q[B]=y[B]/D*(Number(E)||0);return{totalNow:s,targetTotal:j,byCat:x,idealByCat:M,investPlan:Q}}function ba(v){try{return new Intl.NumberFormat(void 0,{style:"currency",currency:"EUR",maximumFractionDigits:0}).format(v)}catch{return`${Math.round(v).toLocaleString()} €`}}function yy({data:v}){const _=Ml.useRef(null);return Ml.useEffect(()=>{const E=_.current;if(!E)return;const s=window.devicePixelRatio||1,x=E.clientWidth*s,q=E.clientHeight*s;E.width=x,E.height=q;const R=E.getContext("2d");if(!R)return;R.clearRect(0,0,x,q);const j=32*s;if(R.strokeStyle="#2a2a2a",R.lineWidth=1,R.beginPath(),R.moveTo(j,j),R.lineTo(j,q-j),R.lineTo(x-j,q-j),R.stroke(),!v.length){R.fillStyle="#9aa0a6",R.font=`${14*s}px ui-sans-serif`,R.fillText("No data",12*s,20*s);return}const M=v.map((hl,ql)=>ql),y=v.map(hl=>hl.value),D=Math.min(...y),Q=Math.max(...y),B=hl=>j+hl/Math.max(1,M.length-1)*(x-2*j),ll=hl=>q-j-(hl-D)/Math.max(1,Q-D)*(q-2*j);R.beginPath();for(let hl=0;hl<M.length;hl++){const ql=B(M[hl]),dl=ll(y[hl]);hl===0?R.moveTo(ql,dl):R.lineTo(ql,dl)}R.lineWidth=2*s,R.strokeStyle="#8ab4f8",R.stroke();const ul=v[v.length-1],pl=B(M[M.length-1]),Dl=ll(y[y.length-1]);R.fillStyle="#e8eaed",R.font=`${12*s}px ui-sans-serif`,R.fillText(`${ul.label}: ${ba(ul.value)}`,pl-100*s,Dl-8*s)},[v]),O.jsx("canvas",{ref:_,className:"w-full h-64 rounded border border-zinc-800 bg-zinc-900"})}function ga({title:v,children:_,right:E}){return O.jsxs("div",{className:"bg-zinc-900/80 border border-zinc-800 rounded-2xl p-4",children:[O.jsxs("div",{className:"flex items-center justify-between mb-3",children:[O.jsx("h2",{className:"text-zinc-200 font-medium",children:v}),E]}),_]})}function Mt({label:v,value:_,onChange:E,type:s="text",placeholder:x="",className:q="",disabled:R=!1}){return O.jsxs("label",{className:`block text-sm ${q}`,children:[O.jsx("span",{className:"text-zinc-400",children:v}),O.jsx("input",{type:s,value:_,placeholder:x,onChange:j=>E(j.target.value),disabled:R,className:"mt-1 w-full rounded-lg bg-zinc-900 border border-zinc-800 px-3 py-2 text-zinc-100 focus:outline-none focus:ring-2 focus:ring-blue-500"})]})}function br({onClick:v,children:_}){return O.jsx("button",{onClick:v,className:"px-3 py-2 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700 text-sm",children:_})}function Sr(v,_){return _[v]?.label||v}function zr(){return Math.random().toString(36).slice(2)}function Au(v,_){const E=_[v]||{fields:[]},s={id:zr(),type:v,value:0};for(const x of E.fields)s[x.key]=x.default||"";return s}function hy(v){const{id:_,...E}=v;return E}const Eu={cash:{label:"Cash",fields:[{key:"currency",label:"Currency",default:"EUR"}]},real_estate:{label:"Real estate",fields:[{key:"description",label:"Description"}]},stock:{label:"Stock",fields:[{key:"ticker",label:"Ticker"}]},private_company_share:{label:"Private company share",fields:[{key:"company",label:"Company"}]}};function my({assets:v,setAssets:_,assetTypes:E}){function s(j,M){_(v.map(y=>y.id===j?{...y,...M}:y))}function x(j){_(v.filter(M=>M.id!==j))}function q(j){_([...v,Au(j,E)])}function R(j,M){_(v.map(y=>y.id!==j?y:{...Au(M,E),id:y.id,value:y.value}))}return O.jsxs("div",{className:"space-y-3",children:[O.jsx("div",{className:"flex flex-wrap gap-2",children:Object.entries(E).map(([j,M])=>O.jsxs(br,{onClick:()=>q(j),children:["Add ",M.label]},j))}),O.jsx("div",{className:"space-y-2",children:v.map(j=>{const M=E[j.type]||{fields:[]};return O.jsxs("div",{className:"grid md:grid-cols-12 gap-2 items-end bg-zinc-900/60 border border-zinc-800 rounded-xl p-3",children:[O.jsxs("div",{className:"md:col-span-2",children:[O.jsx("label",{className:"text-xs text-zinc-400",children:"Type"}),O.jsxs("select",{className:"mt-1 w-full rounded-lg bg-zinc-900 border border-zinc-800 px-3 py-2 text-zinc-100",value:j.type,onChange:y=>R(j.id,y.target.value),children:[!E[j.type]&&O.jsx("option",{value:j.type,children:j.type}),Object.entries(E).map(([y,D])=>O.jsx("option",{value:y,children:D.label},y))]})]}),M.fields.map(y=>O.jsx(Mt,{className:"md:col-span-3",label:y.label,value:j[y.key]||"",onChange:D=>s(j.id,{[y.key]:D})},y.key)),O.jsx(Mt,{className:"md:col-span-3",label:"Value",type:"number",value:String(j.value),onChange:y=>s(j.id,{value:Number(y||0)})}),O.jsx("div",{className:"md:col-span-1 flex justify-end",children:O.jsx("button",{onClick:()=>x(j.id),className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:"Remove"})})]},j.id)})})]})}function gy({assetTypes:v,setAssetTypes:_}){function E(y,D){_({...v,[y]:{...v[y],label:D}})}function s(y,D,Q){const B=v[y],ll=B.fields.map((ul,pl)=>pl===D?{...ul,key:Q}:ul);_({...v,[y]:{...B,fields:ll}})}function x(y,D,Q){const B=v[y],ll=B.fields.map((ul,pl)=>pl===D?{...ul,label:Q}:ul);_({...v,[y]:{...B,fields:ll}})}function q(){const y=window.prompt("New type key","new_type");y&&!v[y]&&_({...v,[y]:{label:y,fields:[]}})}function R(y){const{[y]:D,...Q}=v;_(Q)}function j(y){const D=window.prompt("Field key","field"),Q=window.prompt("Field label","Field");if(D&&Q){const B=v[y];_({...v,[y]:{...B,fields:[...B.fields,{key:D,label:Q}]}})}}function M(y,D){const Q=v[y],B=Q.fields.filter((ll,ul)=>ul!==D);_({...v,[y]:{...Q,fields:B}})}return O.jsxs("div",{className:"space-y-4",children:[Object.entries(v).map(([y,D])=>O.jsxs("div",{className:"border border-zinc-800 bg-zinc-900/60 rounded-xl p-3 space-y-2",children:[O.jsxs("div",{className:"flex items-end gap-2",children:[O.jsx(Mt,{label:"Type",value:y,onChange:()=>{},disabled:!0,className:"md:col-span-3"}),O.jsx(Mt,{label:"Label",value:D.label,onChange:Q=>E(y,Q),className:"md:col-span-3"}),O.jsx("div",{className:"flex-1 text-right",children:O.jsx("button",{onClick:()=>R(y),className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:"Delete"})})]}),D.fields.map((Q,B)=>O.jsxs("div",{className:"flex items-end gap-2",children:[O.jsx(Mt,{label:"Field key",value:Q.key,onChange:ll=>s(y,B,ll)}),O.jsx(Mt,{label:"Field label",value:Q.label,onChange:ll=>x(y,B,ll)}),O.jsx("button",{onClick:()=>M(y,B),className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:"Remove"})]},B)),O.jsx("button",{onClick:()=>j(y),className:"px-3 py-2 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700 text-sm",children:"Add field"})]},y)),O.jsx(br,{onClick:q,children:"Add type"})]})}function by({allocation:v,setAllocation:_,assetTypes:E}){const s=Array.from(new Set([...Object.keys(E),...Object.keys(v||{})])),x=Object.values(v||{}).reduce((M,y)=>M+(Number(y)||0),0);function q(M,y){_({...v,[M]:y})}function R(M){const{[M]:y,...D}=v;_(D)}function j(){const M=window.prompt("New category key (slug)","other");M&&_({...v,[M]:0})}return O.jsxs("div",{className:"space-y-2",children:[s.map(M=>O.jsxs("div",{className:"grid grid-cols-12 items-end gap-2",children:[O.jsx("div",{className:"col-span-6 text-zinc-300",children:Sr(M,E)}),O.jsx("div",{className:"col-span-4",children:O.jsx(Mt,{label:"Target %",type:"number",value:String(v[M]??0),onChange:y=>q(M,Number(y||0))})}),O.jsx("div",{className:"col-span-2 flex justify-end",children:O.jsx("button",{onClick:()=>R(M),className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:"Remove"})})]},M)),O.jsxs("div",{className:"flex items-center justify-between mt-2",children:[O.jsx("button",{onClick:j,className:"px-3 py-2 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700 text-sm",children:"Add category"}),O.jsxs("div",{className:`text-sm ${x===100?"text-emerald-400":"text-amber-400"}`,children:["Total: ",x,"% ",x!==100&&"(will be normalized)"]})]})]})}function Sy({data:v,assetTypes:_}){const E=Object.keys(v.investPlan||{}).sort();return E.length===0?null:O.jsxs("div",{className:"mt-4",children:[O.jsxs("div",{className:"text-sm text-zinc-400 mb-2",children:["Now: ",O.jsx("b",{className:"text-zinc-200",children:ba(v.totalNow)})," → Target: ",O.jsx("b",{className:"text-zinc-200",children:ba(v.targetTotal)})]}),O.jsxs("table",{className:"w-full text-sm",children:[O.jsx("thead",{className:"text-zinc-400",children:O.jsxs("tr",{children:[O.jsx("th",{className:"text-left py-1",children:"Category"}),O.jsx("th",{className:"text-right py-1",children:"Current"}),O.jsx("th",{className:"text-right py-1",children:"Ideal"}),O.jsx("th",{className:"text-right py-1",children:"Invest now"})]})}),O.jsx("tbody",{children:E.map(s=>O.jsxs("tr",{className:"border-t border-zinc-800",children:[O.jsx("td",{className:"py-1 text-zinc-200",children:Sr(s,_)}),O.jsx("td",{className:"py-1 text-right text-zinc-300",children:ba(v.byCat[s]||0)}),O.jsx("td",{className:"py-1 text-right text-zinc-300",children:ba(v.idealByCat[s]||0)}),O.jsx("td",{className:"py-1 text-right font-medium text-zinc-100",children:ba(v.investPlan[s]||0)})]},s))})]})]})}function zy({assetTypes:v,setAssetTypes:_,allocation:E,setAllocation:s}){return O.jsxs("div",{className:"space-y-6",children:[O.jsx(ga,{title:"Asset Types",children:O.jsx(gy,{assetTypes:v,setAssetTypes:_})}),O.jsx(ga,{title:"Allocation",children:O.jsx(by,{allocation:E,setAllocation:s,assetTypes:v})})]})}function Ty(v,_){const E=(v||[]).map(x=>({date:new Date(x.asOf),value:Xn(x.assets||[])}));return Ey(E,_)}function Ey(v,_){const E=x=>_==="monthly"?`${x.getFullYear()}-${String(x.getMonth()+1).padStart(2,"0")}`:`${x.getFullYear()}`,s=new Map;for(const x of v||[]){const q=E(x.date),R=s.get(q);(!R||x.date.getTime()>R.lastDate)&&s.set(q,{label:q,value:x.value,lastDate:x.date.getTime()})}return Array.from(s.values()).sort((x,q)=>x.lastDate-q.lastDate)}function Ay(){const[v,_]=Ml.useState(""),[E,s]=Ml.useState(null),[x,q]=Ml.useState([]),[R,j]=Ml.useState(()=>new Date().toISOString().slice(0,10)),[M,y]=Ml.useState(Eu),[D,Q]=Ml.useState([Au("cash",Eu),Au("real_estate",Eu),Au("stock",Eu)]),[B,ll]=Ml.useState({cash:20,real_estate:50,stock:30}),[ul,pl]=Ml.useState(0),[Dl,hl]=Ml.useState(!1),[ql,dl]=Ml.useState(null),[Vl,P]=Ml.useState("monthly"),[Yl,Ll]=Ml.useState("main");Ml.useEffect(()=>{(async()=>{try{const K=await fy();if(K&&K.queryPermission){const il=await K.queryPermission({mode:"readwrite"});(il==="granted"||il==="prompt")&&s(K)}}catch(K){console.warn("No saved handle",K)}})()},[]);const Kl=Ml.useMemo(()=>Xn(D),[D]),ft=Ml.useMemo(()=>Ty(x,Vl),[x,Vl]),Kt=Ml.useMemo(()=>vy(D,B,ul||0),[D,B,ul]);async function bt(){try{const K=await iy();s(K)}catch(K){dl(K&&K.message?K.message:String(K))}}async function Nl(){if(!E||!v)return dl("Pick a file and enter password first.");hl(!0),dl(null);try{const K=await sy(E,v),il=K.snapshots||[];q(il);const S=il[il.length-1];S&&Q((S.assets||[]).map(N=>({...N,id:zr()}))),ll(K.allocation||{}),y(K.assetTypes||Eu)}catch(K){dl(K&&K.message?K.message:String(K))}finally{hl(!1)}}async function Dt(){if(!E||!v)return dl("Pick a file and enter password first.");hl(!0),dl(null);try{const K={asOf:R,assets:D.map(hy)},il=[...x,K];await oy(E,v,{version:1,assetTypes:M,allocation:B,snapshots:il}),q(il),alert("Saved snapshot")}catch(K){dl(K&&K.message?K.message:String(K))}finally{hl(!1)}}return O.jsx("div",{className:"min-h-screen bg-zinc-950 text-zinc-100",children:O.jsxs("div",{className:"max-w-6xl mx-auto p-6 space-y-6",children:[O.jsxs("header",{className:"flex items-center justify-between",children:[O.jsxs("div",{children:[O.jsx("h1",{className:"text-2xl font-semibold",children:"Local‑First Portfolio Tracker"}),O.jsx("p",{className:"text-sm text-zinc-400",children:"Private by default · Encrypted portfolio file on your disk"})]}),O.jsxs("div",{className:"flex items-end gap-2",children:[O.jsx(Mt,{label:"Password",type:"password",value:v,onChange:_,className:"w-56"}),O.jsx("button",{onClick:bt,className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:E?"Change File":"Choose File"}),O.jsx("button",{onClick:Nl,className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:"Load"}),O.jsx("button",{onClick:Dt,className:"h-10 px-3 rounded-lg bg-blue-600 hover:bg-blue-500",children:"Save snapshot"}),O.jsx("button",{onClick:()=>Ll(Yl==="main"?"config":"main"),className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:Yl==="main"?"Config":"Back"})]})]}),ql&&O.jsx("div",{className:"p-3 rounded-xl bg-red-900/30 border border-red-800 text-red-200",children:ql}),Dl&&O.jsx("div",{className:"p-3 rounded-xl bg-zinc-800 text-zinc-300",children:"Working…"}),Yl==="main"?O.jsxs(O.Fragment,{children:[O.jsxs("div",{className:"grid md:grid-cols-3 gap-6",children:[O.jsx(ga,{title:"As of date",children:O.jsx(Mt,{label:"Date",type:"date",value:R,onChange:j})}),O.jsxs(ga,{title:"Net worth (current)",children:[O.jsx("div",{className:"text-3xl font-semibold",children:ba(Kl)}),O.jsx("div",{className:"text-xs text-zinc-400 mt-1",children:"Computed from asset list"})]}),O.jsx(ga,{title:"History view",right:O.jsxs("select",{value:Vl,onChange:K=>P(K.target.value),className:"bg-zinc-900 border border-zinc-700 rounded-lg px-2 py-1 text-sm",children:[O.jsx("option",{value:"monthly",children:"Monthly"}),O.jsx("option",{value:"yearly",children:"Yearly"})]}),children:O.jsx(yy,{data:ft})})]}),O.jsxs("div",{className:"grid lg:grid-cols-2 gap-6",children:[O.jsx(ga,{title:"Assets",children:O.jsx(my,{assets:D,setAssets:Q,assetTypes:M})}),O.jsxs(ga,{title:"Rebalance",children:[O.jsxs("div",{className:"grid grid-cols-2 gap-4",children:[O.jsx(Mt,{label:"New capital to invest",type:"number",value:String(ul),onChange:K=>pl(Number(K||0))}),O.jsx("div",{className:"text-sm text-zinc-400 flex items-end",children:"Distribute this to move toward target without selling."})]}),O.jsx(Sy,{data:Kt,assetTypes:M})]})]}),O.jsx(ga,{title:"Loaded snapshots",children:x.length===0?O.jsx("div",{className:"text-zinc-400 text-sm",children:"None yet. Create your first snapshot above."}):O.jsx("ul",{className:"text-sm max-h-48 overflow-auto space-y-1",children:x.map((K,il)=>O.jsxs("li",{className:"text-zinc-300",children:[new Date(K.asOf).toISOString().slice(0,10)," — ",ba(Xn(K.assets||[]))]},il))})})]}):O.jsx(zy,{assetTypes:M,setAssetTypes:y,allocation:B,setAllocation:ll})]})})}ly.createRoot(document.getElementById("root")).render(O.jsx(Jv.StrictMode,{children:O.jsx(Ay,{})}));
+`).replace(dv,"")}function Nd(l,t){return t=Dd(t),Dd(l)===t}function xn(){}function sl(l,t,a,e,u,n){switch(a){case"children":typeof e=="string"?t==="body"||t==="textarea"&&e===""||La(l,e):(typeof e=="number"||typeof e=="bigint")&&t!=="body"&&La(l,""+e);break;case"className":Nu(l,"class",e);break;case"tabIndex":Nu(l,"tabindex",e);break;case"dir":case"role":case"viewBox":case"width":case"height":Nu(l,a,e);break;case"style":Uf(l,e,n);break;case"data":if(t!=="object"){Nu(l,"data",e);break}case"src":case"href":if(e===""&&(t!=="a"||a!=="href")){l.removeAttribute(a);break}if(e==null||typeof e=="function"||typeof e=="symbol"||typeof e=="boolean"){l.removeAttribute(a);break}e=ju(""+e),l.setAttribute(a,e);break;case"action":case"formAction":if(typeof e=="function"){l.setAttribute(a,"javascript:throw new Error('A React form was unexpectedly submitted. If you called form.submit() manually, consider using form.requestSubmit() instead. If you\\'re trying to use event.stopPropagation() in a submit event handler, consider also calling event.preventDefault().')");break}else typeof n=="function"&&(a==="formAction"?(t!=="input"&&sl(l,t,"name",u.name,u,null),sl(l,t,"formEncType",u.formEncType,u,null),sl(l,t,"formMethod",u.formMethod,u,null),sl(l,t,"formTarget",u.formTarget,u,null)):(sl(l,t,"encType",u.encType,u,null),sl(l,t,"method",u.method,u,null),sl(l,t,"target",u.target,u,null)));if(e==null||typeof e=="symbol"||typeof e=="boolean"){l.removeAttribute(a);break}e=ju(""+e),l.setAttribute(a,e);break;case"onClick":e!=null&&(l.onclick=xn);break;case"onScroll":e!=null&&F("scroll",l);break;case"onScrollEnd":e!=null&&F("scrollend",l);break;case"dangerouslySetInnerHTML":if(e!=null){if(typeof e!="object"||!("__html"in e))throw Error(s(61));if(a=e.__html,a!=null){if(u.children!=null)throw Error(s(60));l.innerHTML=a}}break;case"multiple":l.multiple=e&&typeof e!="function"&&typeof e!="symbol";break;case"muted":l.muted=e&&typeof e!="function"&&typeof e!="symbol";break;case"suppressContentEditableWarning":case"suppressHydrationWarning":case"defaultValue":case"defaultChecked":case"innerHTML":case"ref":break;case"autoFocus":break;case"xlinkHref":if(e==null||typeof e=="function"||typeof e=="boolean"||typeof e=="symbol"){l.removeAttribute("xlink:href");break}a=ju(""+e),l.setAttributeNS("http://www.w3.org/1999/xlink","xlink:href",a);break;case"contentEditable":case"spellCheck":case"draggable":case"value":case"autoReverse":case"externalResourcesRequired":case"focusable":case"preserveAlpha":e!=null&&typeof e!="function"&&typeof e!="symbol"?l.setAttribute(a,""+e):l.removeAttribute(a);break;case"inert":case"allowFullScreen":case"async":case"autoPlay":case"controls":case"default":case"defer":case"disabled":case"disablePictureInPicture":case"disableRemotePlayback":case"formNoValidate":case"hidden":case"loop":case"noModule":case"noValidate":case"open":case"playsInline":case"readOnly":case"required":case"reversed":case"scoped":case"seamless":case"itemScope":e&&typeof e!="function"&&typeof e!="symbol"?l.setAttribute(a,""):l.removeAttribute(a);break;case"capture":case"download":e===!0?l.setAttribute(a,""):e!==!1&&e!=null&&typeof e!="function"&&typeof e!="symbol"?l.setAttribute(a,e):l.removeAttribute(a);break;case"cols":case"rows":case"size":case"span":e!=null&&typeof e!="function"&&typeof e!="symbol"&&!isNaN(e)&&1<=e?l.setAttribute(a,e):l.removeAttribute(a);break;case"rowSpan":case"start":e==null||typeof e=="function"||typeof e=="symbol"||isNaN(e)?l.removeAttribute(a):l.setAttribute(a,e);break;case"popover":F("beforetoggle",l),F("toggle",l),Du(l,"popover",e);break;case"xlinkActuate":Nt(l,"http://www.w3.org/1999/xlink","xlink:actuate",e);break;case"xlinkArcrole":Nt(l,"http://www.w3.org/1999/xlink","xlink:arcrole",e);break;case"xlinkRole":Nt(l,"http://www.w3.org/1999/xlink","xlink:role",e);break;case"xlinkShow":Nt(l,"http://www.w3.org/1999/xlink","xlink:show",e);break;case"xlinkTitle":Nt(l,"http://www.w3.org/1999/xlink","xlink:title",e);break;case"xlinkType":Nt(l,"http://www.w3.org/1999/xlink","xlink:type",e);break;case"xmlBase":Nt(l,"http://www.w3.org/XML/1998/namespace","xml:base",e);break;case"xmlLang":Nt(l,"http://www.w3.org/XML/1998/namespace","xml:lang",e);break;case"xmlSpace":Nt(l,"http://www.w3.org/XML/1998/namespace","xml:space",e);break;case"is":Du(l,"is",e);break;case"innerText":case"textContent":break;default:(!(2<a.length)||a[0]!=="o"&&a[0]!=="O"||a[1]!=="n"&&a[1]!=="N")&&(a=Qr.get(a)||a,Du(l,a,e))}}function Gi(l,t,a,e,u,n){switch(a){case"style":Uf(l,e,n);break;case"dangerouslySetInnerHTML":if(e!=null){if(typeof e!="object"||!("__html"in e))throw Error(s(61));if(a=e.__html,a!=null){if(u.children!=null)throw Error(s(60));l.innerHTML=a}}break;case"children":typeof e=="string"?La(l,e):(typeof e=="number"||typeof e=="bigint")&&La(l,""+e);break;case"onScroll":e!=null&&F("scroll",l);break;case"onScrollEnd":e!=null&&F("scrollend",l);break;case"onClick":e!=null&&(l.onclick=xn);break;case"suppressContentEditableWarning":case"suppressHydrationWarning":case"innerHTML":case"ref":break;case"innerText":case"textContent":break;default:if(!zf.hasOwnProperty(a))l:{if(a[0]==="o"&&a[1]==="n"&&(u=a.endsWith("Capture"),t=a.slice(2,u?a.length-7:void 0),n=l[wl]||null,n=n!=null?n[a]:null,typeof n=="function"&&l.removeEventListener(t,n,u),typeof e=="function")){typeof n!="function"&&n!==null&&(a in l?l[a]=null:l.hasAttribute(a)&&l.removeAttribute(a)),l.addEventListener(t,e,u);break l}a in l?l[a]=e:e===!0?l.setAttribute(a,""):Du(l,a,e)}}}function Gl(l,t,a){switch(t){case"div":case"span":case"svg":case"path":case"a":case"g":case"p":case"li":break;case"img":F("error",l),F("load",l);var e=!1,u=!1,n;for(n in a)if(a.hasOwnProperty(n)){var c=a[n];if(c!=null)switch(n){case"src":e=!0;break;case"srcSet":u=!0;break;case"children":case"dangerouslySetInnerHTML":throw Error(s(137,t));default:sl(l,t,n,c,a,null)}}u&&sl(l,t,"srcSet",a.srcSet,a,null),e&&sl(l,t,"src",a.src,a,null);return;case"input":F("invalid",l);var i=n=c=u=null,f=null,m=null;for(e in a)if(a.hasOwnProperty(e)){var z=a[e];if(z!=null)switch(e){case"name":u=z;break;case"type":c=z;break;case"checked":f=z;break;case"defaultChecked":m=z;break;case"value":n=z;break;case"defaultValue":i=z;break;case"children":case"dangerouslySetInnerHTML":if(z!=null)throw Error(s(137,t));break;default:sl(l,t,e,z,a,null)}}_f(l,n,i,f,m,c,u,!1),Uu(l);return;case"select":F("invalid",l),e=c=n=null;for(u in a)if(a.hasOwnProperty(u)&&(i=a[u],i!=null))switch(u){case"value":n=i;break;case"defaultValue":c=i;break;case"multiple":e=i;default:sl(l,t,u,i,a,null)}t=n,a=c,l.multiple=!!e,t!=null?Va(l,!!e,t,!1):a!=null&&Va(l,!!e,a,!0);return;case"textarea":F("invalid",l),n=u=e=null;for(c in a)if(a.hasOwnProperty(c)&&(i=a[c],i!=null))switch(c){case"value":e=i;break;case"defaultValue":u=i;break;case"children":n=i;break;case"dangerouslySetInnerHTML":if(i!=null)throw Error(s(91));break;default:sl(l,t,c,i,a,null)}Df(l,e,u,n),Uu(l);return;case"option":for(f in a)if(a.hasOwnProperty(f)&&(e=a[f],e!=null))switch(f){case"selected":l.selected=e&&typeof e!="function"&&typeof e!="symbol";break;default:sl(l,t,f,e,a,null)}return;case"dialog":F("beforetoggle",l),F("toggle",l),F("cancel",l),F("close",l);break;case"iframe":case"object":F("load",l);break;case"video":case"audio":for(e=0;e<su.length;e++)F(su[e],l);break;case"image":F("error",l),F("load",l);break;case"details":F("toggle",l);break;case"embed":case"source":case"link":F("error",l),F("load",l);case"area":case"base":case"br":case"col":case"hr":case"keygen":case"meta":case"param":case"track":case"wbr":case"menuitem":for(m in a)if(a.hasOwnProperty(m)&&(e=a[m],e!=null))switch(m){case"children":case"dangerouslySetInnerHTML":throw Error(s(137,t));default:sl(l,t,m,e,a,null)}return;default:if(tc(t)){for(z in a)a.hasOwnProperty(z)&&(e=a[z],e!==void 0&&Gi(l,t,z,e,a,void 0));return}}for(i in a)a.hasOwnProperty(i)&&(e=a[i],e!=null&&sl(l,t,i,e,a,null))}function rv(l,t,a,e){switch(t){case"div":case"span":case"svg":case"path":case"a":case"g":case"p":case"li":break;case"input":var u=null,n=null,c=null,i=null,f=null,m=null,z=null;for(b in a){var A=a[b];if(a.hasOwnProperty(b)&&A!=null)switch(b){case"checked":break;case"value":break;case"defaultValue":f=A;default:e.hasOwnProperty(b)||sl(l,t,b,null,e,A)}}for(var g in e){var b=e[g];if(A=a[g],e.hasOwnProperty(g)&&(b!=null||A!=null))switch(g){case"type":n=b;break;case"name":u=b;break;case"checked":m=b;break;case"defaultChecked":z=b;break;case"value":c=b;break;case"defaultValue":i=b;break;case"children":case"dangerouslySetInnerHTML":if(b!=null)throw Error(s(137,t));break;default:b!==A&&sl(l,t,g,b,e,A)}}In(l,c,i,f,m,z,n,u);return;case"select":b=c=i=g=null;for(n in a)if(f=a[n],a.hasOwnProperty(n)&&f!=null)switch(n){case"value":break;case"multiple":b=f;default:e.hasOwnProperty(n)||sl(l,t,n,null,e,f)}for(u in e)if(n=e[u],f=a[u],e.hasOwnProperty(u)&&(n!=null||f!=null))switch(u){case"value":g=n;break;case"defaultValue":i=n;break;case"multiple":c=n;default:n!==f&&sl(l,t,u,n,e,f)}t=i,a=c,e=b,g!=null?Va(l,!!a,g,!1):!!e!=!!a&&(t!=null?Va(l,!!a,t,!0):Va(l,!!a,a?[]:"",!1));return;case"textarea":b=g=null;for(i in a)if(u=a[i],a.hasOwnProperty(i)&&u!=null&&!e.hasOwnProperty(i))switch(i){case"value":break;case"children":break;default:sl(l,t,i,null,e,u)}for(c in e)if(u=e[c],n=a[c],e.hasOwnProperty(c)&&(u!=null||n!=null))switch(c){case"value":g=u;break;case"defaultValue":b=u;break;case"children":break;case"dangerouslySetInnerHTML":if(u!=null)throw Error(s(91));break;default:u!==n&&sl(l,t,c,u,e,n)}Mf(l,g,b);return;case"option":for(var L in a)if(g=a[L],a.hasOwnProperty(L)&&g!=null&&!e.hasOwnProperty(L))switch(L){case"selected":l.selected=!1;break;default:sl(l,t,L,null,e,g)}for(f in e)if(g=e[f],b=a[f],e.hasOwnProperty(f)&&g!==b&&(g!=null||b!=null))switch(f){case"selected":l.selected=g&&typeof g!="function"&&typeof g!="symbol";break;default:sl(l,t,f,g,e,b)}return;case"img":case"link":case"area":case"base":case"br":case"col":case"embed":case"hr":case"keygen":case"meta":case"param":case"source":case"track":case"wbr":case"menuitem":for(var X in a)g=a[X],a.hasOwnProperty(X)&&g!=null&&!e.hasOwnProperty(X)&&sl(l,t,X,null,e,g);for(m in e)if(g=e[m],b=a[m],e.hasOwnProperty(m)&&g!==b&&(g!=null||b!=null))switch(m){case"children":case"dangerouslySetInnerHTML":if(g!=null)throw Error(s(137,t));break;default:sl(l,t,m,g,e,b)}return;default:if(tc(t)){for(var ol in a)g=a[ol],a.hasOwnProperty(ol)&&g!==void 0&&!e.hasOwnProperty(ol)&&Gi(l,t,ol,void 0,e,g);for(z in e)g=e[z],b=a[z],!e.hasOwnProperty(z)||g===b||g===void 0&&b===void 0||Gi(l,t,z,g,e,b);return}}for(var r in a)g=a[r],a.hasOwnProperty(r)&&g!=null&&!e.hasOwnProperty(r)&&sl(l,t,r,null,e,g);for(A in e)g=e[A],b=a[A],!e.hasOwnProperty(A)||g===b||g==null&&b==null||sl(l,t,A,g,e,b)}var Xi=null,Qi=null;function _n(l){return l.nodeType===9?l:l.ownerDocument}function Ud(l){switch(l){case"http://www.w3.org/2000/svg":return 1;case"http://www.w3.org/1998/Math/MathML":return 2;default:return 0}}function Rd(l,t){if(l===0)switch(t){case"svg":return 1;case"math":return 2;default:return 0}return l===1&&t==="foreignObject"?0:l}function Zi(l,t){return l==="textarea"||l==="noscript"||typeof t.children=="string"||typeof t.children=="number"||typeof t.children=="bigint"||typeof t.dangerouslySetInnerHTML=="object"&&t.dangerouslySetInnerHTML!==null&&t.dangerouslySetInnerHTML.__html!=null}var Vi=null;function vv(){var l=window.event;return l&&l.type==="popstate"?l===Vi?!1:(Vi=l,!0):(Vi=null,!1)}var jd=typeof setTimeout=="function"?setTimeout:void 0,yv=typeof clearTimeout=="function"?clearTimeout:void 0,Hd=typeof Promise=="function"?Promise:void 0,hv=typeof queueMicrotask=="function"?queueMicrotask:typeof Hd<"u"?function(l){return Hd.resolve(null).then(l).catch(mv)}:jd;function mv(l){setTimeout(function(){throw l})}function da(l){return l==="head"}function qd(l,t){var a=t,e=0,u=0;do{var n=a.nextSibling;if(l.removeChild(a),n&&n.nodeType===8)if(a=n.data,a==="/$"){if(0<e&&8>e){a=e;var c=l.ownerDocument;if(a&1&&du(c.documentElement),a&2&&du(c.body),a&4)for(a=c.head,du(a),c=a.firstChild;c;){var i=c.nextSibling,f=c.nodeName;c[xe]||f==="SCRIPT"||f==="STYLE"||f==="LINK"&&c.rel.toLowerCase()==="stylesheet"||a.removeChild(c),c=i}}if(u===0){l.removeChild(n),Su(t);return}u--}else a==="$"||a==="$?"||a==="$!"?u++:e=a.charCodeAt(0)-48;else e=0;a=n}while(a);Su(t)}function Li(l){var t=l.firstChild;for(t&&t.nodeType===10&&(t=t.nextSibling);t;){var a=t;switch(t=t.nextSibling,a.nodeName){case"HTML":case"HEAD":case"BODY":Li(a),kn(a);continue;case"SCRIPT":case"STYLE":continue;case"LINK":if(a.rel.toLowerCase()==="stylesheet")continue}l.removeChild(a)}}function gv(l,t,a,e){for(;l.nodeType===1;){var u=a;if(l.nodeName.toLowerCase()!==t.toLowerCase()){if(!e&&(l.nodeName!=="INPUT"||l.type!=="hidden"))break}else if(e){if(!l[xe])switch(t){case"meta":if(!l.hasAttribute("itemprop"))break;return l;case"link":if(n=l.getAttribute("rel"),n==="stylesheet"&&l.hasAttribute("data-precedence"))break;if(n!==u.rel||l.getAttribute("href")!==(u.href==null||u.href===""?null:u.href)||l.getAttribute("crossorigin")!==(u.crossOrigin==null?null:u.crossOrigin)||l.getAttribute("title")!==(u.title==null?null:u.title))break;return l;case"style":if(l.hasAttribute("data-precedence"))break;return l;case"script":if(n=l.getAttribute("src"),(n!==(u.src==null?null:u.src)||l.getAttribute("type")!==(u.type==null?null:u.type)||l.getAttribute("crossorigin")!==(u.crossOrigin==null?null:u.crossOrigin))&&n&&l.hasAttribute("async")&&!l.hasAttribute("itemprop"))break;return l;default:return l}}else if(t==="input"&&l.type==="hidden"){var n=u.name==null?null:""+u.name;if(u.type==="hidden"&&l.getAttribute("name")===n)return l}else return l;if(l=Tt(l.nextSibling),l===null)break}return null}function bv(l,t,a){if(t==="")return null;for(;l.nodeType!==3;)if((l.nodeType!==1||l.nodeName!=="INPUT"||l.type!=="hidden")&&!a||(l=Tt(l.nextSibling),l===null))return null;return l}function Ki(l){return l.data==="$!"||l.data==="$?"&&l.ownerDocument.readyState==="complete"}function Sv(l,t){var a=l.ownerDocument;if(l.data!=="$?"||a.readyState==="complete")t();else{var e=function(){t(),a.removeEventListener("DOMContentLoaded",e)};a.addEventListener("DOMContentLoaded",e),l._reactRetry=e}}function Tt(l){for(;l!=null;l=l.nextSibling){var t=l.nodeType;if(t===1||t===3)break;if(t===8){if(t=l.data,t==="$"||t==="$!"||t==="$?"||t==="F!"||t==="F")break;if(t==="/$")return null}}return l}var Ji=null;function Yd(l){l=l.previousSibling;for(var t=0;l;){if(l.nodeType===8){var a=l.data;if(a==="$"||a==="$!"||a==="$?"){if(t===0)return l;t--}else a==="/$"&&t++}l=l.previousSibling}return null}function Bd(l,t,a){switch(t=_n(a),l){case"html":if(l=t.documentElement,!l)throw Error(s(452));return l;case"head":if(l=t.head,!l)throw Error(s(453));return l;case"body":if(l=t.body,!l)throw Error(s(454));return l;default:throw Error(s(451))}}function du(l){for(var t=l.attributes;t.length;)l.removeAttributeNode(t[0]);kn(l)}var gt=new Map,Cd=new Set;function Mn(l){return typeof l.getRootNode=="function"?l.getRootNode():l.nodeType===9?l:l.ownerDocument}var Lt=N.d;N.d={f:zv,r:Tv,D:Ev,C:Av,L:Ov,m:pv,X:_v,S:xv,M:Mv};function zv(){var l=Lt.f(),t=Sn();return l||t}function Tv(l){var t=Ga(l);t!==null&&t.tag===5&&t.type==="form"?eo(t):Lt.r(l)}var Se=typeof document>"u"?null:document;function Gd(l,t,a){var e=Se;if(e&&typeof t=="string"&&t){var u=ot(t);u='link[rel="'+l+'"][href="'+u+'"]',typeof a=="string"&&(u+='[crossorigin="'+a+'"]'),Cd.has(u)||(Cd.add(u),l={rel:l,crossOrigin:a,href:t},e.querySelector(u)===null&&(t=e.createElement("link"),Gl(t,"link",l),Ul(t),e.head.appendChild(t)))}}function Ev(l){Lt.D(l),Gd("dns-prefetch",l,null)}function Av(l,t){Lt.C(l,t),Gd("preconnect",l,t)}function Ov(l,t,a){Lt.L(l,t,a);var e=Se;if(e&&l&&t){var u='link[rel="preload"][as="'+ot(t)+'"]';t==="image"&&a&&a.imageSrcSet?(u+='[imagesrcset="'+ot(a.imageSrcSet)+'"]',typeof a.imageSizes=="string"&&(u+='[imagesizes="'+ot(a.imageSizes)+'"]')):u+='[href="'+ot(l)+'"]';var n=u;switch(t){case"style":n=ze(l);break;case"script":n=Te(l)}gt.has(n)||(l=D({rel:"preload",href:t==="image"&&a&&a.imageSrcSet?void 0:l,as:t},a),gt.set(n,l),e.querySelector(u)!==null||t==="style"&&e.querySelector(ru(n))||t==="script"&&e.querySelector(vu(n))||(t=e.createElement("link"),Gl(t,"link",l),Ul(t),e.head.appendChild(t)))}}function pv(l,t){Lt.m(l,t);var a=Se;if(a&&l){var e=t&&typeof t.as=="string"?t.as:"script",u='link[rel="modulepreload"][as="'+ot(e)+'"][href="'+ot(l)+'"]',n=u;switch(e){case"audioworklet":case"paintworklet":case"serviceworker":case"sharedworker":case"worker":case"script":n=Te(l)}if(!gt.has(n)&&(l=D({rel:"modulepreload",href:l},t),gt.set(n,l),a.querySelector(u)===null)){switch(e){case"audioworklet":case"paintworklet":case"serviceworker":case"sharedworker":case"worker":case"script":if(a.querySelector(vu(n)))return}e=a.createElement("link"),Gl(e,"link",l),Ul(e),a.head.appendChild(e)}}}function xv(l,t,a){Lt.S(l,t,a);var e=Se;if(e&&l){var u=Xa(e).hoistableStyles,n=ze(l);t=t||"default";var c=u.get(n);if(!c){var i={loading:0,preload:null};if(c=e.querySelector(ru(n)))i.loading=5;else{l=D({rel:"stylesheet",href:l,"data-precedence":t},a),(a=gt.get(n))&&wi(l,a);var f=c=e.createElement("link");Ul(f),Gl(f,"link",l),f._p=new Promise(function(m,z){f.onload=m,f.onerror=z}),f.addEventListener("load",function(){i.loading|=1}),f.addEventListener("error",function(){i.loading|=2}),i.loading|=4,Dn(c,t,e)}c={type:"stylesheet",instance:c,count:1,state:i},u.set(n,c)}}}function _v(l,t){Lt.X(l,t);var a=Se;if(a&&l){var e=Xa(a).hoistableScripts,u=Te(l),n=e.get(u);n||(n=a.querySelector(vu(u)),n||(l=D({src:l,async:!0},t),(t=gt.get(u))&&Wi(l,t),n=a.createElement("script"),Ul(n),Gl(n,"link",l),a.head.appendChild(n)),n={type:"script",instance:n,count:1,state:null},e.set(u,n))}}function Mv(l,t){Lt.M(l,t);var a=Se;if(a&&l){var e=Xa(a).hoistableScripts,u=Te(l),n=e.get(u);n||(n=a.querySelector(vu(u)),n||(l=D({src:l,async:!0,type:"module"},t),(t=gt.get(u))&&Wi(l,t),n=a.createElement("script"),Ul(n),Gl(n,"link",l),a.head.appendChild(n)),n={type:"script",instance:n,count:1,state:null},e.set(u,n))}}function Xd(l,t,a,e){var u=(u=J.current)?Mn(u):null;if(!u)throw Error(s(446));switch(l){case"meta":case"title":return null;case"style":return typeof a.precedence=="string"&&typeof a.href=="string"?(t=ze(a.href),a=Xa(u).hoistableStyles,e=a.get(t),e||(e={type:"style",instance:null,count:0,state:null},a.set(t,e)),e):{type:"void",instance:null,count:0,state:null};case"link":if(a.rel==="stylesheet"&&typeof a.href=="string"&&typeof a.precedence=="string"){l=ze(a.href);var n=Xa(u).hoistableStyles,c=n.get(l);if(c||(u=u.ownerDocument||u,c={type:"stylesheet",instance:null,count:0,state:{loading:0,preload:null}},n.set(l,c),(n=u.querySelector(ru(l)))&&!n._p&&(c.instance=n,c.state.loading=5),gt.has(l)||(a={rel:"preload",as:"style",href:a.href,crossOrigin:a.crossOrigin,integrity:a.integrity,media:a.media,hrefLang:a.hrefLang,referrerPolicy:a.referrerPolicy},gt.set(l,a),n||Dv(u,l,a,c.state))),t&&e===null)throw Error(s(528,""));return c}if(t&&e!==null)throw Error(s(529,""));return null;case"script":return t=a.async,a=a.src,typeof a=="string"&&t&&typeof t!="function"&&typeof t!="symbol"?(t=Te(a),a=Xa(u).hoistableScripts,e=a.get(t),e||(e={type:"script",instance:null,count:0,state:null},a.set(t,e)),e):{type:"void",instance:null,count:0,state:null};default:throw Error(s(444,l))}}function ze(l){return'href="'+ot(l)+'"'}function ru(l){return'link[rel="stylesheet"]['+l+"]"}function Qd(l){return D({},l,{"data-precedence":l.precedence,precedence:null})}function Dv(l,t,a,e){l.querySelector('link[rel="preload"][as="style"]['+t+"]")?e.loading=1:(t=l.createElement("link"),e.preload=t,t.addEventListener("load",function(){return e.loading|=1}),t.addEventListener("error",function(){return e.loading|=2}),Gl(t,"link",a),Ul(t),l.head.appendChild(t))}function Te(l){return'[src="'+ot(l)+'"]'}function vu(l){return"script[async]"+l}function Zd(l,t,a){if(t.count++,t.instance===null)switch(t.type){case"style":var e=l.querySelector('style[data-href~="'+ot(a.href)+'"]');if(e)return t.instance=e,Ul(e),e;var u=D({},a,{"data-href":a.href,"data-precedence":a.precedence,href:null,precedence:null});return e=(l.ownerDocument||l).createElement("style"),Ul(e),Gl(e,"style",u),Dn(e,a.precedence,l),t.instance=e;case"stylesheet":u=ze(a.href);var n=l.querySelector(ru(u));if(n)return t.state.loading|=4,t.instance=n,Ul(n),n;e=Qd(a),(u=gt.get(u))&&wi(e,u),n=(l.ownerDocument||l).createElement("link"),Ul(n);var c=n;return c._p=new Promise(function(i,f){c.onload=i,c.onerror=f}),Gl(n,"link",e),t.state.loading|=4,Dn(n,a.precedence,l),t.instance=n;case"script":return n=Te(a.src),(u=l.querySelector(vu(n)))?(t.instance=u,Ul(u),u):(e=a,(u=gt.get(n))&&(e=D({},a),Wi(e,u)),l=l.ownerDocument||l,u=l.createElement("script"),Ul(u),Gl(u,"link",e),l.head.appendChild(u),t.instance=u);case"void":return null;default:throw Error(s(443,t.type))}else t.type==="stylesheet"&&(t.state.loading&4)===0&&(e=t.instance,t.state.loading|=4,Dn(e,a.precedence,l));return t.instance}function Dn(l,t,a){for(var e=a.querySelectorAll('link[rel="stylesheet"][data-precedence],style[data-precedence]'),u=e.length?e[e.length-1]:null,n=u,c=0;c<e.length;c++){var i=e[c];if(i.dataset.precedence===t)n=i;else if(n!==u)break}n?n.parentNode.insertBefore(l,n.nextSibling):(t=a.nodeType===9?a.head:a,t.insertBefore(l,t.firstChild))}function wi(l,t){l.crossOrigin==null&&(l.crossOrigin=t.crossOrigin),l.referrerPolicy==null&&(l.referrerPolicy=t.referrerPolicy),l.title==null&&(l.title=t.title)}function Wi(l,t){l.crossOrigin==null&&(l.crossOrigin=t.crossOrigin),l.referrerPolicy==null&&(l.referrerPolicy=t.referrerPolicy),l.integrity==null&&(l.integrity=t.integrity)}var Nn=null;function Vd(l,t,a){if(Nn===null){var e=new Map,u=Nn=new Map;u.set(a,e)}else u=Nn,e=u.get(a),e||(e=new Map,u.set(a,e));if(e.has(l))return e;for(e.set(l,null),a=a.getElementsByTagName(l),u=0;u<a.length;u++){var n=a[u];if(!(n[xe]||n[Xl]||l==="link"&&n.getAttribute("rel")==="stylesheet")&&n.namespaceURI!=="http://www.w3.org/2000/svg"){var c=n.getAttribute(t)||"";c=l+c;var i=e.get(c);i?i.push(n):e.set(c,[n])}}return e}function Ld(l,t,a){l=l.ownerDocument||l,l.head.insertBefore(a,t==="title"?l.querySelector("head > title"):null)}function Nv(l,t,a){if(a===1||t.itemProp!=null)return!1;switch(l){case"meta":case"title":return!0;case"style":if(typeof t.precedence!="string"||typeof t.href!="string"||t.href==="")break;return!0;case"link":if(typeof t.rel!="string"||typeof t.href!="string"||t.href===""||t.onLoad||t.onError)break;switch(t.rel){case"stylesheet":return l=t.disabled,typeof t.precedence=="string"&&l==null;default:return!0}case"script":if(t.async&&typeof t.async!="function"&&typeof t.async!="symbol"&&!t.onLoad&&!t.onError&&t.src&&typeof t.src=="string")return!0}return!1}function Kd(l){return!(l.type==="stylesheet"&&(l.state.loading&3)===0)}var yu=null;function Uv(){}function Rv(l,t,a){if(yu===null)throw Error(s(475));var e=yu;if(t.type==="stylesheet"&&(typeof a.media!="string"||matchMedia(a.media).matches!==!1)&&(t.state.loading&4)===0){if(t.instance===null){var u=ze(a.href),n=l.querySelector(ru(u));if(n){l=n._p,l!==null&&typeof l=="object"&&typeof l.then=="function"&&(e.count++,e=Un.bind(e),l.then(e,e)),t.state.loading|=4,t.instance=n,Ul(n);return}n=l.ownerDocument||l,a=Qd(a),(u=gt.get(u))&&wi(a,u),n=n.createElement("link"),Ul(n);var c=n;c._p=new Promise(function(i,f){c.onload=i,c.onerror=f}),Gl(n,"link",a),t.instance=n}e.stylesheets===null&&(e.stylesheets=new Map),e.stylesheets.set(t,l),(l=t.state.preload)&&(t.state.loading&3)===0&&(e.count++,t=Un.bind(e),l.addEventListener("load",t),l.addEventListener("error",t))}}function jv(){if(yu===null)throw Error(s(475));var l=yu;return l.stylesheets&&l.count===0&&ki(l,l.stylesheets),0<l.count?function(t){var a=setTimeout(function(){if(l.stylesheets&&ki(l,l.stylesheets),l.unsuspend){var e=l.unsuspend;l.unsuspend=null,e()}},6e4);return l.unsuspend=t,function(){l.unsuspend=null,clearTimeout(a)}}:null}function Un(){if(this.count--,this.count===0){if(this.stylesheets)ki(this,this.stylesheets);else if(this.unsuspend){var l=this.unsuspend;this.unsuspend=null,l()}}}var Rn=null;function ki(l,t){l.stylesheets=null,l.unsuspend!==null&&(l.count++,Rn=new Map,t.forEach(Hv,l),Rn=null,Un.call(l))}function Hv(l,t){if(!(t.state.loading&4)){var a=Rn.get(l);if(a)var e=a.get(null);else{a=new Map,Rn.set(l,a);for(var u=l.querySelectorAll("link[data-precedence],style[data-precedence]"),n=0;n<u.length;n++){var c=u[n];(c.nodeName==="LINK"||c.getAttribute("media")!=="not all")&&(a.set(c.dataset.precedence,c),e=c)}e&&a.set(null,e)}u=t.instance,c=u.getAttribute("data-precedence"),n=a.get(c)||e,n===e&&a.set(null,u),a.set(c,u),this.count++,e=Un.bind(this),u.addEventListener("load",e),u.addEventListener("error",e),n?n.parentNode.insertBefore(u,n.nextSibling):(l=l.nodeType===9?l.head:l,l.insertBefore(u,l.firstChild)),t.state.loading|=4}}var hu={$$typeof:dl,Provider:null,Consumer:null,_currentValue:V,_currentValue2:V,_threadCount:0};function qv(l,t,a,e,u,n,c,i){this.tag=1,this.containerInfo=l,this.pingCache=this.current=this.pendingChildren=null,this.timeoutHandle=-1,this.callbackNode=this.next=this.pendingContext=this.context=this.cancelPendingCommit=null,this.callbackPriority=0,this.expirationTimes=Kn(-1),this.entangledLanes=this.shellSuspendCounter=this.errorRecoveryDisabledLanes=this.expiredLanes=this.warmLanes=this.pingedLanes=this.suspendedLanes=this.pendingLanes=0,this.entanglements=Kn(0),this.hiddenUpdates=Kn(null),this.identifierPrefix=e,this.onUncaughtError=u,this.onCaughtError=n,this.onRecoverableError=c,this.pooledCache=null,this.pooledCacheLanes=0,this.formState=i,this.incompleteTransitions=new Map}function Jd(l,t,a,e,u,n,c,i,f,m,z,A){return l=new qv(l,t,a,c,i,f,m,A),t=1,n===!0&&(t|=24),n=at(3,null,null,t),l.current=n,n.stateNode=l,t=Nc(),t.refCount++,l.pooledCache=t,t.refCount++,n.memoizedState={element:e,isDehydrated:a,cache:t},Hc(n),l}function wd(l){return l?(l=Pa,l):Pa}function Wd(l,t,a,e,u,n){u=wd(u),e.context===null?e.context=u:e.pendingContext=u,e=Pt(t),e.payload={element:a},n=n===void 0?null:n,n!==null&&(e.callback=n),a=It(l,e,t),a!==null&&(it(a,l,t),Ke(a,l,t))}function kd(l,t){if(l=l.memoizedState,l!==null&&l.dehydrated!==null){var a=l.retryLane;l.retryLane=a!==0&&a<t?a:t}}function $i(l,t){kd(l,t),(l=l.alternate)&&kd(l,t)}function $d(l){if(l.tag===13){var t=Fa(l,67108864);t!==null&&it(t,l,67108864),$i(l,67108864)}}var jn=!0;function Yv(l,t,a,e){var u=S.T;S.T=null;var n=N.p;try{N.p=2,Fi(l,t,a,e)}finally{N.p=n,S.T=u}}function Bv(l,t,a,e){var u=S.T;S.T=null;var n=N.p;try{N.p=8,Fi(l,t,a,e)}finally{N.p=n,S.T=u}}function Fi(l,t,a,e){if(jn){var u=Pi(e);if(u===null)Ci(l,t,e,Hn,a),Pd(l,e);else if(Gv(u,l,t,a,e))e.stopPropagation();else if(Pd(l,e),t&4&&-1<Cv.indexOf(l)){for(;u!==null;){var n=Ga(u);if(n!==null)switch(n.tag){case 3:if(n=n.stateNode,n.current.memoizedState.isDehydrated){var c=Sa(n.pendingLanes);if(c!==0){var i=n;for(i.pendingLanes|=2,i.entangledLanes|=2;c;){var f=1<<31-lt(c);i.entanglements[1]|=f,c&=~f}_t(n),(nl&6)===0&&(gn=Et()+500,fu(0))}}break;case 13:i=Fa(n,2),i!==null&&it(i,n,2),Sn(),$i(n,2)}if(n=Pi(e),n===null&&Ci(l,t,e,Hn,a),n===u)break;u=n}u!==null&&e.stopPropagation()}else Ci(l,t,e,null,a)}}function Pi(l){return l=ec(l),Ii(l)}var Hn=null;function Ii(l){if(Hn=null,l=Ca(l),l!==null){var t=q(l);if(t===null)l=null;else{var a=t.tag;if(a===13){if(l=R(t),l!==null)return l;l=null}else if(a===3){if(t.stateNode.current.memoizedState.isDehydrated)return t.tag===3?t.stateNode.containerInfo:null;l=null}else t!==l&&(l=null)}}return Hn=l,null}function Fd(l){switch(l){case"beforetoggle":case"cancel":case"click":case"close":case"contextmenu":case"copy":case"cut":case"auxclick":case"dblclick":case"dragend":case"dragstart":case"drop":case"focusin":case"focusout":case"input":case"invalid":case"keydown":case"keypress":case"keyup":case"mousedown":case"mouseup":case"paste":case"pause":case"play":case"pointercancel":case"pointerdown":case"pointerup":case"ratechange":case"reset":case"resize":case"seeked":case"submit":case"toggle":case"touchcancel":case"touchend":case"touchstart":case"volumechange":case"change":case"selectionchange":case"textInput":case"compositionstart":case"compositionend":case"compositionupdate":case"beforeblur":case"afterblur":case"beforeinput":case"blur":case"fullscreenchange":case"focus":case"hashchange":case"popstate":case"select":case"selectstart":return 2;case"drag":case"dragenter":case"dragexit":case"dragleave":case"dragover":case"mousemove":case"mouseout":case"mouseover":case"pointermove":case"pointerout":case"pointerover":case"scroll":case"touchmove":case"wheel":case"mouseenter":case"mouseleave":case"pointerenter":case"pointerleave":return 8;case"message":switch(Ar()){case of:return 2;case df:return 8;case pu:case Or:return 32;case rf:return 268435456;default:return 32}default:return 32}}var lf=!1,ra=null,va=null,ya=null,mu=new Map,gu=new Map,ha=[],Cv="mousedown mouseup touchcancel touchend touchstart auxclick dblclick pointercancel pointerdown pointerup dragend dragstart drop compositionend compositionstart keydown keypress keyup input textInput copy cut paste click change contextmenu reset".split(" ");function Pd(l,t){switch(l){case"focusin":case"focusout":ra=null;break;case"dragenter":case"dragleave":va=null;break;case"mouseover":case"mouseout":ya=null;break;case"pointerover":case"pointerout":mu.delete(t.pointerId);break;case"gotpointercapture":case"lostpointercapture":gu.delete(t.pointerId)}}function bu(l,t,a,e,u,n){return l===null||l.nativeEvent!==n?(l={blockedOn:t,domEventName:a,eventSystemFlags:e,nativeEvent:n,targetContainers:[u]},t!==null&&(t=Ga(t),t!==null&&$d(t)),l):(l.eventSystemFlags|=e,t=l.targetContainers,u!==null&&t.indexOf(u)===-1&&t.push(u),l)}function Gv(l,t,a,e,u){switch(t){case"focusin":return ra=bu(ra,l,t,a,e,u),!0;case"dragenter":return va=bu(va,l,t,a,e,u),!0;case"mouseover":return ya=bu(ya,l,t,a,e,u),!0;case"pointerover":var n=u.pointerId;return mu.set(n,bu(mu.get(n)||null,l,t,a,e,u)),!0;case"gotpointercapture":return n=u.pointerId,gu.set(n,bu(gu.get(n)||null,l,t,a,e,u)),!0}return!1}function Id(l){var t=Ca(l.target);if(t!==null){var a=q(t);if(a!==null){if(t=a.tag,t===13){if(t=R(a),t!==null){l.blockedOn=t,Rr(l.priority,function(){if(a.tag===13){var e=ct();e=Jn(e);var u=Fa(a,e);u!==null&&it(u,a,e),$i(a,e)}});return}}else if(t===3&&a.stateNode.current.memoizedState.isDehydrated){l.blockedOn=a.tag===3?a.stateNode.containerInfo:null;return}}}l.blockedOn=null}function qn(l){if(l.blockedOn!==null)return!1;for(var t=l.targetContainers;0<t.length;){var a=Pi(l.nativeEvent);if(a===null){a=l.nativeEvent;var e=new a.constructor(a.type,a);ac=e,a.target.dispatchEvent(e),ac=null}else return t=Ga(a),t!==null&&$d(t),l.blockedOn=a,!1;t.shift()}return!0}function lr(l,t,a){qn(l)&&a.delete(t)}function Xv(){lf=!1,ra!==null&&qn(ra)&&(ra=null),va!==null&&qn(va)&&(va=null),ya!==null&&qn(ya)&&(ya=null),mu.forEach(lr),gu.forEach(lr)}function Yn(l,t){l.blockedOn===t&&(l.blockedOn=null,lf||(lf=!0,v.unstable_scheduleCallback(v.unstable_NormalPriority,Xv)))}var Bn=null;function tr(l){Bn!==l&&(Bn=l,v.unstable_scheduleCallback(v.unstable_NormalPriority,function(){Bn===l&&(Bn=null);for(var t=0;t<l.length;t+=3){var a=l[t],e=l[t+1],u=l[t+2];if(typeof e!="function"){if(Ii(e||a)===null)continue;break}var n=Ga(a);n!==null&&(l.splice(t,3),t-=3,Ic(n,{pending:!0,data:u,method:a.method,action:e},e,u))}}))}function Su(l){function t(f){return Yn(f,l)}ra!==null&&Yn(ra,l),va!==null&&Yn(va,l),ya!==null&&Yn(ya,l),mu.forEach(t),gu.forEach(t);for(var a=0;a<ha.length;a++){var e=ha[a];e.blockedOn===l&&(e.blockedOn=null)}for(;0<ha.length&&(a=ha[0],a.blockedOn===null);)Id(a),a.blockedOn===null&&ha.shift();if(a=(l.ownerDocument||l).$$reactFormReplay,a!=null)for(e=0;e<a.length;e+=3){var u=a[e],n=a[e+1],c=u[wl]||null;if(typeof n=="function")c||tr(a);else if(c){var i=null;if(n&&n.hasAttribute("formAction")){if(u=n,c=n[wl]||null)i=c.formAction;else if(Ii(u)!==null)continue}else i=c.action;typeof i=="function"?a[e+1]=i:(a.splice(e,3),e-=3),tr(a)}}}function tf(l){this._internalRoot=l}Cn.prototype.render=tf.prototype.render=function(l){var t=this._internalRoot;if(t===null)throw Error(s(409));var a=t.current,e=ct();Wd(a,e,l,t,null,null)},Cn.prototype.unmount=tf.prototype.unmount=function(){var l=this._internalRoot;if(l!==null){this._internalRoot=null;var t=l.containerInfo;Wd(l.current,2,null,l,null,null),Sn(),t[Ba]=null}};function Cn(l){this._internalRoot=l}Cn.prototype.unstable_scheduleHydration=function(l){if(l){var t=gf();l={blockedOn:null,target:l,priority:t};for(var a=0;a<ha.length&&t!==0&&t<ha[a].priority;a++);ha.splice(a,0,l),a===0&&Id(l)}};var ar=_.version;if(ar!=="19.1.1")throw Error(s(527,ar,"19.1.1"));N.findDOMNode=function(l){var t=l._reactInternals;if(t===void 0)throw typeof l.render=="function"?Error(s(188)):(l=Object.keys(l).join(","),Error(s(268,l)));return l=M(t),l=l!==null?y(l):null,l=l===null?null:l.stateNode,l};var Qv={bundleType:0,version:"19.1.1",rendererPackageName:"react-dom",currentDispatcherRef:S,reconcilerVersion:"19.1.1"};if(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__<"u"){var Gn=__REACT_DEVTOOLS_GLOBAL_HOOK__;if(!Gn.isDisabled&&Gn.supportsFiber)try{Ae=Gn.inject(Qv),Il=Gn}catch{}}return Tu.createRoot=function(l,t){if(!x(l))throw Error(s(299));var a=!1,e="",u=bo,n=So,c=zo,i=null;return t!=null&&(t.unstable_strictMode===!0&&(a=!0),t.identifierPrefix!==void 0&&(e=t.identifierPrefix),t.onUncaughtError!==void 0&&(u=t.onUncaughtError),t.onCaughtError!==void 0&&(n=t.onCaughtError),t.onRecoverableError!==void 0&&(c=t.onRecoverableError),t.unstable_transitionCallbacks!==void 0&&(i=t.unstable_transitionCallbacks)),t=Jd(l,1,!1,null,null,a,e,u,n,c,i,null),l[Ba]=t.current,Bi(l),new tf(t)},Tu.hydrateRoot=function(l,t,a){if(!x(l))throw Error(s(299));var e=!1,u="",n=bo,c=So,i=zo,f=null,m=null;return a!=null&&(a.unstable_strictMode===!0&&(e=!0),a.identifierPrefix!==void 0&&(u=a.identifierPrefix),a.onUncaughtError!==void 0&&(n=a.onUncaughtError),a.onCaughtError!==void 0&&(c=a.onCaughtError),a.onRecoverableError!==void 0&&(i=a.onRecoverableError),a.unstable_transitionCallbacks!==void 0&&(f=a.unstable_transitionCallbacks),a.formState!==void 0&&(m=a.formState)),t=Jd(l,1,!0,t,a??null,e,u,n,c,i,f,m),t.context=wd(null),a=t.current,e=ct(),e=Jn(e),u=Pt(e),u.callback=null,It(a,u,e),a=e,t.current.lanes=a,pe(t,a),_t(t),l[Ba]=t.current,Bi(l),new Cn(t)},Tu.version="19.1.1",Tu}var rr;function Pv(){if(rr)return uf.exports;rr=1;function v(){if(!(typeof __REACT_DEVTOOLS_GLOBAL_HOOK__>"u"||typeof __REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE!="function"))try{__REACT_DEVTOOLS_GLOBAL_HOOK__.checkDCE(v)}catch(_){console.error(_)}}return v(),uf.exports=Fv(),uf.exports}var Iv=Pv();const ly=yr(Iv);function ga({title:v,children:_,right:E}){return O.jsxs("div",{className:"bg-zinc-900/80 border border-zinc-800 rounded-2xl p-4",children:[O.jsxs("div",{className:"flex items-center justify-between mb-3",children:[O.jsx("h2",{className:"text-zinc-200 font-medium",children:v}),E]}),_]})}function Mt({label:v,value:_,onChange:E,type:s="text",placeholder:x="",className:q="",disabled:R=!1}){return O.jsxs("label",{className:`block text-sm ${q}`,children:[O.jsx("span",{className:"text-zinc-400",children:v}),O.jsx("input",{type:s,value:_,placeholder:x,onChange:j=>E(j.target.value),disabled:R,className:"mt-1 w-full rounded-lg bg-zinc-900 border border-zinc-800 px-3 py-2 text-zinc-100 focus:outline-none focus:ring-2 focus:ring-blue-500"})]})}function hr(){return Math.random().toString(36).slice(2)}function mr(v,_){return _[v]?.label||v}function Au(v,_){const E=_[v]||{fields:[]},s={id:hr(),type:v,value:0};for(const x of E.fields)s[x.key]=x.default||"";return s}function ty(v){const{id:_,...E}=v;return E}function ba(v){try{return new Intl.NumberFormat(void 0,{style:"currency",currency:"EUR",maximumFractionDigits:0}).format(v)}catch{return`${Math.round(v).toLocaleString()} €`}}function ay({data:v}){const _=Ml.useRef(null);return Ml.useEffect(()=>{const E=_.current;if(!E)return;const s=window.devicePixelRatio||1,x=E.clientWidth*s,q=E.clientHeight*s;E.width=x,E.height=q;const R=E.getContext("2d");if(!R)return;R.clearRect(0,0,x,q);const j=32*s;if(R.strokeStyle="#2a2a2a",R.lineWidth=1,R.beginPath(),R.moveTo(j,j),R.lineTo(j,q-j),R.lineTo(x-j,q-j),R.stroke(),!v.length){R.fillStyle="#9aa0a6",R.font=`${14*s}px ui-sans-serif`,R.fillText("No data",12*s,20*s);return}const M=v.map((hl,ql)=>ql),y=v.map(hl=>hl.value),D=Math.min(...y),Q=Math.max(...y),B=hl=>j+hl/Math.max(1,M.length-1)*(x-2*j),ll=hl=>q-j-(hl-D)/Math.max(1,Q-D)*(q-2*j);R.beginPath();for(let hl=0;hl<M.length;hl++){const ql=B(M[hl]),dl=ll(y[hl]);hl===0?R.moveTo(ql,dl):R.lineTo(ql,dl)}R.lineWidth=2*s,R.strokeStyle="#8ab4f8",R.stroke();const ul=v[v.length-1],pl=B(M[M.length-1]),Dl=ll(y[y.length-1]);R.fillStyle="#e8eaed",R.font=`${12*s}px ui-sans-serif`,R.fillText(`${ul.label}: ${ba(ul.value)}`,pl-100*s,Dl-8*s)},[v]),O.jsx("canvas",{ref:_,className:"w-full h-64 rounded border border-zinc-800 bg-zinc-900"})}function gr({onClick:v,children:_}){return O.jsx("button",{onClick:v,className:"px-3 py-2 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700 text-sm",children:_})}function ey({assets:v,setAssets:_,assetTypes:E}){function s(j,M){_(v.map(y=>y.id===j?{...y,...M}:y))}function x(j){_(v.filter(M=>M.id!==j))}function q(j){_([...v,Au(j,E)])}function R(j,M){_(v.map(y=>y.id!==j?y:{...Au(M,E),id:y.id,value:y.value}))}return O.jsxs("div",{className:"space-y-3",children:[O.jsx("div",{className:"flex flex-wrap gap-2",children:Object.entries(E).map(([j,M])=>O.jsxs(gr,{onClick:()=>q(j),children:["Add ",M.label]},j))}),O.jsx("div",{className:"space-y-2",children:v.map(j=>{const M=E[j.type]||{fields:[]};return O.jsxs("div",{className:"grid md:grid-cols-12 gap-2 items-end bg-zinc-900/60 border border-zinc-800 rounded-xl p-3",children:[O.jsxs("div",{className:"md:col-span-2",children:[O.jsx("label",{className:"text-xs text-zinc-400",children:"Type"}),O.jsxs("select",{className:"mt-1 w-full rounded-lg bg-zinc-900 border border-zinc-800 px-3 py-2 text-zinc-100",value:j.type,onChange:y=>R(j.id,y.target.value),children:[!E[j.type]&&O.jsx("option",{value:j.type,children:j.type}),Object.entries(E).map(([y,D])=>O.jsx("option",{value:y,children:D.label},y))]})]}),M.fields.map(y=>O.jsx(Mt,{className:"md:col-span-3",label:y.label,value:j[y.key]||"",onChange:D=>s(j.id,{[y.key]:D})},y.key)),O.jsx(Mt,{className:"md:col-span-3",label:"Value",type:"number",value:String(j.value),onChange:y=>s(j.id,{value:Number(y||0)})}),O.jsx("div",{className:"md:col-span-1 flex justify-end",children:O.jsx("button",{onClick:()=>x(j.id),className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:"Remove"})})]},j.id)})})]})}function uy({data:v,assetTypes:_}){const E=Object.keys(v.investPlan||{}).sort();return E.length===0?null:O.jsxs("div",{className:"mt-4",children:[O.jsxs("div",{className:"text-sm text-zinc-400 mb-2",children:["Now: ",O.jsx("b",{className:"text-zinc-200",children:ba(v.totalNow)})," → Target: ",O.jsx("b",{className:"text-zinc-200",children:ba(v.targetTotal)})]}),O.jsxs("table",{className:"w-full text-sm",children:[O.jsx("thead",{className:"text-zinc-400",children:O.jsxs("tr",{children:[O.jsx("th",{className:"text-left py-1",children:"Category"}),O.jsx("th",{className:"text-right py-1",children:"Current"}),O.jsx("th",{className:"text-right py-1",children:"Ideal"}),O.jsx("th",{className:"text-right py-1",children:"Invest now"})]})}),O.jsx("tbody",{children:E.map(s=>O.jsxs("tr",{className:"border-t border-zinc-800",children:[O.jsx("td",{className:"py-1 text-zinc-200",children:mr(s,_)}),O.jsx("td",{className:"py-1 text-right text-zinc-300",children:ba(v.byCat[s]||0)}),O.jsx("td",{className:"py-1 text-right text-zinc-300",children:ba(v.idealByCat[s]||0)}),O.jsx("td",{className:"py-1 text-right font-medium text-zinc-100",children:ba(v.investPlan[s]||0)})]},s))})]})]})}function ny({assetTypes:v,setAssetTypes:_}){function E(y,D){_({...v,[y]:{...v[y],label:D}})}function s(y,D,Q){const B=v[y],ll=B.fields.map((ul,pl)=>pl===D?{...ul,key:Q}:ul);_({...v,[y]:{...B,fields:ll}})}function x(y,D,Q){const B=v[y],ll=B.fields.map((ul,pl)=>pl===D?{...ul,label:Q}:ul);_({...v,[y]:{...B,fields:ll}})}function q(){const y=window.prompt("New type key","new_type");y&&!v[y]&&_({...v,[y]:{label:y,fields:[]}})}function R(y){const{[y]:D,...Q}=v;_(Q)}function j(y){const D=window.prompt("Field key","field"),Q=window.prompt("Field label","Field");if(D&&Q){const B=v[y];_({...v,[y]:{...B,fields:[...B.fields,{key:D,label:Q}]}})}}function M(y,D){const Q=v[y],B=Q.fields.filter((ll,ul)=>ul!==D);_({...v,[y]:{...Q,fields:B}})}return O.jsxs("div",{className:"space-y-4",children:[Object.entries(v).map(([y,D])=>O.jsxs("div",{className:"border border-zinc-800 bg-zinc-900/60 rounded-xl p-3 space-y-2",children:[O.jsxs("div",{className:"flex items-end gap-2",children:[O.jsx(Mt,{label:"Type",value:y,onChange:()=>{},disabled:!0,className:"md:col-span-3"}),O.jsx(Mt,{label:"Label",value:D.label,onChange:Q=>E(y,Q),className:"md:col-span-3"}),O.jsx("div",{className:"flex-1 text-right",children:O.jsx("button",{onClick:()=>R(y),className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:"Delete"})})]}),D.fields.map((Q,B)=>O.jsxs("div",{className:"flex items-end gap-2",children:[O.jsx(Mt,{label:"Field key",value:Q.key,onChange:ll=>s(y,B,ll)}),O.jsx(Mt,{label:"Field label",value:Q.label,onChange:ll=>x(y,B,ll)}),O.jsx("button",{onClick:()=>M(y,B),className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:"Remove"})]},B)),O.jsx("button",{onClick:()=>j(y),className:"px-3 py-2 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700 text-sm",children:"Add field"})]},y)),O.jsx(gr,{onClick:q,children:"Add type"})]})}function cy({allocation:v,setAllocation:_,assetTypes:E}){const s=Array.from(new Set([...Object.keys(E),...Object.keys(v||{})])),x=Object.values(v||{}).reduce((M,y)=>M+(Number(y)||0),0);function q(M,y){_({...v,[M]:y})}function R(M){const{[M]:y,...D}=v;_(D)}function j(){const M=window.prompt("New category key (slug)","other");M&&_({...v,[M]:0})}return O.jsxs("div",{className:"space-y-2",children:[s.map(M=>O.jsxs("div",{className:"grid grid-cols-12 items-end gap-2",children:[O.jsx("div",{className:"col-span-6 text-zinc-300",children:mr(M,E)}),O.jsx("div",{className:"col-span-4",children:O.jsx(Mt,{label:"Target %",type:"number",value:String(v[M]??0),onChange:y=>q(M,Number(y||0))})}),O.jsx("div",{className:"col-span-2 flex justify-end",children:O.jsx("button",{onClick:()=>R(M),className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:"Remove"})})]},M)),O.jsxs("div",{className:"flex items-center justify-between mt-2",children:[O.jsx("button",{onClick:j,className:"px-3 py-2 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700 text-sm",children:"Add category"}),O.jsxs("div",{className:`text-sm ${x===100?"text-emerald-400":"text-amber-400"}`,children:["Total: ",x,"% ",x!==100&&"(will be normalized)"]})]})]})}function iy({assetTypes:v,setAssetTypes:_,allocation:E,setAllocation:s}){return O.jsxs("div",{className:"space-y-6",children:[O.jsx(ga,{title:"Asset Types",children:O.jsx(ny,{assetTypes:v,setAssetTypes:_})}),O.jsx(ga,{title:"Allocation",children:O.jsx(cy,{allocation:E,setAllocation:s,assetTypes:v})})]})}const Eu={cash:{label:"Cash",fields:[{key:"currency",label:"Currency",default:"EUR"}]},real_estate:{label:"Real estate",fields:[{key:"description",label:"Description"}]},stock:{label:"Stock",fields:[{key:"ticker",label:"Ticker"}]},private_company_share:{label:"Private company share",fields:[{key:"company",label:"Company"}]}};function Xn(v){return(v||[]).reduce((_,E)=>_+(Number(E.value)||0),0)}function fy(v){const _={};for(const E of v||[]){const s=E.type;_[s]=(_[s]||0)+(Number(E.value)||0)}return _}function sy(v){const _=Object.values(v).reduce((s,x)=>s+(Number(x)||0),0)||1,E={};for(const s of Object.keys(v))E[s]=(Number(v[s])||0)/_;return E}function oy(v,_,E){const s=Xn(v),x=fy(v),q=sy(_),R=Array.from(new Set([...Object.keys(x),...Object.keys(q)])),j=s+(Number(E)||0),M={};for(const B of R)M[B]=(q[B]||0)*j;const y={};for(const B of R)y[B]=Math.max(0,(M[B]||0)-(x[B]||0));const D=Object.values(y).reduce((B,ll)=>B+ll,0)||1,Q={};for(const B of R)Q[B]=y[B]/D*(Number(E)||0);return{totalNow:s,targetTotal:j,byCat:x,idealByCat:M,investPlan:Q}}function dy(v,_){const E=x=>_==="monthly"?`${x.getFullYear()}-${String(x.getMonth()+1).padStart(2,"0")}`:`${x.getFullYear()}`,s=new Map;for(const x of v||[]){const q=E(x.date),R=s.get(q);(!R||x.date.getTime()>R.lastDate)&&s.set(q,{label:q,value:x.value,lastDate:x.date.getTime()})}return Array.from(s.values()).sort((x,q)=>x.lastDate-q.lastDate)}function ry(v,_){const E=(v||[]).map(x=>({date:new Date(x.asOf),value:Xn(x.assets||[])}));return dy(E,_)}const vy="portfolio-tracker-db",Ee="handles";function br(){return new Promise((v,_)=>{const E=indexedDB.open(vy,1);E.onupgradeneeded=()=>{const s=E.result;s.objectStoreNames.contains(Ee)||s.createObjectStore(Ee)},E.onsuccess=()=>v(E.result),E.onerror=()=>_(E.error)})}async function vr(v,_){const E=await br();await new Promise((s,x)=>{const j=E.transaction(Ee,"readwrite").objectStore(Ee).put(_,v);j.onsuccess=()=>s(),j.onerror=()=>x(j.error)})}async function yy(v){const _=await br();return await new Promise((E,s)=>{const R=_.transaction(Ee,"readonly").objectStore(Ee).get(v);R.onsuccess=()=>E(R.result),R.onerror=()=>s(R.error)})}const Sr=new TextEncoder().encode("PTv1.enc");async function zr(v,_){const E=new TextEncoder,s=await crypto.subtle.importKey("raw",E.encode(v),{name:"PBKDF2"},!1,["deriveKey"]);return await crypto.subtle.deriveKey({name:"PBKDF2",salt:_,iterations:15e4,hash:"SHA-256"},s,{name:"AES-GCM",length:256},!1,["encrypt","decrypt"])}function hy(...v){const _=v.reduce((x,q)=>x+q.length,0),E=new Uint8Array(_);let s=0;for(const x of v)E.set(x,s),s+=x.length;return E}async function my(v,_){const E=new TextEncoder().encode(JSON.stringify(v,null,2)),s=crypto.getRandomValues(new Uint8Array(16)),x=crypto.getRandomValues(new Uint8Array(12)),q=await zr(_,s),R=new Uint8Array(await crypto.subtle.encrypt({name:"AES-GCM",iv:x},q,E));return hy(new Uint8Array(Sr),s,x,R)}async function gy(v,_){const E=new Uint8Array(v),s=E.slice(0,8);if(!by(s,Sr))throw new Error("Invalid file format");const x=E.slice(8,24),q=E.slice(24,36),R=E.slice(36),j=await zr(_,x),M=await crypto.subtle.decrypt({name:"AES-GCM",iv:q},j,R),y=new TextDecoder().decode(new Uint8Array(M));return JSON.parse(y)}function by(v,_){if(v.length!==_.length)return!1;for(let E=0;E<v.length;E++)if(v[E]!==_[E])return!1;return!0}async function Sy(){try{const[v]=await window.showOpenFilePicker({types:[{description:"Portfolio",accept:{"application/octet-stream":[".enc"]}}]});return await vr("fileHandle",v),v}catch{const _=await window.showSaveFilePicker({suggestedName:"portfolio.enc",types:[{description:"Portfolio",accept:{"application/octet-stream":[".enc"]}}]});return await vr("fileHandle",_),_}}async function zy(){return await yy("fileHandle")}async function Ty(v,_){const E=await v.getFile();if(E.size===0)return{version:1,assetTypes:{},allocation:{},snapshots:[]};const s=await E.arrayBuffer();return await gy(s,_)}async function Ey(v,_,E){const s=await my(E,_),x=await v.createWritable();await x.write(s),await x.close()}function Ay(){const[v,_]=Ml.useState(""),[E,s]=Ml.useState(null),[x,q]=Ml.useState([]),[R,j]=Ml.useState(()=>new Date().toISOString().slice(0,10)),[M,y]=Ml.useState(Eu),[D,Q]=Ml.useState([Au("cash",Eu),Au("real_estate",Eu),Au("stock",Eu)]),[B,ll]=Ml.useState({cash:20,real_estate:50,stock:30}),[ul,pl]=Ml.useState(0),[Dl,hl]=Ml.useState(!1),[ql,dl]=Ml.useState(null),[Vl,P]=Ml.useState("monthly"),[Yl,Ll]=Ml.useState("main");Ml.useEffect(()=>{(async()=>{try{const K=await zy();if(K&&K.queryPermission){const il=await K.queryPermission({mode:"readwrite"});(il==="granted"||il==="prompt")&&s(K)}}catch(K){console.warn("No saved handle",K)}})()},[]);const Kl=Ml.useMemo(()=>Xn(D),[D]),ft=Ml.useMemo(()=>ry(x,Vl),[x,Vl]),Kt=Ml.useMemo(()=>oy(D,B,ul||0),[D,B,ul]);async function bt(){try{const K=await Sy();s(K)}catch(K){dl(K&&K.message?K.message:String(K))}}async function Nl(){if(!E||!v)return dl("Pick a file and enter password first.");hl(!0),dl(null);try{const K=await Ty(E,v),il=K.snapshots||[];q(il);const S=il[il.length-1];S&&Q((S.assets||[]).map(N=>({...N,id:hr()}))),ll(K.allocation||{}),y(K.assetTypes||Eu)}catch(K){dl(K&&K.message?K.message:String(K))}finally{hl(!1)}}async function Dt(){if(!E||!v)return dl("Pick a file and enter password first.");hl(!0),dl(null);try{const K={asOf:R,assets:D.map(ty)},il=[...x,K];await Ey(E,v,{version:1,assetTypes:M,allocation:B,snapshots:il}),q(il),alert("Saved snapshot")}catch(K){dl(K&&K.message?K.message:String(K))}finally{hl(!1)}}return O.jsx("div",{className:"min-h-screen bg-zinc-950 text-zinc-100",children:O.jsxs("div",{className:"max-w-6xl mx-auto p-6 space-y-6",children:[O.jsxs("header",{className:"flex items-center justify-between",children:[O.jsxs("div",{children:[O.jsx("h1",{className:"text-2xl font-semibold",children:"Local‑First Portfolio Tracker"}),O.jsx("p",{className:"text-sm text-zinc-400",children:"Private by default · Encrypted portfolio file on your disk"})]}),O.jsxs("div",{className:"flex items-end gap-2",children:[O.jsx(Mt,{label:"Password",type:"password",value:v,onChange:_,className:"w-56"}),O.jsx("button",{onClick:bt,className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:E?"Change File":"Choose File"}),O.jsx("button",{onClick:Nl,className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:"Load"}),O.jsx("button",{onClick:Dt,className:"h-10 px-3 rounded-lg bg-blue-600 hover:bg-blue-500",children:"Save snapshot"}),O.jsx("button",{onClick:()=>Ll(Yl==="main"?"config":"main"),className:"h-10 px-3 rounded-lg bg-zinc-800 border border-zinc-700 hover:bg-zinc-700",children:Yl==="main"?"Config":"Back"})]})]}),ql&&O.jsx("div",{className:"p-3 rounded-xl bg-red-900/30 border border-red-800 text-red-200",children:ql}),Dl&&O.jsx("div",{className:"p-3 rounded-xl bg-zinc-800 text-zinc-300",children:"Working…"}),Yl==="main"?O.jsxs(O.Fragment,{children:[O.jsxs("div",{className:"grid md:grid-cols-3 gap-6",children:[O.jsx(ga,{title:"As of date",children:O.jsx(Mt,{label:"Date",type:"date",value:R,onChange:j})}),O.jsxs(ga,{title:"Net worth (current)",children:[O.jsx("div",{className:"text-3xl font-semibold",children:ba(Kl)}),O.jsx("div",{className:"text-xs text-zinc-400 mt-1",children:"Computed from asset list"})]}),O.jsx(ga,{title:"History view",right:O.jsxs("select",{value:Vl,onChange:K=>P(K.target.value),className:"bg-zinc-900 border border-zinc-700 rounded-lg px-2 py-1 text-sm",children:[O.jsx("option",{value:"monthly",children:"Monthly"}),O.jsx("option",{value:"yearly",children:"Yearly"})]}),children:O.jsx(ay,{data:ft})})]}),O.jsxs("div",{className:"grid lg:grid-cols-2 gap-6",children:[O.jsx(ga,{title:"Assets",children:O.jsx(ey,{assets:D,setAssets:Q,assetTypes:M})}),O.jsxs(ga,{title:"Rebalance",children:[O.jsxs("div",{className:"grid grid-cols-2 gap-4",children:[O.jsx(Mt,{label:"New capital to invest",type:"number",value:String(ul),onChange:K=>pl(Number(K||0))}),O.jsx("div",{className:"text-sm text-zinc-400 flex items-end",children:"Distribute this to move toward target without selling."})]}),O.jsx(uy,{data:Kt,assetTypes:M})]})]}),O.jsx(ga,{title:"Loaded snapshots",children:x.length===0?O.jsx("div",{className:"text-zinc-400 text-sm",children:"None yet. Create your first snapshot above."}):O.jsx("ul",{className:"text-sm max-h-48 overflow-auto space-y-1",children:x.map((K,il)=>O.jsxs("li",{className:"text-zinc-300",children:[new Date(K.asOf).toISOString().slice(0,10)," — ",ba(Xn(K.assets||[]))]},il))})})]}):O.jsx(iy,{assetTypes:M,setAssetTypes:y,allocation:B,setAllocation:ll})]})})}ly.createRoot(document.getElementById("root")).render(O.jsx(Jv.StrictMode,{children:O.jsx(Ay,{})}));
