+++ docs/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Portfolio Tracker</title>
-    <script type="module" crossorigin src="./assets/index-B7sdhgmR.js"></script>
+    <script type="module" crossorigin src="./assets/index-BQTPNGku.js"></script>
     <link rel="stylesheet" crossorigin href="./assets/index-BGR37Dks.css">
   </head>
   <body>
