+++ src/components/ConfigPage.jsx
@@ -1,0 +1,16 @@
+import Section from "./Section.jsx";
+import AssetTypeManager from "./AssetTypeManager.jsx";
+import AllocationEditor from "./AllocationEditor.jsx";
+
+export default function ConfigPage({ assetTypes, setAssetTypes, allocation, setAllocation }) {
+  return (
+    <div className="space-y-6">
+      <Section title="Asset Types">
+        <AssetTypeManager assetTypes={assetTypes} setAssetTypes={setAssetTypes} />
+      </Section>
+      <Section title="Allocation">
+        <AllocationEditor allocation={allocation} setAllocation={setAllocation} assetTypes={assetTypes} />
+      </Section>
+    </div>
+  );
+}
